% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drought_utils.R
\name{bf_forecast_start_date}
\alias{bf_forecast_start_date}
\title{bf_forecast_start_date}
\usage{
bf_forecast_start_date(
  start_date,
  adjust_start_date,
  gage_data,
  flow_col = "value",
  date_col = "time"
)
}
\arguments{
\item{start_date}{A character or date vector of length 1 identifying the
starting point of the user analysis that will be adjusted by
adjust_start_date}

\item{adjust_start_date}{Either an integer vector of length 1 or a
character/Date vector of length 2. If a integer is entered, an appropriate
start date will be searched for in the prior adjust_start_date days before
start_date. Otherwise, if a character/Date vector is input, an appropriate
starting date will be searched between the input dates}

\item{gage_data}{A data frame containing flow and date field identified by
flow_col and date_col, often derived via \code{WaterGageDaily()}}

\item{flow_col}{The field containing flow values that will be searched for
minimum values in the user dates of interest}

\item{date_col}{The field containing date values that contain the date
indexing}
}
\value{
A Date representing the adjusted start date, which may be identical to
start_date
}
\description{
Find appropriate start dates for baseflow forecasting
}
\details{
This function assists users in finding an appropriate start date to
conduct a baseflow forecast using the methods on \code{WaterGageDaily()}.
Such forecasts are recommended to start from a low point in the season of
interest where ET is constant and no recharge is occurring. This function
allows users to enter a start date and then find an appropriate start in the
prior x-days or between two dates by identifying valid minimum flow.
}
\examples{
\dontrun{
gage_obj <- WaterGageDaily$new(
 gage_id = '02069700', ds_in = ds,
 start_date = "2002-01-01",end_date = "2002-12-31"
)
bf_forecast_start_date(start_date = "2002-08-01",
                      adjust_start_date = 30,
                      gage_data = gage_obj$gage_data,
                      flow_col = gage_obj$flow_col,
                      date_col = gage_obj$date_col)
bf_forecast_start_date(start_date = "2002-08-01",
                      adjust_start_date = c("2002-05-01","2002-07-01"),
                      gage_data = gage_obj$gage_data,
                      flow_col = gage_obj$flow_col,
                      date_col = gage_obj$date_col)
}
}
