% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CEDSEntity.R
\name{CEDSEntity}
\alias{CEDSEntity}
\title{Generic Entity from CEDS}
\value{
Instance of CEDSFacility with populated facility information
}
\description{
CEDSEntity is the base entity R6 that others like CEDSFacility inherit
}
\details{
Contains generic standard methods for searching CEDS and generic fields
for storing entity information. These are inherited by other objects like \code{CEDSFacility}
and provide a shared basis for querying methods, like get_mps, so they can work for
both \code{CEDSFacility} and \code{CEDSPermit}. This will rarely be called by itself,
and is only called by the children functions.
}
\examples{
\dontrun{
#Create a CEDSFacility using pkid
fac <- CEDSFacility$new(dsCEDS, pkid = 200000203181)
# Using the config to specify other names
fac2 <- CEDSFacility$new(dsCEDS, config = list("Facility Name" = 'Schaefer Landis Farm LLC'))
}
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{pkid}}{Primary ID value of the entity}

    \item{\code{ds}}{Datasource for this entity, created by \code{CEDSDataSource}}

    \item{\code{entity_type}}{What type of record is this object representing}

    \item{\code{tbl_info}}{Result of fn_get_table_names to identify key columns for the entity_type}

    \item{\code{tbl_df}}{Row of the entity's respective view for this record}

    \item{\code{permits}}{List/df of all permits created by \code{CEDSEntity$get_permits()}}

    \item{\code{sf}}{Coordinates of the entity in an sf data.frame}

    \item{\code{contacts}}{Contacts associated with the entity, created by
\code{CEDSEntity$get_contacts()}}

    \item{\code{mps}}{list/df of measuring points on Withdrawal permits/registraitons, created by \code{CEDSEntity$get_mps()}}

    \item{\code{locality}}{Locality that the entity is in}

    \item{\code{five_yr_avg}}{The five year average withdrawal of the entity (sum of all MPs)}

    \item{\code{withdrawals}}{Data.frame of withdrawals of this entity's mps}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-CEDS_Facility-initialize}{\code{CEDS_Facility$new()}}
    \item \href{#method-CEDS_Facility-get_mps}{\code{CEDS_Facility$get_mps()}}
    \item \href{#method-CEDS_Facility-get_contacts}{\code{CEDS_Facility$get_contacts()}}
    \item \href{#method-CEDS_Facility-get_withdrawals}{\code{CEDS_Facility$get_withdrawals()}}
    \item \href{#method-CEDS_Facility-clone}{\code{CEDS_Facility$clone()}}
  }
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CEDS_Facility-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-CEDS_Facility-initialize}{}}}
\subsection{\code{CEDS_Facility$new()}}{
  Initializes the instance of CEDSEntity This can then be fed into other CEDS r6 objects.
This method is called by the \code{initialize} of other CEDS objects, like \code{CEDSFacility}
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CEDS_Facility$new(datasource, entity_type, pkid = NA, config = list())}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{datasource}}{\code{CEDSDataSource} object, which serves as the connection for this object
to ODS}
      \item{\code{entity_type}}{A character string specifying what the feature of interest is. Could
be "facility" or the type of permit ("WWR","VWP","GWP","VPDES")}
      \item{\code{pkid}}{The CEDS ID of the entity of interest. Must be the ID of the entity itself,
so the facility ID for a facility the permit ID of the permit.  This or config are required to
find the entity If pkid is provided, config is ignored}
      \item{\code{config}}{A list of column name value pairs to search for the entity by. These must be
column names from the entity table, else it will return an error}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    object instance of CEDSEntity
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CEDS_Facility-get_mps"></a>}}
\if{latex}{\out{\hypertarget{method-CEDS_Facility-get_mps}{}}}
\subsection{\code{CEDS_Facility$get_mps()}}{
  Pulls the measuring points of a withdrawal permit. This method takes apermit CEDS ID and
returns the row of the measuring point view of all measuring points of the desired source type
(specified). This method only returns MP information, notwithdrawals.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CEDS_Facility$get_mps(permit_id, source = c("SW/GW", "SW", "GW", "All"))}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{permit_id}}{The permit CEDS ID. This can be either a VWP, GWP, or WWR. But it does not
matter which type, since the measuring point table has a permit ID column agnostic of type}
      \item{\code{source}}{The desired measuring points to pull. By default it is SW/GW (surface water +
groundwater). This can be limited to one or the other, or "All", which includes transfers}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A data.frame of the rows from the measuring point view for MPs of the selected
source type for that permit.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CEDS_Facility-get_contacts"></a>}}
\if{latex}{\out{\hypertarget{method-CEDS_Facility-get_contacts}{}}}
\subsection{\code{CEDS_Facility$get_contacts()}}{
  Pulls contacts associated with the entity. For everything but WWR, this is the contacts associated
directly with that entity. For WWR, it is the contacts of the associated facility (since they cannot
be directly attached to a WWR). Each entity has its own contact table, so the names are standardized
in this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CEDS_Facility$get_contacts(limit = TRUE, return_df = TRUE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{limit}}{Should the result be limited to "Water WIthdrawal Contacts"? This only applies to
facilities and WWRs, since permit contacts often have other 'Contact Purposes'. Can be set to
FALSE to include all facility contacts}
      \item{\code{return_df}}{Should a data.frame be returned. Defaults to TRUE}
      \item{\code{pkid}}{The CEDS ID of the entity of interest. Must be the ID of the entity itself,
so the facility ID for a facility the permit ID of the permit.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Data.frame of contacts for the entity and their contact information
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CEDS_Facility-get_withdrawals"></a>}}
\if{latex}{\out{\hypertarget{method-CEDS_Facility-get_withdrawals}{}}}
\subsection{\code{CEDS_Facility$get_withdrawals()}}{
  Pulls the withdrawals from the water.Water_Use_By_Month_Vw for the measuring points on
this entity (as pulled by \code{get_mps}). These withdrawals show either monthly or yearly
use of the all mps in the facility/permit in millions of gallons. This is pulled by the
same method the foundation dataset is pulled, just paired down to only MP/withdrawal fields.
Since this is pulled via a live connection to ODS (still 24 hours behind CEDS), it may be
different from the foundation datset. All arguments are optional and it can run without them,
they just change the format of the returned data.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CEDS_Facility$get_withdrawals(
  years = FALSE,
  period = "yearly",
  return_df = TRUE
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{years}}{Lets you specify years of interest. By default set as FALSE, which will show the
entire period of record. Specifying years (as a numeric vector, i.e. 2015:2025) will only pull
data for those specified years. For entities with many MPs, this can speed up performance.}
      \item{\code{period}}{This will control how the data is aggregated. The data is reported monthly, so by
default there will be no aggregation applied. However, if \code{period} is set to "yearly", then
it will be aggregated by year, showing the total withdrawal of that MP for the given year}
      \item{\code{return_df}}{Should a data.frame be returned. Defaults to TRUE}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Data.frame of the withdrawals (in millions of gallons) for the MPs of the entity
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-CEDS_Facility-clone"></a>}}
\if{latex}{\out{\hypertarget{method-CEDS_Facility-clone}{}}}
\subsection{\code{CEDS_Facility$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{CEDS_Facility$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
