% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cia_utils.R
\name{CIA_data}
\alias{CIA_data}
\title{Cumulative Impact Analysis Data Function}
\usage{
CIA_data(riv_seg, runid1, runid2, flow_metric, AllSegList)
}
\arguments{
\item{riv_seg}{Desired river segment}

\item{runid1}{First runid of interest}

\item{runid2}{Second runid of interest}

\item{flow_metric}{Desired flow metric}

\item{AllSegList}{A list of all river segments}
}
\value{
data frame of river segments downstream and upstream of inputed segment
}
\description{
Extracts metric data from all upstream river segments and downstream segments from inputed segment
}
