% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dHVariablePluginDefault.R
\name{dHOMAlphanumericConstant}
\alias{dHOMAlphanumericConstant}
\title{dHOMAlphanumericConstant}
\value{
reference class of type openmi.om.base.

an updated config if necessary or FALSE if it fails
}
\description{
Simple class to hold string values
}
\details{
Text Constant meta-model object

Has standard methods for managing data and meta data
}
\examples{
NA
}
\seealso{
NA
}
\section{Super class}{
\code{\link[hydrotools:dHVariablePluginDefault]{hydrotools::dHVariablePluginDefault}} -> \code{dHOMAlphanumericConstant}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{what is it called}

\item{\code{object_class}}{model object type}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-dHOMAlphanumericConstant-new}{\code{dHOMAlphanumericConstant$new()}}
\item \href{#method-dHOMAlphanumericConstant-exportOpenMIBase}{\code{dHOMAlphanumericConstant$exportOpenMIBase()}}
\item \href{#method-dHOMAlphanumericConstant-clone}{\code{dHOMAlphanumericConstant$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="dHVariablePluginDefault" data-id="exportOpenMI"><a href='../../hydrotools/html/dHVariablePluginDefault.html#method-dHVariablePluginDefault-exportOpenMI'><code>hydrotools::dHVariablePluginDefault$exportOpenMI()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="dHVariablePluginDefault" data-id="fromOpenMIBase"><a href='../../hydrotools/html/dHVariablePluginDefault.html#method-dHVariablePluginDefault-fromOpenMIBase'><code>hydrotools::dHVariablePluginDefault$fromOpenMIBase()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="dHVariablePluginDefault" data-id="param_info"><a href='../../hydrotools/html/dHVariablePluginDefault.html#method-dHVariablePluginDefault-param_info'><code>hydrotools::dHVariablePluginDefault$param_info()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dHOMAlphanumericConstant-new"></a>}}
\if{latex}{\out{\hypertarget{method-dHOMAlphanumericConstant-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dHOMAlphanumericConstant$new(config = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{list of attributes to set, see also: to_list() for format}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
object instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dHOMAlphanumericConstant-exportOpenMIBase"></a>}}
\if{latex}{\out{\hypertarget{method-dHOMAlphanumericConstant-exportOpenMIBase}{}}}
\subsection{Method \code{exportOpenMIBase()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dHOMAlphanumericConstant$exportOpenMIBase(entity)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entity}}{the local object to work on}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dHOMAlphanumericConstant-clone"></a>}}
\if{latex}{\out{\hypertarget{method-dHOMAlphanumericConstant-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dHOMAlphanumericConstant$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
