% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dHVariablePluginDefault.R
\name{dHOMHydroImpoundment}
\alias{dHOMHydroImpoundment}
\title{dHOMHydroImpoundment}
\value{
reference class of type openmi.om.base.
}
\description{
Simple class to deal with hydro impoundment class model objects
}
\details{
VA Hydro model impoundment plugin

Has standard methods for exporting impoundment properties and subproperties
}
\examples{
NA
}
\seealso{
NA
}
\section{Super classes}{
\code{\link[hydrotools:dHVariablePluginDefault]{hydrotools::dHVariablePluginDefault}} -> \code{\link[hydrotools:dHOMDataMatrix]{hydrotools::dHOMDataMatrix}} -> \code{dHOMHydroImpoundment}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{what is it called}

\item{\code{object_class}}{model object type}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-dHOMHydroImpoundment-exportOpenMIBase}{\code{dHOMHydroImpoundment$exportOpenMIBase()}}
\item \href{#method-dHOMHydroImpoundment-clone}{\code{dHOMHydroImpoundment$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="dHVariablePluginDefault" data-id="exportOpenMI"><a href='../../hydrotools/html/dHVariablePluginDefault.html#method-dHVariablePluginDefault-exportOpenMI'><code>hydrotools::dHVariablePluginDefault$exportOpenMI()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="dHVariablePluginDefault" data-id="fromOpenMIBase"><a href='../../hydrotools/html/dHVariablePluginDefault.html#method-dHVariablePluginDefault-fromOpenMIBase'><code>hydrotools::dHVariablePluginDefault$fromOpenMIBase()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="dHVariablePluginDefault" data-id="initialize"><a href='../../hydrotools/html/dHVariablePluginDefault.html#method-dHVariablePluginDefault-initialize'><code>hydrotools::dHVariablePluginDefault$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="dHOMDataMatrix" data-id="param_info"><a href='../../hydrotools/html/dHOMDataMatrix.html#method-dHOMDataMatrix-param_info'><code>hydrotools::dHOMDataMatrix$param_info()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dHOMHydroImpoundment-exportOpenMIBase"></a>}}
\if{latex}{\out{\hypertarget{method-dHOMHydroImpoundment-exportOpenMIBase}{}}}
\subsection{Method \code{exportOpenMIBase()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dHOMHydroImpoundment$exportOpenMIBase(entity)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entity}}{the local object to work on}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an updated config if necessary or FALSE if it fails
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dHOMHydroImpoundment-clone"></a>}}
\if{latex}{\out{\hypertarget{method-dHOMHydroImpoundment-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dHOMHydroImpoundment$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
