% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dHVariablePluginDefault.R
\name{dHOMHydroImpoundmentSmall}
\alias{dHOMHydroImpoundmentSmall}
\title{dHOMHydroImpoundmentSmall}
\value{
reference class of type openmi.om.base.
}
\description{
Simple class to hold tabular flow by values
}
\details{
hydroimp

Has standard methods for managing data and meta data
}
\examples{
NA
}
\seealso{
NA
}
\section{Super classes}{
\code{\link[hydrotools:dHVariablePluginDefault]{hydrotools::dHVariablePluginDefault}} -> \code{\link[hydrotools:dHOMDataMatrix]{hydrotools::dHOMDataMatrix}} -> \code{\link[hydrotools:dHOMHydroImpoundment]{hydrotools::dHOMHydroImpoundment}} -> \code{dHOMHydroImpoundmentSmall}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{what is it called}

\item{\code{object_class}}{model object type}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-dHOMHydroImpoundmentSmall-exportOpenMIBase}{\code{dHOMHydroImpoundmentSmall$exportOpenMIBase()}}
\item \href{#method-dHOMHydroImpoundmentSmall-exportOpenMIpost}{\code{dHOMHydroImpoundmentSmall$exportOpenMIpost()}}
\item \href{#method-dHOMHydroImpoundmentSmall-clone}{\code{dHOMHydroImpoundmentSmall$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="dHVariablePluginDefault" data-id="exportOpenMI"><a href='../../hydrotools/html/dHVariablePluginDefault.html#method-dHVariablePluginDefault-exportOpenMI'><code>hydrotools::dHVariablePluginDefault$exportOpenMI()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="dHVariablePluginDefault" data-id="fromOpenMIBase"><a href='../../hydrotools/html/dHVariablePluginDefault.html#method-dHVariablePluginDefault-fromOpenMIBase'><code>hydrotools::dHVariablePluginDefault$fromOpenMIBase()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="dHVariablePluginDefault" data-id="initialize"><a href='../../hydrotools/html/dHVariablePluginDefault.html#method-dHVariablePluginDefault-initialize'><code>hydrotools::dHVariablePluginDefault$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="dHVariablePluginDefault" data-id="populate"><a href='../../hydrotools/html/dHVariablePluginDefault.html#method-dHVariablePluginDefault-populate'><code>hydrotools::dHVariablePluginDefault$populate()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="dHOMDataMatrix" data-id="param_info"><a href='../../hydrotools/html/dHOMDataMatrix.html#method-dHOMDataMatrix-param_info'><code>hydrotools::dHOMDataMatrix$param_info()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dHOMHydroImpoundmentSmall-exportOpenMIBase"></a>}}
\if{latex}{\out{\hypertarget{method-dHOMHydroImpoundmentSmall-exportOpenMIBase}{}}}
\subsection{Method \code{exportOpenMIBase()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dHOMHydroImpoundmentSmall$exportOpenMIBase(entity)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entity}}{the local object to work on}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
an updated config if necessary or FALSE if it fails
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dHOMHydroImpoundmentSmall-exportOpenMIpost"></a>}}
\if{latex}{\out{\hypertarget{method-dHOMHydroImpoundmentSmall-exportOpenMIpost}{}}}
\subsection{Method \code{exportOpenMIpost()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dHOMHydroImpoundmentSmall$exportOpenMIpost(entity, export = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entity}}{the local object to work on}

\item{\code{export}}{current export object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Returns a modified list of property and children
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dHOMHydroImpoundmentSmall-clone"></a>}}
\if{latex}{\out{\hypertarget{method-dHOMHydroImpoundmentSmall-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dHOMHydroImpoundmentSmall$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
