% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dHVariablePluginDefault.R
\name{dHOMObjectClass}
\alias{dHOMObjectClass}
\title{Object class of meta-model object}
\value{
reference class of type openmi.om.base.
}
\description{
Simple class to hold text values
}
\details{
Has standard methods for managing data and meta data
}
\examples{
NA
}
\seealso{
NA
}
\section{Super classes}{
\code{\link[hydrotools:dHVariablePluginDefault]{hydrotools::dHVariablePluginDefault}} -> \code{hydrotools::dHOMAlphanumericConstant} -> \code{dHOMObjectClass}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-dHOMObjectClass-clone}{\code{dHOMObjectClass$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="dHVariablePluginDefault" data-id="exportOpenMI"><a href='../../hydrotools/html/dHVariablePluginDefault.html#method-dHVariablePluginDefault-exportOpenMI'><code>hydrotools::dHVariablePluginDefault$exportOpenMI()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="dHOMAlphanumericConstant" data-id="exportOpenMIBase"><a href='../../hydrotools/html/dHOMAlphanumericConstant.html#method-dHOMAlphanumericConstant-exportOpenMIBase'><code>hydrotools::dHOMAlphanumericConstant$exportOpenMIBase()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="dHOMAlphanumericConstant" data-id="initialize"><a href='../../hydrotools/html/dHOMAlphanumericConstant.html#method-dHOMAlphanumericConstant-initialize'><code>hydrotools::dHOMAlphanumericConstant$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dHOMObjectClass-clone"></a>}}
\if{latex}{\out{\hypertarget{method-dHOMObjectClass-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dHOMObjectClass$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
