% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dHVariablePluginDefault.R
\name{dHOMPublicVars}
\alias{dHOMPublicVars}
\title{dHOMPublicVars}
\value{
reference class of type openmi.om.base.
}
\description{
Text that can be any of public vars list
}
\details{
Has standard methods for managing data and meta data
}
\examples{
NA
}
\seealso{
NA
}
\section{Super classes}{
\code{\link[hydrotools:dHVariablePluginDefault]{hydrotools::dHVariablePluginDefault}} -> \code{\link[hydrotools:dHOMAlphanumericConstant]{hydrotools::dHOMAlphanumericConstant}} -> \code{dHVarImage}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-dHVarImage-clone}{\code{dHOMPublicVars$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="dHVariablePluginDefault" data-id="exportOpenMI"><a href='../../hydrotools/html/dHVariablePluginDefault.html#method-dHVariablePluginDefault-exportOpenMI'><code>hydrotools::dHVariablePluginDefault$exportOpenMI()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="dHVariablePluginDefault" data-id="fromOpenMIBase"><a href='../../hydrotools/html/dHVariablePluginDefault.html#method-dHVariablePluginDefault-fromOpenMIBase'><code>hydrotools::dHVariablePluginDefault$fromOpenMIBase()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="dHOMAlphanumericConstant" data-id="exportOpenMIBase"><a href='../../hydrotools/html/dHOMAlphanumericConstant.html#method-dHOMAlphanumericConstant-exportOpenMIBase'><code>hydrotools::dHOMAlphanumericConstant$exportOpenMIBase()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="dHOMAlphanumericConstant" data-id="initialize"><a href='../../hydrotools/html/dHOMAlphanumericConstant.html#method-dHOMAlphanumericConstant-initialize'><code>hydrotools::dHOMAlphanumericConstant$initialize()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-dHVarImage-clone"></a>}}
\if{latex}{\out{\hypertarget{method-dHVarImage-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{dHOMPublicVars$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
