% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{find_name}
\alias{find_name}
\title{Find a name in a nested list. TODO: move to utils.R or in OM (where its used)}
\usage{
find_name(haystack, needle)
}
\arguments{
\item{haystack}{list}

\item{needle}{object name to search for}
}
\value{
Object found in list
}
\description{
Find a name in a nested list. TODO: move to utils.R or in OM (where its used)
}
\examples{
NA
}
