% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cia_utils.R
\name{fn_extract_basin}
\alias{fn_extract_basin}
\title{Extracting Basin Function}
\usage{
fn_extract_basin(cia_data_frame, end_seg)
}
\arguments{
\item{cia_data_frame}{Data frame of cumulative impact data. This must contain
a column with the name "riverseg" that contains all river segments of
interest. This function can ONLY find upstream segments if they are present
in this riverseg column!}

\item{end_seg}{Desired end river segment - river segment that is most downstream}
}
\value{
Trimmed cumulative impact data frame containing the outlet and all upstream segments
}
\description{
Trims data frame to include only segments upstream of end
  segment. It is inclusive of the end segment and will thus include the
  outlet and all upstream segments
}
