% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fn_get_rest}
\alias{fn_get_rest}
\title{Get any entity from a RESTful web service}
\usage{
fn_get_rest(entity_type, pk, inputs, site, token)
}
\arguments{
\item{entity_type}{= dh_feature, dh_properties, ...}

\item{pk}{= primary key column name, e.g. hydroid, pid, ...}

\item{inputs}{contents of record to get in list(pid, propname, propvalue, ...)}

\item{site}{URL of rest server}

\item{token}{for xhttp auth}
}
\description{
Get any entity from a RESTful web service
}
