% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fn_get_rundata}
\alias{fn_get_rundata}
\title{Retrieve Run summary data from Old OM model}
\usage{
fn_get_rundata(
  elementid = -1,
  runid = -1,
  varname = "Qout",
  scenid = 37,
  site = "http://deq2.bse.vt.edu"
)
}
\arguments{
\item{elementid}{integer OM model element id}

\item{runid}{integer run id}

\item{varname}{character - if set will isolate a single column of data}

\item{scenid}{integer - model domain ID}

\item{site}{URL of om server}
}
\value{
reference class of type openmi.om.equation
}
\description{
Retrieve Run summary data from Old OM model
}
\examples{
NA
}
\seealso{
NA
}
