% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fn_get_runfile_info}
\alias{fn_get_runfile_info}
\title{Retrieve Info About Run File from Old OM model}
\usage{
fn_get_runfile_info(
  elementid = -1,
  runid = -1,
  scenid = 37,
  site = "http://deq2.bse.vt.edu"
)
}
\arguments{
\item{elementid}{integer OM model element id}

\item{runid}{integer run id}

\item{scenid}{integer - model domain ID}

\item{site}{URL of om server}
}
\value{
reference class of type openmi.om.equation
}
\description{
Retrieve Info About Run File from Old OM model
}
\examples{
NA
}
\seealso{
NA
}
