% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fn_get_vardef_view}
\alias{fn_get_vardef_view}
\title{Retrieve Variable Def from View (not REST)}
\usage{
fn_get_vardef_view(varkey, site, token, debug = FALSE)
}
\arguments{
\item{varkey}{character variable key}

\item{site}{URL of om server}

\item{token}{for xhttp auth}

\item{debug}{show debugging info}
}
\value{
variable definition as list
}
\description{
Retrieve Variable Def from View (not REST)
}
\examples{
NA
}
\seealso{
NA
}
