% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fn_getVarDefView}
\alias{fn_getVarDefView}
\title{Retrieve Variable Def from View (not REST)}
\usage{
fn_getVarDefView(site, varkey, debug = FALSE)
}
\arguments{
\item{site}{URL of om server}

\item{varkey}{character variable key}
}
\value{
integer variable id
}
\description{
Retrieve Variable Def from View (not REST)
}
\examples{
NA
}
\seealso{
NA
}
