% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cia_utils.R
\name{fn_handletimestamp}
\alias{fn_handletimestamp}
\title{Handle Any Time/Date to Timestamp}
\usage{
fn_handletimestamp(tslike, format = "epoch")
}
\arguments{
\item{tslike}{some type of date or timestamp}

\item{format}{TBD use this to allow alternative return formats}
}
\value{
single numeric value for UNix epoch
}
\description{
Try to bring any date or timestamp into a common format (Unix epoch, TBD)
}
