% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cia_utils.R
\name{fn_iha_flow_extreme}
\alias{fn_iha_flow_extreme}
\title{fn_iha_flow_extreme}
\usage{
fn_iha_flow_extreme(flows, metric, stat = "min", wyear_type = "calendar")
}
\arguments{
\item{flows}{a timeseries flormatted in zoo (required by IHA)}

\item{metric}{the flow period and type (min/max)}

\item{stat}{default=min, options: max, median}

\item{wyear_type}{water year default=calendar, options: water}
}
\value{
singel numeric value for the selected index/stat
}
\description{
provide the quantile of annual period
}
