% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cia_utils.R
\name{fn_iha_mlf}
\alias{fn_iha_mlf}
\title{fn_iha_mlf}
\usage{
fn_iha_mlf(zoots, targetmo, q = 0.5)
}
\arguments{
\item{zoots}{a timeseries flormatted in zoo (required by IHA)}

\item{targetmo}{month in title case string}

\item{q}{numeric what quantile to return (default median/0.5)}
}
\value{
singel numeric value for 7q10
}
\description{
provide the quantile of minimum observed monthly flow for the month over a period of years
}
