% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_plot_cia_dend.R
\name{fn_plot_cia_dend}
\alias{fn_plot_cia_dend}
\title{Dendritic Ploting Function}
\usage{
fn_plot_cia_dend(
  riv_seg,
  AllSegList,
  runid1,
  runid2,
  flow_metric,
  cia_data_frame
)
}
\arguments{
\item{riv_seg}{Downstream river segment that you wish to have no more tributaries below segment - cannot be any more downstream then segment entered into function}

\item{AllSegList}{List of all vahydro river segments}

\item{runid1}{Runid that corresponds to first metric in data frame}

\item{runid2}{Runid that corresponds to second metric in data frame}

\item{flow_metric}{Flow metric that corresponds to data in data frame}

\item{cia_data_frame}{Data frame of cumulative impact data}
}
\value{
A dendritic plot of segments from data frame
}
\description{
Helps locate all upstream segments from the working segment
}
