% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fn_remove_model_warmup}
\alias{fn_remove_model_warmup}
\title{Truncate run data to best guess for warmup and water year periods}
\usage{
fn_remove_model_warmup(dat)
}
\arguments{
\item{dat}{the data timeseries to handle}
}
\value{
a shortened time series
}
\description{
Truncate run data to best guess for warmup and water year periods
}
\examples{
NA
}
\seealso{
NA
}
