% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cia_utils.R
\name{fn_river_network}
\alias{fn_river_network}
\title{fn_river_network()}
\usage{
fn_river_network(riv_seg, AllSegList, cia_data_frame)
}
\arguments{
\item{riv_seg}{Desired river segment}

\item{AllSegList}{List of all vahydro river segments}

\item{cia_data_frame}{CIA data frame with specific columns}
}
\value{
data frame of river segments, and associated river miles
}
\description{
Calculates ands adds river mile column to data frame of river segments
}
