% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fn_search_properties}
\alias{fn_search_properties}
\title{Retrieve Property data from propvalues style data frame}
\usage{
fn_search_properties(config, propvalues_tmp, multiplicity = "default")
}
\arguments{
\item{config}{= list(entity_type, featureid, pid = NULL, varid = NULL, startdate = NULL, enddate = NULL, tscode = NULL, tlid = NULL) timeline ID (not yet used)}

\item{propvalues_tmp}{data frame to search}

\item{multiplicity}{uniqueness criteria. default = startdate_singular which is varid + startdate (all are varid singular)}
}
\value{
data frame of propvalue or FALSE
}
\description{
Retrieve Property data from propvalues style data frame
}
\examples{
NA
}
\seealso{
NA
}
