% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fn_search_tsvalues}
\alias{fn_search_tsvalues}
\title{Retrieve TS data from tsvalues style data frame}
\usage{
fn_search_tsvalues(config, tsvalues_tmp)
}
\arguments{
\item{config}{= list(entity_type, featureid, tid = NULL, varid = NULL, tstime = NULL, tsendtime = NULL, tscode = NULL, tlid = NULL) timeline ID (not yet used)}

\item{tsvalues_tmp}{data frame to search}
}
\value{
data frame of tsvalue or FALSE
}
\description{
Retrieve TS data from tsvalues style data frame
}
\examples{
NA
}
\seealso{
NA
}
