% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fn_search_vardefs}
\alias{fn_search_vardefs}
\title{Retrieve Property data from propvalues style data frame}
\usage{
fn_search_vardefs(config, var_defs_tmp)
}
\arguments{
\item{config}{= list(entity_type, featureid, pid = NULL, varid = NULL, startdate = NULL, enddate = NULL, tscode = NULL, tlid = NULL) timeline ID (not yet used)}

\item{var_defs_tmp}{data frame to search}
}
\value{
data frame of propvalue or FALSE
}
\description{
Retrieve Property data from propvalues style data frame
}
\examples{
NA
}
\seealso{
NA
}
