% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cia_utils.R
\name{fn_upstream}
\alias{fn_upstream}
\title{Find Upstream Segments}
\usage{
fn_upstream(riv.seg, AllSegList)
}
\arguments{
\item{riv.seg}{string ID of segment to fine}

\item{AllSegList}{AllSegList data frame of all segments}
}
\value{
string next upstream segment ID
}
\description{
Trims data frame to include only segments upstream of end segment
}
