% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cia_utils.R
\name{get_all_riversegs}
\alias{get_all_riversegs}
\title{Get data frame of all dbase river segments}
\usage{
get_all_riversegs(ds)
}
\arguments{
\item{ds}{A RomDataSource to allow for querying of additional
information. May be provided by OWS config files.}
}
\value{
data frame with riverseg, hydroid, hydrocode, name from dh_feature
}
\description{
Query datasource for all river segments
}
\details{
A simple query that extracts river segment IDs from dh_feature based
on bundle = 'watershed', ftype = 'vahydro', and hydrocode ilike
'vahydrosw_wshed_\%'
}
