% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dHVariablePluginDefault.R
\name{get_plugin_class}
\alias{get_plugin_class}
\title{get_plugin_class}
\usage{
get_plugin_class(plugin_name, entity)
}
\arguments{
\item{plugin_name}{the actual class name}

\item{entity}{the object to apply the plugin to}
}
\value{
class matching plugin or default generic plugin
}
\description{
Retrieve Plugin object for a variable entity
}
\examples{
NA
}
\seealso{
NA
}
