% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wshd_utils.R
\name{get_weather_raster_data}
\alias{get_weather_raster_data}
\title{get_weather_raster_data}
\usage{
get_weather_raster_data(
  start_epoch = -1,
  end_epoch = as.numeric(Sys.time()),
  hydroid = NULL,
  bundle = "watershed",
  ftype = "usgs_full_drainage",
  varkey = "prism_mod_daily",
  band = 1,
  metric = "mean",
  touched = FALSE
)
}
\arguments{
\item{start_epoch}{A start time (seconds after epoch) to begin the raster analysis}

\item{end_epoch}{A end time (seconds after epoch) to begin the raster analysis}

\item{hydroid}{A vector of hydroids that may be used to identify features.
Alternatively, users may use bundle AND ftype instead}

\item{bundle}{If no hydroid is input, the user can identify features using a
bundle (the general type of entity) and ftype (entity sub-type) to query
dh_feature. Both bundle and ftype are mandatory for this method.}

\item{ftype}{If no hydroid is input, the user can identify features using a
bundle (the general type of entity) and ftype (entity sub-type) to query
dh_feature. Both bundle and ftype are mandatory for this method.}

\item{varkey}{The raster data set of interest. These varkeys are stored in
dh_variabledefinition. defaults to 'prism_mod_daily'}

\item{band}{The raster band to summarize, defaults to 1}

\item{metric}{Character. What kind of raster aggregate summary should be
applied? Defaults to mean. These are taken from PostGIS
\code{ST_SummaryStats()}}

\item{touched}{Defaults to FALSE. Should PostGIS ST_Clip() use the touched
argument to return all raster cells touched by the feature? Or, if FALSE,
should only pixels that have centroids within the feature be considered?}
}
\value{
A list with parameters for the gage_vs_model.Rmd markdown
}
\description{
Summary dh_timeseries_weather raster for one or many entities in
 dh_feature
}
\details{
Summarize raster data over one or many features as identified by the
 user. Users may identify this feature through a vector of hydroids or
 through a general bundle and ftype. By default, this will get the mean value
 of the raster each day.
}
