% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFLOW_xQy.R
\name{getmode}
\alias{getmode}
\title{getmode}
\usage{
getmode(v)
}
\arguments{
\item{v}{A vector}
}
\value{
A vector with the values that are most frequently used in the table
}
\description{
Find the mode of a vector
}
\details{
Find the mode of any kind of vector. In the case of ties, it will
  return all values that tied in frequency
}
