% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TNC_IHA_Ports.R
\name{group2Funs}
\alias{group2Funs}
\title{Calculates group2 statistics from a matrix of rolling means}
\usage{
group2Funs(x)
}
\arguments{
\item{x}{a matrix of rolling means}
}
\value{
a named vector containing the group2 statistics, including the range
of the 1-, 3-, 7-, 30-, and 90-day flows and the base index and days of zero
flow
}
\description{
Calculates group2 statistics from a matrix of rolling means. This include
range, base index (minimum 7-day flow divided by mean daily flow), and the
number of days with zero flow.
}
\author{
jason.e.law@gmail.com (imported to Hydrotools by Connor Brogan,connor.brogan@deq.virginia.gov)
}
