% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cia_utils.R
\name{is.empty}
\alias{is.empty}
\title{is.empty}
\usage{
is.empty(x, ...)
}
\arguments{
\item{x}{an object to check its emptiness.}

\item{...}{additional arguments for sapply}
}
\description{
Checks common NULL, NA, length cases for empty values
}
\details{
Modified from rapportools as the check for a numeric value of 0 was removed.
From rapportools: "Rails-inspired helper that checks if vector values are
"empty", i.e. if it's: NULL, zero-length, NA, NaN, FALSE, an empty string ...
Note that unlike its native R is.\if{html}{\out{<something>}} sibling functions, is.empty
is vectorised (hence the "values")."
}
\examples{
is.empty(data.frame())
is.empty(data.frame(1:2,c(3, NA)))
is.empty(c("",NA, 1, 2,3, "Im Not Empty!", 4))
is.empty(c(TRUE, FALSE, TRUE))
}
