% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFLOW_xQy.R
\name{isLeapYear}
\alias{isLeapYear}
\title{isLeapYear}
\usage{
isLeapYear(inYears)
}
\arguments{
\item{inYears}{A numeric vector representing years to check}
}
\value{
Logical vector, TRUE or FALSE
}
\description{
Check vector for leap years
}
\details{
lapply wrapper for leapYearAssessment(). Used to determine which
  years in a vector of years are leap years
}
