% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFLOW_xQy.R
\name{Ky_Qp3}
\alias{Ky_Qp3}
\title{Ky_Qp3}
\usage{
Ky_Qp3(g, aep)
}
\arguments{
\item{g}{Skewness}

\item{aep}{Exceedance probability}
}
\value{
Qauntile of Peason III distribution for aep based on skewness g
}
\description{
Pearson III quantile
}
\details{
Return the quantile of the Pearson III distribution given skewness
  of g, and annual exceedance probability, aep. This is a wrapper for lmomco
  package functions.
}
