% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFLOW_xQy.R
\name{leapYearAssessment}
\alias{leapYearAssessment}
\title{leapYearAssessment}
\usage{
leapYearAssessment(inYear)
}
\arguments{
\item{inYear}{A single numeric representing a year}
}
\value{
Logical, TRUE or FALSE
}
\description{
Check if year is leap year
}
\details{
Check if the numeric year is a leap year i.e., is the year divisible
  by 4 but not 100, unless it is also divisible by 400?
}
