% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_handlers.R
\name{ModelElementBase}
\alias{ModelElementBase}
\title{ModelElementBase}
\value{
R6Class of type ModelElementBase
}
\description{
Utility class for interacting with a watershed feature/model combo
}
\details{
Watershed Model Node data object

Has standard methods for managing data and meta data
}
\examples{
NA
}
\seealso{
NA
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ds}}{RomDataSource}

\item{\code{ds_om}}{RomDataSource points to legacy model db}

\item{\code{hydroid}}{unique ID from dh_feature database}

\item{\code{hydrocode}}{non-numeric identifier may reference ext source db}

\item{\code{bundle}}{type of feature}

\item{\code{ftype}}{sub-type of feature}

\item{\code{feature}}{a RomFeature entity}

\item{\code{pid}}{integer identifier}

\item{\code{version}}{descriptive code for model}

\item{\code{prop}}{as RomProperty}

\item{\code{json}}{as list}

\item{\code{elementid}}{as integer}

\item{\code{rocode}}{describes runoff type default = 'cbp6_lrseg'}

\item{\code{name}}{what is it called}

\item{\code{riverseg}}{we using code names}

\item{\code{site}}{URI base for the legacy model scripts assumes global var omsite exists for default}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ModelElementBase-new}{\code{ModelElementBase$new()}}
\item \href{#method-ModelElementBase-handle_config}{\code{ModelElementBase$handle_config()}}
\item \href{#method-ModelElementBase-get_feature}{\code{ModelElementBase$get_feature()}}
\item \href{#method-ModelElementBase-get_model_pid}{\code{ModelElementBase$get_model_pid()}}
\item \href{#method-ModelElementBase-get_model}{\code{ModelElementBase$get_model()}}
\item \href{#method-ModelElementBase-get_run_data}{\code{ModelElementBase$get_run_data()}}
\item \href{#method-ModelElementBase-load_json_model}{\code{ModelElementBase$load_json_model()}}
\item \href{#method-ModelElementBase-get_json_model}{\code{ModelElementBase$get_json_model()}}
\item \href{#method-ModelElementBase-om_element_tree}{\code{ModelElementBase$om_element_tree()}}
\item \href{#method-ModelElementBase-render_cia_brief}{\code{ModelElementBase$render_cia_brief()}}
\item \href{#method-ModelElementBase-clone}{\code{ModelElementBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelElementBase-new"></a>}}
\if{latex}{\out{\hypertarget{method-ModelElementBase-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelElementBase$new(ds, config = list(), ds_om = NA, site = omsite)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ds}}{RomDataSource object mandatory}

\item{\code{config}}{list of attributes to set, see also: to_list() for format}

\item{\code{ds_om}}{RomDataSource object pointing to OM (optional)}

\item{\code{site}}{URI base for the legacy model scripts assumes global var omsite exists for default}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
object instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelElementBase-handle_config"></a>}}
\if{latex}{\out{\hypertarget{method-ModelElementBase-handle_config}{}}}
\subsection{Method \code{handle_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelElementBase$handle_config(config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{list of attributes to set/query}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
json of model from dH
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelElementBase-get_feature"></a>}}
\if{latex}{\out{\hypertarget{method-ModelElementBase-get_feature}{}}}
\subsection{Method \code{get_feature()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelElementBase$get_feature()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
json of model from dH
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelElementBase-get_model_pid"></a>}}
\if{latex}{\out{\hypertarget{method-ModelElementBase-get_model_pid}{}}}
\subsection{Method \code{get_model_pid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelElementBase$get_model_pid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
json of model from dH
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelElementBase-get_model"></a>}}
\if{latex}{\out{\hypertarget{method-ModelElementBase-get_model}{}}}
\subsection{Method \code{get_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelElementBase$get_model()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
json of model from dH
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelElementBase-get_run_data"></a>}}
\if{latex}{\out{\hypertarget{method-ModelElementBase-get_run_data}{}}}
\subsection{Method \code{get_run_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelElementBase$get_run_data(
  runid,
  cached = FALSE,
  hydrowarmup = TRUE,
  cleanup = FALSE,
  outaszoo = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{runid}}{run scenario/id to retrieve}

\item{\code{cached}}{- use local cache? default FALSE}

\item{\code{hydrowarmup}}{truncate first 2-9 months?}

\item{\code{cleanup}}{Should the function delete the log file create for the cached argument?}

\item{\code{outaszoo}}{zoo is default, if FALSE returns dataframe}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
timeseries of run file
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelElementBase-load_json_model"></a>}}
\if{latex}{\out{\hypertarget{method-ModelElementBase-load_json_model}{}}}
\subsection{Method \code{load_json_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelElementBase$load_json_model()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TRUE or FALSE
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelElementBase-get_json_model"></a>}}
\if{latex}{\out{\hypertarget{method-ModelElementBase-get_json_model}{}}}
\subsection{Method \code{get_json_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelElementBase$get_json_model()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
json of model from dH
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelElementBase-om_element_tree"></a>}}
\if{latex}{\out{\hypertarget{method-ModelElementBase-om_element_tree}{}}}
\subsection{Method \code{om_element_tree()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelElementBase$om_element_tree(
  elementid,
  include_children = 1,
  max_recursion_level = -1,
  exclude_custom1 = -1,
  target_custom1 = -1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{elementid}}{target container to search}

\item{\code{include_children}}{recurse through children? default = 1}

\item{\code{max_recursion_level}}{integer how far to search if include_children = 1}

\item{\code{exclude_custom1}}{custom1 to exclude (use cova_upstream prevent upstream recursion)}

\item{\code{target_custom1}}{custom1 to find}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
dataframe of elementids
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelElementBase-render_cia_brief"></a>}}
\if{latex}{\out{\hypertarget{method-ModelElementBase-render_cia_brief}{}}}
\subsection{Method \code{render_cia_brief()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelElementBase$render_cia_brief(
  runid,
  export_path = export_path,
  github_location = github_location,
  cu_pre_var = "Qout",
  cu_post_var = "Qout",
  doc_title = "auto",
  cu_threshold = c(-10, -20, -30),
  include_appendices = c()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{runid}}{runid to summarize}

\item{\code{export_path}}{where to store the file}

\item{\code{github_location}}{library to load rendering code}

\item{\code{cu_pre_var}}{variable for "BEFORE" flow values}

\item{\code{cu_post_var}}{variable for "AFTER"}

\item{\code{doc_title}}{document title page text, default generated from feature}

\item{\code{cu_threshold}}{thresholds of change to code yellow, orange, red}

\item{\code{include_appendices}}{options 'hydropower', ...}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
file path of rendered output
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ModelElementBase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ModelElementBase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ModelElementBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
