% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_auth_read.R
\name{om_auth_read}
\alias{om_auth_read}
\title{Retrieve a token from vaydro REST service.  TBD: replace with object to store token and dn connection}
\usage{
om_auth_read(uri, token, ctype = "text/csv", delim = ",", enc = "xml")
}
\arguments{
\item{uri}{character deprecated to be replaced by global datasource}

\item{token}{for CSRF autentication use}

\item{ctype}{content-type}

\item{delim}{delimiter to expect in final data set}

\item{enc}{encoding}
}
\value{
data frame from delimited file
}
\description{
Retrieve a token from vaydro REST service.  TBD: replace with object to store token and dn connection
}
\examples{
NA
}
\seealso{
NA
}
