% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_cia_table.R
\name{om_cia_table}
\alias{om_cia_table}
\title{Generate a table of CIA model result summary statistsics for both Rseg and Facility models, comparing across scenarios}
\usage{
om_cia_table(
  rseg.hydroid = 462757,
  fac.hydroid = 72672,
  runid.list = c("runid_201", "runid_401"),
  fac.metric.list = c("wd_mgd", "ps_mgd", "unmet1_mgd", "unmet7_mgd", "unmet30_mgd",
    "unmet90_mgd"),
  rseg.metric.list = c("Qout", "Qbaseline", "remaining_days_p0", "remaining_days_p10",
    "remaining_days_p50", "l30_Qout", "l90_Qout", "consumptive_use_frac",
    "wd_cumulative_mgd", "ps_cumulative_mgd"),
  site = "https://deq1.bse.vt.edu/d.dh",
  site_base = "https://deq1.bse.vt.edu"
)
}
\arguments{
\item{rseg.hydroid}{riverseg dh_feature hydroid}

\item{fac.hydroid}{facility dh_feature hydroid}

\item{runid.list}{list of runids of interest}

\item{fac.metric.list}{list of facility metrics of interest}

\item{rseg.metric.list}{list of riverseg metrics of interest}

\item{site}{vhydro url}

\item{site_base}{base vahydro url}
}
\value{
stats.df dataframe of summary stats
}
\description{
Generate a table of CIA model result summary statistsics for both Rseg and Facility models, comparing across scenarios
}
\examples{
NA
}
\seealso{
NA
}
