% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fac_utils.R
\name{om_cu_table_data}
\alias{om_cu_table_data}
\title{Difference in monthly percentiles}
\usage{
om_cu_table_data(
  pr_data,
  cu_post_var,
  cu_pre_var,
  output_style = "percent",
  cu_decimals,
  cu_min_valid = 0
)
}
\arguments{
\item{pr_data}{A data frame that contains month and the user input
cu_post_var and cu_pre_var columns. This data frame is likely made up of
multiple model run files pulled via fn_get_runfile}

\item{cu_post_var}{The numeric column in pr_data that represents a "post"
alteration scenario i.e. Qout on a river node from run 600 in the OM}

\item{cu_pre_var}{The numeric column in pr_data that represents a "pre"
alteration scenario i.e. Qout on a river node from run 100 in the OM}

\item{output_style}{Should the differences in monthly percentiles be reported
as percentage change ("percent") from cu_pre_var or as the numeric
difference ("numeric")?}

\item{cu_decimals}{How many decimals should the output be rounded to?}

\item{cu_min_valid}{Minimum flow value to display, otherwise assumed below
model valid range as output will be NA}
}
\value{
A data frame that has either the percent or absolute difference in
  monthly percentile between cu_pre_var and cu_post_var depending on
  output_style. The month is included as the first column. Subsequent columns
  are those output by cu_flow_table i.e. minimum, 5%, 10%, 25%, 50%, and
  median
}
\description{
Using a data frame provided by user, find the monthly
  percentiles of cu_pre_var and cu_post_var using om_flow_table. Report the
  differences in these monthly percentiles or the percentage change (post to
  pre)
}
