% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cia_utils.R
\name{om_find_dh_elid}
\alias{om_find_dh_elid}
\title{Model element finder}
\usage{
om_find_dh_elid(elid, ds)
}
\arguments{
\item{elid}{Desired elementid}

\item{ds}{RomDataSource}
}
\value{
data frame of info about the target elementid
}
\description{
Finds the model, and its parent feature or prop given an OM elementid
}
