% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fac_utils.R
\name{om_flow_table}
\alias{om_flow_table}
\title{Display the monthly percentiles of flows from minimum to 50% and mean flows
of the given flow data}
\usage{
om_flow_table(df2sum, q_col = "Qout", mo_col = "month", rdigits = 1)
}
\arguments{
\item{df2sum}{A data frame that contains at least a column to be summarized
(q_col, typically flow) and a grouping column (mo_col, typically month)}

\item{q_col}{The name of the column in df2sum to be summarized and grouped by mo_col}

\item{mo_col}{A grouping column used to derive the percentiles of q_col}

\item{rdigits}{The number of digits to report/round percentiles to}
}
\value{
A data frame that has 8 columns: Month, Minimum, 5%, 10%, 25%, 30%,
  50%, Mean. These columns represent the correponsing percentile or statistic
  of q_col grouped by mo_col and rounded to rdigits. the names of the data
  frame correspond to these statistics. The data is organized by month.
}
\description{
Using a data frame provided by user, summarize the q_col to
  create an output that shows the minimum, 5 - 50%, and mean of that column,
  grouped by mo_col. The most common use is to quickly see flow percentiles
  of model or gage data across months
}
