% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_get_model_scenario.R
\name{om_get_model_scenario}
\alias{om_get_model_scenario}
\title{om_get_model_scenario Function: Retrieves a scenario for saving model data, creates scenario if needed}
\usage{
om_get_model_scenario(
  ds,
  model,
  scenario_name,
  model_varkey = "om_model_element"
)
}
\arguments{
\item{ds}{RomDataSource}

\item{model}{object as RomProperty}

\item{scenario_name}{Name of scenario to create/retrieve}

\item{model_varkey}{Variable key to use for model if one needs to be created, default=om_model_element}
}
\value{
scenario object as RomProperty
}
\description{
Retrieves a model scenario
}
