% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_vahydro_metric_grid.R
\name{om_metric_matrix}
\alias{om_metric_matrix}
\title{Build om_vahydro_metric_grid data frame input}
\usage{
om_metric_matrix(m_version, runids, metrics)
}
\arguments{
\item{m_version}{character vector. This will
be used to determine which models to find on \code{entity_type}}

\item{runids}{Either a character vector of model run ids or a dataframe with
model_version, runid, metric, and runlabel specified. The data frame is
generally deprecated in favor of using the individual arguments on this
function}

\item{metrics}{character vector of requested metrics}
}
\value{
data frame to provide to \code{om_vahydro_metric_grid()}
}
\description{
Helper function to generate old-style inputs for
 om_vahydro_metric_grid
}
\details{
Builds a data frame of runids, metrics, and runlabels to provide as
 the runids argument in \code{om_vahydro_metric_grid()}
}
\examples{
m_version <- c('vahydro-1.0')
runids <- c('runid_600','runid_400')
metrics <- c("WA_90_mgd", "Qavailable_90_mgd","Smin_L90_mg")
omgrid_df <- om_metric_matrix(m_version, runids, metrics) 
}
