% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_model_object.R
\name{om_model_object}
\alias{om_model_object}
\title{om_model_object Function: Retrieves a model object, creates if needed}
\usage{
om_model_object(
  ds,
  feature,
  model_version,
  model_name = FALSE,
  model_varkey = "om_model_element"
)
}
\arguments{
\item{ds}{RomDataSource}

\item{feature}{Feature object RomFeature}

\item{model_version}{Model version}

\item{model_name}{Model name, defaults to feature name + model_Version}

\item{model_varkey}{Variable key to use for model if one needs to be created, default=om_model_element}
}
\value{
model object as RomProperty
}
\description{
Retrieves a model object, creates if needed
}
