% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cia_utils.R
\name{om_model_run_monitor}
\alias{om_model_run_monitor}
\title{List of models that have been run or are running}
\usage{
om_model_run_monitor(ds_model, minspast = 60, elids = FALSE, limit = 100)
}
\arguments{
\item{ds_model}{RomDataSource connectio to model runtime database}

\item{minspast}{integer, number of minutes from present to limit results}

\item{elids}{Desired elementid(s) FALSE retrieves all}

\item{limit}{total number of records to return}
}
\value{
data frame of info about the target elementid
}
\description{
Finds the model, and its parent feature or prop given an OM elementid
}
