% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_model_table.R
\name{om_model_table}
\alias{om_model_table}
\title{Generate a table of CIA model result summary statistics for a model, comparing across scenarios}
\usage{
om_model_table(
  model_info = FALSE,
  runid.list = c("runid_6011", "runid_6012"),
  metric.list = c("Qout", "Qbaseline", "l30_Qout", "l90_Qout", "consumptive_use_frac",
    "wd_cumulative_mgd", "ps_cumulative_mgd"),
  include.elfgen = FALSE,
  site = "http://deq1.bse.vt.edu:81/d.dh",
  site_base = "http://deq1.bse.vt.edu:81",
  rdigits = 2
)
}
\arguments{
\item{model_info}{model json list}

\item{runid.list}{list of runids of interest}

\item{metric.list}{list of metrics of interest}

\item{include.elfgen}{include elfgen metrics TRUE or FALSE}

\item{site}{vhydro url}

\item{site_base}{base vahydro url}

\item{rdigits}{Number of digits to round}
}
\value{
stats.df dataframe of summary stats
}
\description{
Generate a table of CIA model result summary statistics for a model, comparing across scenarios
}
\examples{
NA
}
\seealso{
NA
}
