% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fac_utils.R
\name{om_quantile_table}
\alias{om_quantile_table}
\title{Difference in monthly percentiles}
\usage{
om_quantile_table(
  rundata_df,
  metrics = c("Qintake", "Runit"),
  quantiles = c(0, 0.1, 0.25, 0.5, 0.75, 0.9, 1),
  rdigits = 2,
  na.rm = TRUE
)
}
\arguments{
\item{rundata_df}{Model run timeseries dataframe.}

\item{metrics}{Which metrics to calculate quantiles for? These should be
column names of rundata_df}

\item{quantiles}{Which quantiles to calculate?}

\item{rdigits}{Number of Digits to round column values.}

\item{na.rm}{Logical, default TRUE. Should NA values be removed prior to
calculating quantiles of each metric?}
}
\value{
A data frame that if has one row per variable with quantiles in columns
}
\description{
Using a data frame and column list provided by user, calculate quantiles
for desired variables.
}
