% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_vahydro_get_token.R
\name{om_vahydro_get_token}
\alias{om_vahydro_get_token}
\title{Retrieve a token from vaydro REST service.  TBD: replace with object to store token and dn connection}
\usage{
om_vahydro_get_token(base_url = "http://deq2.bse.vt.edu/d.dh")
}
\arguments{
\item{base_url}{character deprecated to be replaced by global datasource}
}
\value{
token from VAHydro REST web service
}
\description{
Retrieve a token from vaydro REST service.  TBD: replace with object to store token and dn connection
}
\examples{
NA
}
\seealso{
NA
}
