% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_vahydro_token.R
\name{om_vahydro_token}
\alias{om_vahydro_token}
\title{Retrieve a token from vahydro REST service.  TBD: replace with object to store token and dn connection}
\usage{
om_vahydro_token(base_url = "http://deq2.bse.vt.edu/d.dh", rest_uname = NULL)
}
\arguments{
\item{base_url}{character deprecated to be replaced by global datasource}

\item{rest_uname}{username to authenticate (if omitted will prompt)}
}
\value{
token character from VAHydro REST web service
}
\description{
Retrieve a token from vahydro REST service.  TBD: replace with object to store token and dn connection
}
\examples{
NA
}
\seealso{
NA
}
