% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openmi.om.base.R
\name{openmi.om.base}
\alias{openmi.om.base}
\title{The base object class for meta-model components, utilizing a openmi to
leverage the VA Hydro Operation Model components}
\value{
R6 class of type openmi.om.base
}
\description{
Class providing the minimum attributes and methods for a model
  component of the VA Hydro OM
}
\details{
Has standard methods for iterating through timesteps and connecting
  with other model components
}
\examples{
NA
}
\seealso{
https://doi.org/10.1016/j.envsoft.2018.07.015
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{A unique identifier for this controller}

\item{\code{code}}{A unique identifier for this controller}

\item{\code{debug}}{mode on/off}

\item{\code{value}}{The value at the current instance}

\item{\code{data}}{Data from the larger context, including the parent. was arData in original om}

\item{\code{state}}{This objects local state. was state in om}

\item{\code{inputs}}{Inputs linked to this object}

\item{\code{components}}{Components contained by this object often equations,
matrices, etc.}

\item{\code{host}}{Host database name}

\item{\code{type}}{Type of component}

\item{\code{compid}}{compid is a unique identifier in this simulation domain}

\item{\code{vars}}{vars is an array of variables that this requires for solving, determines op order}

\item{\code{timer}}{Timer is the object keepign time in simulation (set by parent controller)}

\item{\code{id}}{id is identifier (compid, name duplicates?)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-openmi.om.base-new}{\code{openmi.om.base$new()}}
\item \href{#method-openmi.om.base-settable}{\code{openmi.om.base$settable()}}
\item \href{#method-openmi.om.base-set_prop}{\code{openmi.om.base$set_prop()}}
\item \href{#method-openmi.om.base-parse_openmi}{\code{openmi.om.base$parse_openmi()}}
\item \href{#method-openmi.om.base-parse_class_specific}{\code{openmi.om.base$parse_class_specific()}}
\item \href{#method-openmi.om.base-set_sub_prop}{\code{openmi.om.base$set_sub_prop()}}
\item \href{#method-openmi.om.base-init}{\code{openmi.om.base$init()}}
\item \href{#method-openmi.om.base-set_vars}{\code{openmi.om.base$set_vars()}}
\item \href{#method-openmi.om.base-log_debug}{\code{openmi.om.base$log_debug()}}
\item \href{#method-openmi.om.base-getInputs}{\code{openmi.om.base$getInputs()}}
\item \href{#method-openmi.om.base-prepare}{\code{openmi.om.base$prepare()}}
\item \href{#method-openmi.om.base-step}{\code{openmi.om.base$step()}}
\item \href{#method-openmi.om.base-stepChildren}{\code{openmi.om.base$stepChildren()}}
\item \href{#method-openmi.om.base-update}{\code{openmi.om.base$update()}}
\item \href{#method-openmi.om.base-finish}{\code{openmi.om.base$finish()}}
\item \href{#method-openmi.om.base-validate}{\code{openmi.om.base$validate()}}
\item \href{#method-openmi.om.base-logState}{\code{openmi.om.base$logState()}}
\item \href{#method-openmi.om.base-getValue}{\code{openmi.om.base$getValue()}}
\item \href{#method-openmi.om.base-addInput}{\code{openmi.om.base$addInput()}}
\item \href{#method-openmi.om.base-addComponent}{\code{openmi.om.base$addComponent()}}
\item \href{#method-openmi.om.base-add_component}{\code{openmi.om.base$add_component()}}
\item \href{#method-openmi.om.base-get_component_id}{\code{openmi.om.base$get_component_id()}}
\item \href{#method-openmi.om.base-orderOperations}{\code{openmi.om.base$orderOperations()}}
\item \href{#method-openmi.om.base-asJSON}{\code{openmi.om.base$asJSON()}}
\item \href{#method-openmi.om.base-clone}{\code{openmi.om.base$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.base-new"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.base-new}{}}}
\subsection{Method \code{new()}}{
Create an openmi.om.base property with name or host
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.base$new(elem_list = list(), format = "raw")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{elem_list}}{Which properties to set on creation?}

\item{\code{format}}{Format of elem_list}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
R6 class object with populated name or host fields
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.base-settable"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.base-settable}{}}}
\subsection{Method \code{settable()}}{
What fields may be set on this object?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.base$settable()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
array of settable properties
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.base-set_prop"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.base-set_prop}{}}}
\subsection{Method \code{set_prop()}}{
Sets a property value on this object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.base$set_prop(propname, propvalue, format = "raw")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{propname}}{which attribute}

\item{\code{propvalue}}{what value}

\item{\code{format}}{The intended format of propvalue}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing, but sets host or name via \code{openmi.om.base$set_sub_prop()}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.base-parse_openmi"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.base-parse_openmi}{}}}
\subsection{Method \code{parse_openmi()}}{
Derives the value from an openmi formatted property list to find
relevant modelling info based on model object type
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.base$parse_openmi(propvalue)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{propvalue}}{openmi formatted property list}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a settable data value from openmi json type data description
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.base-parse_class_specific"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.base-parse_class_specific}{}}}
\subsection{Method \code{parse_class_specific()}}{
Method to allow special parsing of model property values
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.base$parse_class_specific(propvalue)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{propvalue}}{from some custom classformat reading implementation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a settable data value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.base-set_sub_prop"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.base-set_sub_prop}{}}}
\subsection{Method \code{set_sub_prop()}}{
set the value of a contained component (not a local class attribute)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.base$set_sub_prop(propname, propvalue, format = "raw")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{propname}}{which attribute}

\item{\code{propvalue}}{what value}

\item{\code{format}}{of propvalue}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a settable data value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.base-init"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.base-init}{}}}
\subsection{Method \code{init()}}{
initialize this component after first object creation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.base$init()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Nothing, but will run \code{self$set_vars()} to populate fields
  and will initialize each component (using the \code{init()} method on
  each component)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.base-set_vars"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.base-set_vars}{}}}
\subsection{Method \code{set_vars()}}{
set_vars finds all the input var names for this function
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.base$set_vars()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.base-log_debug"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.base-log_debug}{}}}
\subsection{Method \code{log_debug()}}{
log_debug handles debug info
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.base$log_debug(debug_mesg)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{debug_mesg}}{message to add to logger}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.base-getInputs"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.base-getInputs}{}}}
\subsection{Method \code{getInputs()}}{
get all input values from linked components and internal
  timeseried feeds
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.base$getInputs()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Nothing, but sets data fields
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.base-prepare"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.base-prepare}{}}}
\subsection{Method \code{prepare()}}{
execute things to do before model timestep execution
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.base$prepare()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Nothing, but runs \code{self$getInputs()}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.base-step"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.base-step}{}}}
\subsection{Method \code{step()}}{
perform model timestep execution via \code{self$prepare()},
\code{self$stepChildren()}, \code{self$update()}, \code{self$finish()},
 \code{self$logState()}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.base$step()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Nothing, but runs various state methods
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.base-stepChildren"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.base-stepChildren}{}}}
\subsection{Method \code{stepChildren()}}{
execute child model timestep code by calling \code{step()}
  for each component opn this object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.base$stepChildren()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.base-update"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.base-update}{}}}
\subsection{Method \code{update()}}{
execute model timestep code
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.base$update()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.base-finish"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.base-finish}{}}}
\subsection{Method \code{finish()}}{
do things at end of model step and run for each component
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.base$finish()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.base-validate"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.base-validate}{}}}
\subsection{Method \code{validate()}}{
Is this object valid?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.base$validate()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
logical TRUE/FALSE
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.base-logState"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.base-logState}{}}}
\subsection{Method \code{logState()}}{
log data at end of timestep
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.base$logState()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.base-getValue"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.base-getValue}{}}}
\subsection{Method \code{getValue()}}{
get value of this object currently
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.base$getValue(name = "value")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{a specific name from the state array, not just the default. TBD}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.base-addInput"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.base-addInput}{}}}
\subsection{Method \code{addInput()}}{
connect an input to this component
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.base$addInput(
  local_name = character(),
  object = openmi.om.base,
  remote_name = "",
  input_type = "numeric"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{local_name}}{name that will be referred to in local contest}

\item{\code{object}}{is the actual R6 class to connect to}

\item{\code{remote_name}}{is what property on the remote object are we accessing}

\item{\code{input_type}}{is this is a number (most common), text or other?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Sets inputs field
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.base-addComponent"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.base-addComponent}{}}}
\subsection{Method \code{addComponent()}}{
add a contained sub-component (i.e. not linked)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.base$addComponent(thiscomp = openmi.om.base)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thiscomp}}{an R6 classof open.mi type}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.base-add_component"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.base-add_component}{}}}
\subsection{Method \code{add_component()}}{
add a contained sub-component (i.e. not linked)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.base$add_component(thiscomp = openmi.om.base)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thiscomp}}{an R6 classof open.mi type}
}
\if{html}{\out{</div>}}
}
\subsection{Details}{
components are small objects that reside inside this object
these *were* called processors in the old version of the model other
models are handled exclusively by the runtime controller/container and
they *may be* known as model_entities -- or they may *also* be known as
components: is there any reason to have a separate concept of components
and sub_components? run hierarchy?
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.base-get_component_id"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.base-get_component_id}{}}}
\subsection{Method \code{get_component_id()}}{
return unique ID of component
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.base$get_component_id(thiscomp)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thiscomp}}{an R6 classof open.mi type}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
integer
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.base-orderOperations"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.base-orderOperations}{}}}
\subsection{Method \code{orderOperations()}}{
order contained sub-components
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.base$orderOperations()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.base-asJSON"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.base-asJSON}{}}}
\subsection{Method \code{asJSON()}}{
format this object as openmi json
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.base$asJSON()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
json text
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.base-clone"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.base-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.base$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
