% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openmi.om.linkableComponent.R
\name{openmi.om.linkableComponent}
\alias{openmi.om.linkableComponent}
\title{The base class for linkable meta-model components.}
\value{
R6 class of type openmi.om.linkableComponent
}
\description{
The base class for linkable meta-model components.
}
\section{Super class}{
\code{\link[hydrotools:openmi.om.base]{openmi.om.base}} -> \code{openmi.om.linkableComponent}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-openmi.om.linkableComponent-clone}{\code{openmi.om.linkableComponent$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="addComponent"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-addComponent'><code>openmi.om.base$addComponent()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="addInput"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-addInput'><code>openmi.om.base$addInput()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="add_component"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-add_component'><code>openmi.om.base$add_component()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="asJSON"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-asJSON'><code>openmi.om.base$asJSON()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="finish"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-finish'><code>openmi.om.base$finish()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="getInputs"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-getInputs'><code>openmi.om.base$getInputs()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="getValue"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-getValue'><code>openmi.om.base$getValue()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="get_component_id"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-get_component_id'><code>openmi.om.base$get_component_id()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="init"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-init'><code>openmi.om.base$init()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="initialize"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-initialize'><code>openmi.om.base$initialize()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="logState"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-logState'><code>openmi.om.base$logState()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="log_debug"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-log_debug'><code>openmi.om.base$log_debug()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="orderOperations"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-orderOperations'><code>openmi.om.base$orderOperations()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="parse_class_specific"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-parse_class_specific'><code>openmi.om.base$parse_class_specific()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="parse_openmi"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-parse_openmi'><code>openmi.om.base$parse_openmi()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="prepare"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-prepare'><code>openmi.om.base$prepare()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="set_prop"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-set_prop'><code>openmi.om.base$set_prop()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="set_sub_prop"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-set_sub_prop'><code>openmi.om.base$set_sub_prop()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="set_vars"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-set_vars'><code>openmi.om.base$set_vars()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="settable"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-settable'><code>openmi.om.base$settable()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="step"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-step'><code>openmi.om.base$step()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="stepChildren"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-stepChildren'><code>openmi.om.base$stepChildren()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="update"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-update'><code>openmi.om.base$update()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="validate"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-validate'><code>openmi.om.base$validate()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.linkableComponent-clone"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.linkableComponent-clone}{}}}
\subsection{\code{openmi.om.linkableComponent$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{openmi.om.linkableComponent$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
