% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fac_utils.R
\name{plot_boxplot_context}
\alias{plot_boxplot_context}
\title{Monthly boxplot with single year highlight}
\usage{
plot_boxplot_context(
  targetYear = as.numeric(format(Sys.Date(), "\%Y")),
  include_months = 1:12,
  AYS = 10,
  dataframe_in = NULL,
  value_col = "flow",
  date_col = "time",
  gage_id = NA,
  metric = mean,
  use_y_log = TRUE,
  ylab = "Monthly Mean Flow (cfs)"
)
}
\arguments{
\item{targetYear}{What year should be highlighted on the plot? Defaults
to this year. If left as NULL, no additional point data will be included}

\item{include_months}{What months should be included in plot/analysis?}

\item{AYS}{Which month begins the analysis year? The plot will start in
this month and then proceed through the next 12 months or those
included by user in include_months}

\item{dataframe_in}{A data frame to summarize and use for the plot. Must
contain the fields input in value_col and date_col}

\item{value_col}{A field name in dataframe_in used to control which column is
plotted from dataframe_in}

\item{date_col}{A field name in dataframe_in that must contain R Dates, to be
used in developing the monthly metrics}

\item{gage_id}{A name to display in the plot title}

\item{metric}{A function to pass to \code{dplyr::summarize()} to control
how data monthly data is aggregated}

\item{use_y_log}{Boolean. Should the y axis be on a log10 scale? Defaults
to TRUE}

\item{ylab}{Character. The text to display on the y-axis of the plot.}
}
\value{
A boxplot of monthly mean flows with highlighted points
}
\description{
Create a boxplot of the monthly metric requested by user at
  this gage (or of a user input data frame) and then plot the target year
  as highlighted points on the plot. Often used in OWS to assess
  potential winter baseflow recharge when using gage data and evaluting
  between October - March (include_months = c(10,11,12,1,2,3)).
}
