% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cia_utils.R
\name{renameNHD}
\alias{renameNHD}
\title{Use VA Hydro Varkeys on NHDPlus}
\usage{
renameNHD(get_nhdplus_df, returnPlotName = FALSE)
}
\arguments{
\item{get_nhdplus_df}{A data frame output from
\code{nhdPlusTools::get_nhdplus()} or the \code{simple_nhdplus()} wrapper.
Alternatively, a vector of nhd flow names}

\item{returnPlotName}{Logical. Defaults to FALSE. If TRUE, returns a character
vector of names that describe the flow column}
}
\value{
An identical data frame with flow columns renamed using appropriate
 varkeys if returnPlotName is FALSE or a character vector of readible names
}
\description{
A function to rename the flows from NHDPlus to VAHydro varkeys
}
\details{
This function takes a data.frame or tibble output from 
\code{nhdPlusTools::get_nhdplus()} and renames the relevant flow columns using
VA Hydro varkeys from dh_variabledefinition to allow for consistent comparison
to data from VA Hydro databases. Alternatively, it takes in a vector of flow
names from \code{nhdPlusTools::get_nhdplus()} and returns a vector of readible
names (when returnPlotName is TRUE).
}
\examples{
\dontrun{
out <- simple_nhdPlusFlows(ds = ds, hydroid = 68069, huc_level = "huc8")
out <- renameNHD(out)
names(out)
}
}
