% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RomDataSource.R
\name{RomDataSource}
\alias{RomDataSource}
\title{Global object store holds references to all objects of a given domain
and pointers to remote data source if applicable.
if remote data source is set, all data can be synched back to it
no mixing of data sources is allowed}
\value{
reference class of type openmi.om.base.
}
\description{
Data Source containing tables and methods of Model data objects
}
\details{
Provides integrated, queryable universe of features, observations and meta data
}
\examples{
NA
}
\seealso{
NA
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{site}}{URL of some RESTful repository}

\item{\code{json_obj_url}}{URL for retrieving full objects}

\item{\code{rest_uname}}{username to connect to RESTful repository}

\item{\code{timeline}}{for default time series data}

\item{\code{tsvalues}}{table of time series data}

\item{\code{propvalues}}{table of object properties (can be contained by objects)}

\item{\code{features}}{table of physical features}

\item{\code{var_defs}}{table of variable definitions}

\item{\code{admin_features}}{table of adminreg features}

\item{\code{prop_json_cache}}{list of json objects retrieved, keyed by ID}

\item{\code{ts_cache}}{list of ts objects instantiated}

\item{\code{feature_cache}}{list of feature objects instantiated}

\item{\code{prop_cache}}{list of prop objects instantiated}

\item{\code{var_cache}}{list of var objects instantiated}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RomDataSource-new}{\code{RomDataSource$new()}}
\item \href{#method-RomDataSource-reset}{\code{RomDataSource$reset()}}
\item \href{#method-RomDataSource-get_token}{\code{RomDataSource$get_token()}}
\item \href{#method-RomDataSource-get_vardef}{\code{RomDataSource$get_vardef()}}
\item \href{#method-RomDataSource-get_prop}{\code{RomDataSource$get_prop()}}
\item \href{#method-RomDataSource-get_ts}{\code{RomDataSource$get_ts()}}
\item \href{#method-RomDataSource-auth_read}{\code{RomDataSource$auth_read()}}
\item \href{#method-RomDataSource-set_ts}{\code{RomDataSource$set_ts()}}
\item \href{#method-RomDataSource-set_prop}{\code{RomDataSource$set_prop()}}
\item \href{#method-RomDataSource-set_vardef}{\code{RomDataSource$set_vardef()}}
\item \href{#method-RomDataSource-store_features}{\code{RomDataSource$store_features()}}
\item \href{#method-RomDataSource-set_feature}{\code{RomDataSource$set_feature()}}
\item \href{#method-RomDataSource-get}{\code{RomDataSource$get()}}
\item \href{#method-RomDataSource-post}{\code{RomDataSource$post()}}
\item \href{#method-RomDataSource-get_json_prop}{\code{RomDataSource$get_json_prop()}}
\item \href{#method-RomDataSource-clone}{\code{RomDataSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomDataSource-new"></a>}}
\if{latex}{\out{\hypertarget{method-RomDataSource-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomDataSource$new(site, rest_uname = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{site}}{URL of some RESTful repository}

\item{\code{rest_uname}}{username to connect to RESTful repository}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
object instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomDataSource-reset"></a>}}
\if{latex}{\out{\hypertarget{method-RomDataSource-reset}{}}}
\subsection{Method \code{reset()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomDataSource$reset(table)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table}}{which table. Default 'all'}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing clears data tables
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomDataSource-get_token"></a>}}
\if{latex}{\out{\hypertarget{method-RomDataSource-get_token}{}}}
\subsection{Method \code{get_token()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomDataSource$get_token(rest_pw = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{rest_pw}}{to use, if NULL will prompt}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing sets internal private token
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomDataSource-get_vardef"></a>}}
\if{latex}{\out{\hypertarget{method-RomDataSource-get_vardef}{}}}
\subsection{Method \code{get_vardef()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomDataSource$get_vardef(varkey, force_update = FALSE, debug = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{varkey}}{= variable key}

\item{\code{force_update}}{Check remote repository for new info?}

\item{\code{debug}}{show info}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing sets internal private token
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomDataSource-get_prop"></a>}}
\if{latex}{\out{\hypertarget{method-RomDataSource-get_prop}{}}}
\subsection{Method \code{get_prop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomDataSource$get_prop(
  config,
  return_type = "data.frame",
  force_refresh = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{= list(entity_type, featureid, tid = NULL, varid = NULL, tstime = NULL, tsendtime = NULL, tscode = NULL, tlid = NULL) timeline ID (not yet used)}

\item{\code{return_type}}{'data.frame' or 'list'}

\item{\code{force_refresh}}{if this ds has a remote source, whether to pull anew}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing sets internal private token
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomDataSource-get_ts"></a>}}
\if{latex}{\out{\hypertarget{method-RomDataSource-get_ts}{}}}
\subsection{Method \code{get_ts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomDataSource$get_ts(config, return_type = "data.frame", force_refresh = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{= list(entity_type, featureid, tid = NULL, varid = NULL, tstime = NULL, tsendtime = NULL, tscode = NULL, tlid = NULL) timeline ID (not yet used)}

\item{\code{return_type}}{'data.frame' or 'list'}

\item{\code{force_refresh}}{if this ds has a remote source, whether to pull anew}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing sets internal private token
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomDataSource-auth_read"></a>}}
\if{latex}{\out{\hypertarget{method-RomDataSource-auth_read}{}}}
\subsection{Method \code{auth_read()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomDataSource$auth_read(
  uri,
  content_type = "text/csv",
  delim = ",",
  enc = "xml"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{remote address to retrieve data}

\item{\code{content_type}}{http content-type}

\item{\code{delim}}{delimiter}

\item{\code{enc}}{encoding}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
result of web request
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomDataSource-set_ts"></a>}}
\if{latex}{\out{\hypertarget{method-RomDataSource-set_ts}{}}}
\subsection{Method \code{set_ts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomDataSource$set_ts(ts)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ts}}{= list(entity_type, featureid, tid = NULL, varid = NULL, tstime = NULL, tsendtime = NULL, tscode = NULL, tlid = NULL) timeline ID (not yet used)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing sets internal private token
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomDataSource-set_prop"></a>}}
\if{latex}{\out{\hypertarget{method-RomDataSource-set_prop}{}}}
\subsection{Method \code{set_prop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomDataSource$set_prop(prop)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{prop}}{= list(entity_type, featureid, pid = NULL, varid = NULL, tstime = NULL, tsendtime = NULL, tscode = NULL, tlid = NULL) timeline ID (not yet used)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing seprop internal private token
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomDataSource-set_vardef"></a>}}
\if{latex}{\out{\hypertarget{method-RomDataSource-set_vardef}{}}}
\subsection{Method \code{set_vardef()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomDataSource$set_vardef(var_def)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var_def}}{= list(varid, varkey, varname, varunits, varcode,...)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
local df index?
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomDataSource-store_features"></a>}}
\if{latex}{\out{\hypertarget{method-RomDataSource-store_features}{}}}
\subsection{Method \code{store_features()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomDataSource$store_features(features)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{features}}{= list(entity_type, featureid, pid = NULL, varid = NULL, tstime = NULL, tsendtime = NULL, tscode = NULL, tlid = NULL) timeline ID (not yet used)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing seprop internal private token
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomDataSource-set_feature"></a>}}
\if{latex}{\out{\hypertarget{method-RomDataSource-set_feature}{}}}
\subsection{Method \code{set_feature()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomDataSource$set_feature(feature)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{feature}}{= list(entity_type, featureid, pid = NULL, varid = NULL, tstime = NULL, tsendtime = NULL, tscode = NULL, tlid = NULL) timeline ID (not yet used)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing seprop internal private token
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomDataSource-get"></a>}}
\if{latex}{\out{\hypertarget{method-RomDataSource-get}{}}}
\subsection{Method \code{get()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomDataSource$get(entity_type, pk, config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entity_type}}{= dh_feature, dh_properties, ...}

\item{\code{pk}}{= primary key column name, e.g. hydroid, pid, ...}

\item{\code{config}}{= contents of record to post in list(pid, propname, propvalue, ...)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
local df index?
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomDataSource-post"></a>}}
\if{latex}{\out{\hypertarget{method-RomDataSource-post}{}}}
\subsection{Method \code{post()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomDataSource$post(entity_type, pk, config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entity_type}}{= dh_feature, dh_properties, ...}

\item{\code{pk}}{= primary key column name, e.g. hydroid, pid, ...}

\item{\code{config}}{= contents of record to post in list(pid, propname, propvalue, ...)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
local df index?
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomDataSource-get_json_prop"></a>}}
\if{latex}{\out{\hypertarget{method-RomDataSource-get_json_prop}{}}}
\subsection{Method \code{get_json_prop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomDataSource$get_json_prop(pid)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pid}}{= object pid}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
unserialized json as list, with object stored in ds$prop_json_cache
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomDataSource-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RomDataSource-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomDataSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
