% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RomEntity.R
\name{RomEntity}
\alias{RomEntity}
\title{Base entity data object}
\value{
reference class of type openmi.om.base.
}
\description{
Object for storing a single entity with attribute and timeseries related
}
\details{
Has standard methods for managing data and meta data
}
\examples{
NA
}
\seealso{
NA
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{base_entity_type}}{kind of entity}

\item{\code{pk_name}}{the name of this entity's pk column}

\item{\code{datasource}}{RomDataSource}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RomEntity-get_id}{\code{RomEntity$get_id()}}
\item \href{#method-RomEntity-propvalues}{\code{RomEntity$propvalues()}}
\item \href{#method-RomEntity-tsvalues}{\code{RomEntity$tsvalues()}}
\item \href{#method-RomEntity-new}{\code{RomEntity$new()}}
\item \href{#method-RomEntity-clone}{\code{RomEntity$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomEntity-get_id"></a>}}
\if{latex}{\out{\hypertarget{method-RomEntity-get_id}{}}}
\subsection{Method \code{get_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomEntity$get_id()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
get_id the unique id of this entity alias to remote pkid, subclassed as function
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomEntity-propvalues"></a>}}
\if{latex}{\out{\hypertarget{method-RomEntity-propvalues}{}}}
\subsection{Method \code{propvalues()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomEntity$propvalues(propname = NULL, varid = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{propname}}{optional name to filter}

\item{\code{varid}}{option variable to filter}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
propvalues unique properties of this entity
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomEntity-tsvalues"></a>}}
\if{latex}{\out{\hypertarget{method-RomEntity-tsvalues}{}}}
\subsection{Method \code{tsvalues()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomEntity$tsvalues(varkey = NULL, tstime = NULL, tsendtime = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{varkey}}{option variable to filter}

\item{\code{tstime}}{timespan begin}

\item{\code{tsendtime}}{timespan end}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
tsvalues unique timeseries records for this entity
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomEntity-new"></a>}}
\if{latex}{\out{\hypertarget{method-RomEntity-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomEntity$new(datasource = NULL, config = list(), load_remote = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datasource}}{RESTful repository (optional)}

\item{\code{config}}{list of attributes to set, see also: to_list() for format}

\item{\code{load_remote}}{automatically query REST data source for matches?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
object instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomEntity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RomEntity-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomEntity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
