% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RomProperty.R
\name{RomProperty}
\alias{RomProperty}
\title{Time Series data object}
\value{
reference class of type openmi.om.base.
}
\description{
Object for storing single time series value
}
\details{
Has standard methods for managing data and meta data
}
\examples{
NA
}
\seealso{
NA
}
\section{Super class}{
\code{\link[hydrotools:RomEntity]{hydrotools::RomEntity}} -> \code{RomProperty}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{base_entity_type}}{kind of entity}

\item{\code{pk_name}}{the name of this entity's pk column}

\item{\code{pid}}{unique ID in this RomDataSource}

\item{\code{featureid}}{id of entity this is attached to}

\item{\code{entity_type}}{type of entity this is attached to}

\item{\code{propname}}{locally unique name}

\item{\code{startdate}}{begin timestamp}

\item{\code{enddate}}{end time stamp}

\item{\code{propvalue}}{numerical value}

\item{\code{propcode}}{alphanumeric value}

\item{\code{varid}}{variable ID from RomDataSource}

\item{\code{bundle}}{(for future use)}

\item{\code{uid}}{user id of creator}

\item{\code{modified}}{timestamp}

\item{\code{datasource}}{RomDataSource}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-get_id}{\code{RomProperty$get_id()}}
\item \href{#method-new}{\code{RomProperty$new()}}
\item \href{#method-from_list}{\code{RomProperty$from_list()}}
\item \href{#method-to_list}{\code{RomProperty$to_list()}}
\item \href{#method-set_prop}{\code{RomProperty$set_prop()}}
\item \href{#method-save}{\code{RomProperty$save()}}
\item \href{#method-clone}{\code{RomProperty$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="hydrotools" data-topic="RomEntity" data-id="propvalues">}\href{../../hydrotools/html/RomEntity.html#method-propvalues}{\code{hydrotools::RomEntity$propvalues()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="hydrotools" data-topic="RomEntity" data-id="tsvalues">}\href{../../hydrotools/html/RomEntity.html#method-tsvalues}{\code{hydrotools::RomEntity$tsvalues()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_id"></a>}}
\if{latex}{\out{\hypertarget{method-get_id}{}}}
\subsection{Method \code{get_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomProperty$get_id()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
get_id the id of this entity alias to remote pkid, subclassed as function
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomProperty$new(datasource = NULL, config = list(), load_remote = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datasource}}{optional RomDataSource for remote and local storage}

\item{\code{config}}{list of attributes to set, see also: to_list() for format}

\item{\code{load_remote}}{automatically query REST data source for matches?}

\item{\code{site}}{URL of some RESTful repository}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
object instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomProperty$from_list(config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{list of attributes to set, see also: to_list() for format}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomProperty$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of object attributes suitable for input to new() and from_list() methods
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_prop"></a>}}
\if{latex}{\out{\hypertarget{method-set_prop}{}}}
\subsection{Method \code{set_prop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomProperty$set_prop(name, value, format = "default")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{attribute name}

\item{\code{value}}{attribute value}

\item{\code{format}}{type of data being sent (default, json, ...)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
object instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save"></a>}}
\if{latex}{\out{\hypertarget{method-save}{}}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomProperty$save(push_remote = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{push_remote}}{whether to automatically propagate changes to remote data source}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
boolean TRUE on success, FALSE on failure
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomProperty$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
