% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RomTS.R
\name{RomTS}
\alias{RomTS}
\title{Time Series data object}
\value{
reference class of type openmi.om.base.
}
\description{
Object for storing single time series value
}
\details{
Has standard methods for managing data and meta data
}
\examples{
NA
}
\seealso{
NA
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{tid}}{unique ID in this RomDataSource}

\item{\code{featureid}}{id of entity this is attached to}

\item{\code{entity_type}}{type of entity this is attached to}

\item{\code{tstime}}{begin timestamp}

\item{\code{tsendtime}}{end time stamp}

\item{\code{tsvalue}}{numerical value}

\item{\code{tscode}}{alphanumerical value}

\item{\code{tlid}}{timeline ID (for future use)}

\item{\code{varid}}{variable ID from RomDataSource}

\item{\code{bundle}}{(for future use)}

\item{\code{uid}}{user id of creator}

\item{\code{datasource}}{RomDataSource}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RomTS$new()}}
\item \href{#method-from_list}{\code{RomTS$from_list()}}
\item \href{#method-to_list}{\code{RomTS$to_list()}}
\item \href{#method-set_prop}{\code{RomTS$set_prop()}}
\item \href{#method-clone}{\code{RomTS$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomTS$new(datasource = NULL, config = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datasource}}{optional RomDataSource for remote and local storage}

\item{\code{config}}{list of attributes to set, see also: to_list() for format}

\item{\code{site}}{URL of some RESTful repository}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
object instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomTS$from_list(config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{list of attributes to set, see also: to_list() for format}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomTS$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of object attributes suitable for input to new() and from_list() methods
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_prop"></a>}}
\if{latex}{\out{\hypertarget{method-set_prop}{}}}
\subsection{Method \code{set_prop()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomTS$set_prop(name, value, format = "default")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name}}{attribute name}

\item{\code{value}}{attribute value}

\item{\code{format}}{type of data being sent (default, json, ...)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
object instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomTS$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
