% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wshd_utils.R
\name{set_zoom}
\alias{set_zoom}
\title{set_zoom}
\usage{
set_zoom(pb)
}
\arguments{
\item{pb}{a bounding box vector often from \code{sf::st_bbox()}}
}
\description{
Determine general zoom level for maptiles
}
\details{
This function was taken from \code{nhdplusTools:::set_zoom} verbatim
 and is used to set default zoom levels to grab maptiles background maps
}
