% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cia_utils.R
\name{simple_elfgen}
\alias{simple_elfgen}
\title{Simple Hydro Elfgen Wrapper}
\usage{
simple_elfgen(
  ds,
  hydroid,
  huc_level = "huc8",
  dataset = "VAHydro-EDAS",
  ws_varkey = "erom_q0001e_aug",
  quantile = 0.8,
  breakpt = "ymax",
  yaxis_thresh = NULL,
  blo = 0,
  bhi = yaxis_thresh
)
}
\arguments{
\item{ds}{A datasource provided by the user, usally the RomDataSource instance
create in DEQ config.R files, querying drupal.dh03}

\item{hydroid}{The hydroid of a feature of interest. This is used to find
intersecting NHD Plus catchments and to ultimately identify the NHD feature
of interest based on the huc_level input}

\item{huc_level}{Which NHD waterhsed level should be used to generate the
\code{elfgen()} relationships? Defaults to "huc8", representing HUC8}

\item{dataset}{Should \code{elfgen()} use USGS Icthys data or DEQ EDAS data
(via Hydro, which may be outdated but is state-specific data). Options are
"IchthyMaps" for USGS or "VAHydro-EDAS" for DEQ data}

\item{ws_varkey}{Which NHDPlus flow for the \code{huc_level} should be used to
generate the \code{elfgen()} ecologic limit function? The options are mean
flow for any month but MUST be in the form "erom_q0001e_aug". Use
"erom_q0001e_mean" to use mean annual flow.}

\item{quantile}{A specified value for the quantile of interest - 0.95 equals
the 95th percentile. This is used in the eflgen ecologic limit function
analysis}

\item{breakpt}{A breakpoint - either user-supplied fixed value or may be
derived using elfgen breakpoint functions \code{bkpt_pwit()} or
\code{bkpt_ymax()} by inputting the character value "ymax" or "pwit"
respectively. If using "pwit", it is recommended the user provide a
\code{blo} and \code{bhi} value as documented in the elfgen pacakge.
\code{blo} will default to 0 and \code{bhi} to the \code{yaxis_thresh} value
input, which itself defaults to the maximum number of taxa}

\item{yaxis_thresh}{The maximum value to be used in the plot. If this value is
left NULL by user, it will be set as the maximum value of the dataset. If
using "pwit" as the breakpt, this will also serve as the \code{bhi} value if
not otherwise specified by user.}

\item{blo}{When using the "pwit" breakpt analysis, this is the "bound low"
value, or the lower bound of the piecewise range and defaults to 0. See
\code{elfgen::bkpt_pwit()} for additional details.}

\item{bhi}{When using the "pwit" breakpt analysis, this is the "bound high"
value, or the upper bound of the piecewise range and defaults to
\code{yaxis_thresh}. See \code{elfgen::bkpt_pwit()} for additional details.}
}
\value{
Object containing plot image and dataframe of ELF statistics. See
 \code{elfgen::elfgen()} for more details
}
\description{
A wrapper that makes it easier to access VA Hydro's imported EDAS
 dataset and provides simple arguments to eflgen package.
}
\details{
This function provides the necessary inputs to call the
 elfgen::elfgen() function. Simply, this wrapper function uses easily
 accessible input to find the necessary NHD Plus watersheds and fish taxa
 data to create the \code{watershed.df} input in elfgen() and returns its
 output. The function will use the largest NHDPlus feature of the input code
 that is contained within the user input hydro feature based on the provided
 hydroid
}
