% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cia_utils.R
\name{simple_nhdPlusFlows}
\alias{simple_nhdPlusFlows}
\title{Get NHDPlus HUC Monthly Flows}
\usage{
simple_nhdPlusFlows(ds, hydroid, huc_level = "huc8")
}
\arguments{
\item{ds}{A datasource provided by the user, usally the RomDataSource instance
create in DEQ config.R files, querying drupal.dh03}

\item{hydroid}{The hydroid of a feature of interest. This is used to find
intersecting NHD Plus catchments and to ultimately identify the NHD feature
of interest based on the huc_level input}

\item{huc_level}{Which NHD waterhsed level should be used to generate the
\code{elfgen()} relationships? Defaults to "huc8", representing HUC8}
}
\value{
A list that contains the HUC code of interest and data on the nhdPlus
 segment
}
\description{
A wrapper function that makes it easy to obtain HUC flows from a
 hydro feature
}
\details{
This function provides the necessary inputs to call nhdPlusTools
 functions to extract the relevant HUC flows from an area of interest. The
 function will use the largest NHDPlus feature of the input code that is
 contained within the user input hydro feature based on the provided hydroid.
 Output flow columns will be renamed using VA Hydro varkeys.
}
\examples{
\dontrun{simple_nhdPlusFlows(ds = ds, hydroid = 68069, huc_level = "huc8")}
}
