% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wshd_utils.R
\name{usgs_bankfull_properties}
\alias{usgs_bankfull_properties}
\title{usgs_bankfull_properties}
\usage{
usgs_bankfull_properties(prov, da)
}
\arguments{
\item{prov}{Integer 1 - 4. Geomorphic province 1 = appalachian plateau, 2 =
valley and ridge, 3 = Piedmont, 4 = coastal plain}

\item{da}{Numeric. The drainage area of your channel.}
}
\value{
A list with the bank full stage (h), bank full width (bf), base width
 (b), side slope (z), and mannings roughness (n)
}
\description{
Uses USGS bankfull regional curves for each VA geomorphic
 province to calculate bankfull channel stage, width, full width, and side
 slope
}
\details{
See USGS bankfull regression reports for additional information. The
 empirical bankfull equations are used to calculate channel properties based
 on which province the channel is in and the drainage area at that point.
 This data is often used in OM to set local_channel properties. Returns a list 
 with bank full stage "max height before floodplain", bank full width "max width before floodplain",
 base width "width at lowest stage", side slope of channel, and mannings Roughness
}
