% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wshd_utils.R
\name{usgs_calib_rarray}
\alias{usgs_calib_rarray}
\title{usgs_calib_rarray}
\usage{
usgs_calib_rarray(riverseg_json, gage_info, model_runid)
}
\arguments{
\item{riverseg_json}{a valid RomFeature for the watershed of interest}

\item{gage_info}{a dataframe returned from dataRetrieval::readNWISsite(gageid)}

\item{model_runid}{a model run id/scenario}
}
\description{
Create a best guess calibration render script param set.
}
\details{
Prepares a params list for the gage_vs_model.Rmd in hydro-tools/USGS.
}
