% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wshd_utils.R
\name{usgs_nearest_gage}
\alias{usgs_nearest_gage}
\title{usgs_nearest_gage}
\usage{
usgs_nearest_gage(watershed_feature, watershed_json)
}
\arguments{
\item{watershed_feature}{a valid RomFeature for the watershed of interest}

\item{watershed_json}{a valid json model for the watershed of interest}
}
\description{
Find the closest gage for a watershed feature and model.
}
\details{
This function finds full drainage watersheds that contain a feature, and make an area match.
}
