% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{vahydro_post_metric_to_scenprop}
\alias{vahydro_post_metric_to_scenprop}
\title{Send a property to VAHydro with a simple one line command}
\usage{
vahydro_post_metric_to_scenprop(
  pid,
  varkey,
  propcode,
  propname,
  propvalue,
  ds = FALSE
)
}
\arguments{
\item{pid}{The unique property ID}

\item{varkey}{The variable key for the property}

\item{propcode}{The propcode, null OK}

\item{propname}{The property name - must be non-null}

\item{propvalue}{The numerical value}

\item{ds}{A valid Rom datasource}
}
\value{
A RomProperty object
}
\description{
Send a property to VAHydro with a simple one line command
}
\examples{
NA
}
\seealso{
NA
}
