% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TNC_IHA_Ports.R
\name{Water Month}
\alias{Water Month}
\alias{water.month}
\title{Water Month}
\usage{
water.month(x, label = FALSE, abbr = TRUE)
}
\arguments{
\item{x}{a date or time object or a numeric representing a calendar year month}

\item{label}{logical TRUE will return an ordered factor for month with the
month name as labels, FALSE will return a numeric}

\item{abbr}{logical. FALSE will abbreviate the name of the month in the ordered factor.}
}
\value{
If x is numeric, then returns a numeric. Otherwise, if label is TRUE,
 returns an ordered factor indicating the equivalent month of the water year.
 Otherwise, returns only the numeric that represents the month.
}
\description{
Extract the month of the water year from a date or time object.
}
\details{
Extract the month of the water year from a date or time object.  Can be returned
as a number 1:12, for October through September respectively, or an ordered factor. 
A user may alternatively enter a calendar year month as a numeric to return
the equivalent water year month
}
\author{
jason.e.law@gmail.com (imported to Hydrotools by Connor Brogan,connor.brogan@deq.virginia.gov)
}
