% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TNC_IHA_Ports.R
\name{Water Year}
\alias{Water Year}
\alias{water.year}
\title{Water Year}
\usage{
water.year(x)
}
\arguments{
\item{x}{a date-time object which can be handled by lubridate}
}
\value{
Returns the equivalent water year of the date x input by the uee
}
\description{
Return a numeric giving the water year for a date
}
\details{
Returns a number specifying the water year (\code{2010}) for a date. Water year is hardcoded here as Oct 1 - Sept
30
}
\author{
jason.e.law@gmail.com (imported to Hydrotools by Connor Brogan,connor.brogan@deq.virginia.gov)
}
