% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wshd_utils.R
\name{watershed_model_da}
\alias{watershed_model_da}
\title{watershed_model_da}
\usage{
watershed_model_da(watershed_json)
}
\arguments{
\item{watershed_json}{a valid json model for the watershed of interest}
}
\description{
Find the drainage area from a json model collection.
}
\details{
This function handles the varying area data formats that these models have historically used.
}
