% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WatershedModelNode.R
\name{WatershedModelNode}
\alias{WatershedModelNode}
\title{WatershedModelNode}
\value{
R6Class of type WatershedModelNode
}
\description{
Utility class for interacting with a watershed feature/model combo
}
\details{
Watershed Model Node data object

Has standard methods for managing data and meta data
}
\examples{
NA
}
\seealso{
NA
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ds}}{RomDataSource}

\item{\code{ds_om}}{RomDataSource points to legacy model db}

\item{\code{hydroid}}{unique ID from dh_feature database}

\item{\code{hydrocode}}{non-numeric identifier may reference ext source db}

\item{\code{bundle}}{type of feature}

\item{\code{ftype}}{sub-type of feature}

\item{\code{feature}}{a RomFeature entity}

\item{\code{pid}}{integer identifier}

\item{\code{version}}{descriptive code for model}

\item{\code{prop}}{as RomProperty}

\item{\code{json}}{as list}

\item{\code{elementid}}{as integer}

\item{\code{rocode}}{describes runoff type default = 'cbp6_lrseg'}

\item{\code{name}}{what is it called}

\item{\code{riverseg}}{we using code names}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-watershed_model_node-new}{\code{WatershedModelNode$new()}}
\item \href{#method-watershed_model_node-handle_config}{\code{WatershedModelNode$handle_config()}}
\item \href{#method-watershed_model_node-get_feature}{\code{WatershedModelNode$get_feature()}}
\item \href{#method-watershed_model_node-get_model_pid}{\code{WatershedModelNode$get_model_pid()}}
\item \href{#method-watershed_model_node-get_model}{\code{WatershedModelNode$get_model()}}
\item \href{#method-watershed_model_node-load_json_model}{\code{WatershedModelNode$load_json_model()}}
\item \href{#method-watershed_model_node-get_json_model}{\code{WatershedModelNode$get_json_model()}}
\item \href{#method-watershed_model_node-om_element_tree}{\code{WatershedModelNode$om_element_tree()}}
\item \href{#method-watershed_model_node-get_runoff_elids}{\code{WatershedModelNode$get_runoff_elids()}}
\item \href{#method-watershed_model_node-nearest_gage}{\code{WatershedModelNode$nearest_gage()}}
\item \href{#method-watershed_model_node-clone}{\code{WatershedModelNode$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-watershed_model_node-new"></a>}}
\if{latex}{\out{\hypertarget{method-watershed_model_node-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WatershedModelNode$new(ds, config = list(), ds_om = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ds}}{RomDataSource for remote and local storage (required)}

\item{\code{config}}{list of attributes to set, see also: to_list() for format}

\item{\code{ds_om}}{RomDataSource for legacy model connection (optional)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
object instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-watershed_model_node-handle_config"></a>}}
\if{latex}{\out{\hypertarget{method-watershed_model_node-handle_config}{}}}
\subsection{Method \code{handle_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WatershedModelNode$handle_config(config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{list of attributes to set/query}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
json of model from dH
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-watershed_model_node-get_feature"></a>}}
\if{latex}{\out{\hypertarget{method-watershed_model_node-get_feature}{}}}
\subsection{Method \code{get_feature()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WatershedModelNode$get_feature()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
json of model from dH
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-watershed_model_node-get_model_pid"></a>}}
\if{latex}{\out{\hypertarget{method-watershed_model_node-get_model_pid}{}}}
\subsection{Method \code{get_model_pid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WatershedModelNode$get_model_pid()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
json of model from dH
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-watershed_model_node-get_model"></a>}}
\if{latex}{\out{\hypertarget{method-watershed_model_node-get_model}{}}}
\subsection{Method \code{get_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WatershedModelNode$get_model()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
json of model from dH
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-watershed_model_node-load_json_model"></a>}}
\if{latex}{\out{\hypertarget{method-watershed_model_node-load_json_model}{}}}
\subsection{Method \code{load_json_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WatershedModelNode$load_json_model()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
TRUE or FALSE
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-watershed_model_node-get_json_model"></a>}}
\if{latex}{\out{\hypertarget{method-watershed_model_node-get_json_model}{}}}
\subsection{Method \code{get_json_model()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WatershedModelNode$get_json_model()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
json of model from dH
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-watershed_model_node-om_element_tree"></a>}}
\if{latex}{\out{\hypertarget{method-watershed_model_node-om_element_tree}{}}}
\subsection{Method \code{om_element_tree()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WatershedModelNode$om_element_tree(
  elementid,
  include_children = 1,
  max_recursion_level = -1,
  exclude_custom1 = -1,
  target_custom1 = -1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{elementid}}{target container to search}

\item{\code{include_children}}{recurse through children? default = 1}

\item{\code{max_recursion_level}}{integer how far to search if include_children = 1}

\item{\code{exclude_custom1}}{custom1 to exclude (use cova_upstream prevent upstream recursion)}

\item{\code{target_custom1}}{custom1 to find}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
dataframe of elementids
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-watershed_model_node-get_runoff_elids"></a>}}
\if{latex}{\out{\hypertarget{method-watershed_model_node-get_runoff_elids}{}}}
\subsection{Method \code{get_runoff_elids()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WatershedModelNode$get_runoff_elids()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of elementids corresponding to runoff props uses rocode prop to filter
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-watershed_model_node-nearest_gage"></a>}}
\if{latex}{\out{\hypertarget{method-watershed_model_node-nearest_gage}{}}}
\subsection{Method \code{nearest_gage()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WatershedModelNode$nearest_gage()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data frame with the closest gage
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-watershed_model_node-clone"></a>}}
\if{latex}{\out{\hypertarget{method-watershed_model_node-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WatershedModelNode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
