% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFLOW_xQy.R
\name{xQy}
\alias{xQy}
\title{xQy}
\usage{
xQy(
  gageDataIn,
  flowColumn = "Flow",
  dateColumn = "Date",
  AYS = "04-01",
  AYE = "03-31",
  startYear = NULL,
  endYear = NULL,
  x = 7,
  y = 10,
  IncludeSummerFlow = FALSE
)
}
\arguments{
\item{gageDataIn}{A data frame or 1-D zoo timeseries containing a date and
flow column of daily streamflows. If a zoo is passed in, the flowColumn
argument the flowColumn and dateColumn arguments are ignored}

\item{flowColumn}{The column name in gageDataIn that contains the flow values}

\item{dateColumn}{The column name in gageDataIn that contains the date
values. Will be converted to the date class.}

\item{AYS}{The start of the analysis season in the format MM-DD. For
instance, to calculate low flows across the water year, AYS = '10-01'}

\item{AYE}{The end of the analysis season in the format MM-DD. For instance,
to calculate low flows across the water year, AYE = '09-31'}

\item{startYear}{Any calendar year prior to this year will not be included.
User can leave this as NULL (default) to include all data.}

\item{endYear}{Any calendar year after this year will not be included. User
can leave this as NULL (default) to include all data.}

\item{x}{User may optionally calculate a custom xQy flow. 1Q10, 7Q10, 30Q10,
90Q10, and 30Q5 are included in the results already. This input serves as
the averaging period.}

\item{y}{User may optionally calculate a custom xQy flow. 1Q10, 7Q10, 30Q10,
90Q10, and 30Q5 are included in the results already. This input serves as
the return period}

\item{IncludeSummerFlow}{Included to match inputs of VPDES function. MDE
Summer flows are particular VPDES low flows used in rare analyses. not
included in hydrotools.}
}
\value{
A list of relevant objects that includes a list of the calculated low
flows Flows, the percentile of the calculate low flows pctg, the formatted
gage data formatted_data (which includes only data considered in analysis and the
rolling average flows), a data frame (annualMinimums) that includes a
summary of the rolling flows in each year and identifies the minimum flows,
data frames that identify when minimum flows occurred in each analysis year
for each flow (e.g., n1Q10_annDate), a series of vectors line
n1Q10_ModeMin that identifies the month when annual low flows of that
averaging period most commonly occur, a numeric timecheck that indicates
the number of days in each analysis year on a non-leap year
}
\description{
Compute VPDES low flows and other xQy style flows
}
\details{
This function calculates xQy style low flows from USGS or other flow
gage data. It takes in a dataframe that has at least a flow column and date
column specified by the user. From there, it divides the data into user
defined analysis seasons, using only complete seasons (adjusted for leap
years) in the analysis. It then used xQyComp() to calculte the probability
adjusted xQy flows and includes by deafult the 1Q10, 7Q10, 30Q10, 30Q5,
90Q10, and the harmonic mean.\cr
Users may enter analysis years or seasons such that the low flows can be
computed across entire user defined years (like the climate or water year)
or across only certain months. The function is careful to use only
independent analysis years, removing any averaged flow that may extend into
the next analysis year. This makes choosing the correct analysis season
very important.\cr
This is functionally equivalent to the VPDES' DFLOW R Shiny application.
}
\examples{
\dontrun{
gageDat <- dataRetrieval::readNWISdv("01631000","00060")
gageDat <- dataRetrieval::renameNWISColumns(gageDat)
gageDat <- gageDat[!grepl("P",gageDat$Flow_cd),]
low_flows <- xQy(gageDataIn = gageDat,
   flowColumn = "Flow", dateColumn = "Date",
   AYS = "07-13", AYE = "02-15",
   startYear = NULL, endYear = NULL,
   x = 7, y = 10,
   IncludeSummerFlow = FALSE)
low_flows$Flows
}
}
