% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DFLOW_xQy.R
\name{xQyComp}
\alias{xQyComp}
\title{xQyComp}
\usage{
xQyComp(xQy_ann, y)
}
\arguments{
\item{xQy_ann}{A vector of minimum annual averaged flows across duration x.
For instance, these should be annual minimum 7-day flows to calculate a
7Q10}

\item{y}{Return period i.e. the y in xQy style low flows. For instance, a
7Q10 requires y = 10}
}
\value{
The xQy flow or NA if insufficient data exists
}
\description{
Compute xQy low flow
}
\details{
xQyComp is used within the xQy() function to calculate the 7Q10 and
other relevant low flows. It takes in a vector of annual minimum flows
(xQy_ann) and the return period y to determine the xQy flow. These flows are
probability adjusted for years with 0 minimum flow using the USGS approach
documented in Austin et al. 2011
}
