% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drought_utils.R
\name{auto_forecast_cs}
\alias{auto_forecast_cs}
\title{auto_forecast_cs}
\usage{
auto_forecast_cs(
  gage_obj,
  yr = 2002,
  best_selection = "minimum",
  AGWRC = "lm_variable",
  start_mmdd = "03-01",
  end_mmdd = "10-01",
  min_error = -20,
  max_error = 20,
  trough_min_width = 15
)
}
\arguments{
\item{gage_obj}{A WaterGageDaily R6 object of a USGS gage}

\item{yr}{Numeric, default 2002. Which year should be analyzed for potential
forecasting lookback}

\item{best_selection}{Character, default "minimum". How should the best forecast
be identified? If "minimum", the longest forecast that estimated the observed low flow
within user tolerance and does not overpredict any troughs outside of
tolerance is selected.}

\item{AGWRC}{A of numeric or character given to
\code{WaterGageDaily$baseflow_forecast}. Value is the decay coefficient to
regress the flow on start_date for forecast. May be a single numeric to
allow for a constant forecast or a vector of numeric values to allow for a
variable forecast but must be of length days. Otherwise, may be
"lm_constant" to calculate a constant value from m and b or "lm_variable" to
have a variable value.}

\item{start_mmdd}{Character, default "03-01". The start of the period in
"mm-dd" to identify potential troughs and minimum observed flow (not used in
start date detection, which is hard coded to March - August)}

\item{end_mmdd}{Character, default "10-01". The end of the period in "mm-dd"
to identify potential troughs and minimum observed flow (not used in start
date detection, which is hard coded to March - August)}

\item{min_error}{Numeric, default -20. The low error bound in \% on trough/minimum
flow prediction such that an acceptable forecast has an error of at least
\code{min_error} (i.e. underprediction by at most 20\%)}

\item{max_error}{Numeric, default 10. The high error bound in \% on trough/minimum
flow prediction such that an acceptable forecast has an error of at most
\code{min_error} (i.e. overprediction by at most 10\%)}

\item{trough_min_width}{Numeric, default 15. The interval with which to decide
if a flow is a "trough", a local minimum worth evaluating as a potential
start point and/or forecast performance point}
}
\value{
A list with five entries:
\itemize{
\item beststart The start date of the best performing forecast
\item all_data A list of data frames named by forecast start date. These
data frames have 365 day forecasts with trough prediction performance for
each potential start date
\item analysis_data A list of data frames named by forecast start date. These
data frames are identical to all_data but have been filtered to only include
trough or minimum flows to demonstrate how the algorithim selected beststart
\item fc_plot A plot of the best selected forecast beginning at beststart and
terminating either 90 days later or 14 days after the last trough/minimum
flow
\item fc_all_plot A plot of all tested forecasts with start dates as red points.
The best forecast (which begins at beststart) is displayed in blue, and
forecasts before and after beststart are in orange and grey respectively
}
}
\description{
Find potential historic case study forecasts for a given year
}
\details{
This function finds all potential start dates in March - August to
run a summer no-rain baseflow forecast. It iterates through each potential
start date and evaluates which forecast consecutively predicts the most
troughs accurately or that which predicts the observed minimum flow in the
user period the most accurately based on user input. A list is returned to
the user with the best forecast plotted, a plot of all tested forecasts, a
list of forecasts from each start date, and a list of the analysis data from
each start date.
\itemize{
\item Users can set \code{start_mmdd} and \code{end_mmdd} to determine the period
between which minimum flows and troughs should be identified. Start dates
are only identified in the period March - August
\item \code{best_selection} determines how the functions decides which forecast
performs the best. "minimum" results in the selection of the forecast that
most closely predicts the observed minimum flow as long as no trough is
overpredicted by more than \code{max_error}. "consecutive" selects the method
that most accurately predicts the most troughs consecutively based on user
error input (any trough out of range, even if its the first, iterates the
function to the next potential). If no start date predicts troughs
consecutively, that which predicts the most overall (including
non-consecutive) is selected. The user is provided a warning if no gage was
found
\item \code{min_error} and \code{max_error} can be set to specify trough/minimum
flow prediction tolerance before program iterates to next start date
}
}
\examples{
\dontrun{
gage_obj <- hydrotools::WaterGageDaily$new(  gage_id = '02029000', ds_in = ds )
#Run iterative forecasts using variable AGWRC method and selecting best picked
start date via minimum observed flow performance
fc_cs <- auto_forecast_cs(
  gage_obj = gage_obj,
  yr = 1932,
  AGWRC = "lm_variable",
  best_selection = "minimum",
  start_mmdd = "03-01",
  end_mmdd = "09-30",
  min_error = -20,
  max_error = 20,
  trough_min_width = 15
)
#Best forecast starts on:
fc_cs$beststart
#Show all tested forecasts:
fc_cs$fc_all_plot
}
}
