% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cia_utils.R
\name{fn_ALL.upstream}
\alias{fn_ALL.upstream}
\title{Find Upstream Segments: A basic function that finds the segments immediately
upstream of a given segment(s). riv.seg may be a character vector of segments.}
\usage{
fn_ALL.upstream(AllSegList, riv.seg = NULL, getAllSegmentsOnly = TRUE)
}
\arguments{
\item{AllSegList}{Data frame that must contain riverseg and all river}

\item{riv.seg}{If user wishes to only return certain segments, they can}

\item{getAllSegmentsOnly}{If the user only wishes to get the ids of upstream}
}
\value{
string next upstream segment ID
}
\description{
Returns a vector of the segments immediately upstream of riv.seg
  based on the list in AllSegList
}
