% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn_extract_basin.R
\name{fn_extract_basin}
\alias{fn_extract_basin}
\title{Extracting Basin Function}
\usage{
fn_extract_basin(cia_data_frame, end_seg)
}
\arguments{
\item{cia_data_frame}{Data frame of cumulative impact data}

\item{end_seg}{Desired end river segment - river segment that is most downstream}
}
\value{
Trimmed cumulative impact data frame
}
\description{
Trims data frame to include only segments upstream of end segment
}
