% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fn_get_runfile}
\alias{fn_get_runfile}
\title{Retrieve Entire Run Log File from Old OM model}
\usage{
fn_get_runfile(elementid = -1, runid = -1, scenid = 37,
  site = "http://deq2.bse.vt.edu", cached = TRUE, outaszoo = TRUE,
  use_tz = FALSE)
}
\arguments{
\item{elementid}{integer OM model element id}

\item{runid}{integer run id}

\item{scenid}{integer - model domain ID}

\item{site}{URL of om server}

\item{cached}{boolean - use local copy or force refresh}

\item{outaszoo}{boolean return as a zoo timeseries if TRUE, or as data frame}

\item{usetz}{character pass in a custom timezone for zoo}
}
\value{
reference class of type openmi.om.equation
}
\description{
Retrieve Entire Run Log File from Old OM model
}
\examples{
NA
}
\seealso{
NA
}
