% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fn_get_timeseries}
\alias{fn_get_timeseries}
\title{Retrieve time series value from RESTful web service}
\usage{
fn_get_timeseries(inputs, site, token)
}
\arguments{
\item{inputs}{list of query terms (see RomTS::to_list())}

\item{site}{URL of om server}

\item{token}{for httpx authentication}
}
\value{
integer variable id
}
\description{
Retrieve time series value from RESTful web service
}
\examples{
NA
}
\seealso{
NA
}
