% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odbc_utils.R
\name{fn_guess_sql_where}
\alias{fn_guess_sql_where}
\title{Guess an SQL query from a simple list of inputs}
\usage{
fn_guess_sql_where(entity_type, pk, inputs, alias = "")
}
\arguments{
\item{entity_type}{= dh_feature, dh_properties, ...}

\item{pk}{= primary key column name, e.g. hydroid, pid, ...}

\item{inputs}{contents of record to post in list(pid=X, propname='nom', propvalue, ...)}

\item{alias}{assign an alias to a give table/view relation?}

}
\description{
Guess an SQL query from a simple list of inputs
}
\examples{
NA
}
