% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cia_utils.R
\name{fn_handletimestamp}
\alias{fn_handletimestamp}
\title{Handle Any Time/Date to Timestamp}
\usage{
fn_handletimestamp(ts)
}
\arguments{
\item{ts}{some type of date or timestamp}
}
\value{
single numeric value for UNix epoch
}
\description{
Try to bring any date or timestamp into a common format (Unix epoch, TBD)
}
