% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cia_utils.R
\name{fn_iha_7q10}
\alias{fn_iha_7q10}
\title{fn_iha_7q10}
\usage{
fn_iha_7q10(zoots)
}
\arguments{
\item{zoots}{a timeseries flormatted in zoo (required by IHA)}
}
\value{
singel numeric value for 7q10
}
\description{
provide the 7q10 from a given flow timeseries
}
\examples{
NA
}
\seealso{
NA
}
