% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cia_utils.R
\name{fn_iha_7q10}
\alias{fn_iha_7q10}
\title{fn_iha_7q10}
\usage{
fn_iha_7q10(zoots, flowColumnIn = "Flow", dateColumnIn = "Date")
}
\arguments{
\item{zoots}{a timeseries formatted in zoo or a data frame that contains a
date and flow column, to be specified by user}

\item{flowColumnIn}{If a dataframe is provided to the function, this is the
name of the column that contains the flow data for analysis. Ignored if a
zoo timeseries was provided.}

\item{dateColumnIn}{If a dataframe is provided to the function, this is the
name of the column that contains the dates. Ignored if a zoo timeseries was
provided}
}
\value{
single numeric value for 7Q10
}
\description{
Calculate the 7Q10 from a flow timeseries
}
\details{
This function was originally inspired by the Nature Conservancy's IHA package
and relied on \code{group2()}. However, now this function is just a simple
wrapper of \code{xQy()} to ensure consistency with the rest of DEQ. It takes
in a data frame or zoo timeseries and outputs the 7Q10 as a single value
numeric.
}
\examples{
flows <- dataRetrieval::readNWISdv("01631000","00060")
flows <- dataRetrieval::renameNWISColumns(flows)
#Convert flows to zoo
flows_zoo <- zoo::as.zoo(x = flows$Flow)
zoo::index(flows_zoo) <- flows$Date
fn_iha_7q10(flows_zoo)
}
\seealso{
xQy
}
