% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fn_search_vardefs}
\alias{fn_search_vardefs}
\title{fn_search_vardefs}
\usage{
fn_search_vardefs(config, var_defs_tmp)
}
\arguments{
\item{config}{= list(entity_type, featureid, pid = NULL, varid = NULL, startdate = NULL, enddate = NULL, tscode = NULL, tlid = NULL) timeline ID (not yet used)
Definitions for these items are defined in the hydrotools readme.}

\item{var_defs_tmp}{data frame to search}
}
\value{
data frame of propvalue or FALSE
}
\description{
Retrieve Property data from propvalues style data frame provided by user
}
\examples{
NA
#ds <- RomDataSource$new()
#fn_search_vardefs(config = list(hydroid = 1456),)
}
