% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cia_utils.R
\name{is.empty}
\alias{is.empty}
\title{is.empty}
\usage{
is.empty(x, ...)
}
\arguments{
\item{x}{an object to check its emptiness.}

\item{...}{additional arguments for sapply}
}
\value{
TRUE/FALSE
}
\description{
Checks common NULL, NA, length cases for empty values
}
\details{
Imported from rapportools and should be used sparingly as explicit length,
NA, and NULL checks are better for ensuring consistent data strucutre. From
rapportools: "Rails-inspired helper that checks if vector values are "empty",
i.e. if it's: NULL, zero-length, NA, NaN, FALSE, an empty string or 0. Note
that unlike its native R is.<something> sibling functions, is.empty is
vectorised (hence the "values")."
}
