% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_cia_table.R
\name{om_cia_table}
\alias{om_cia_table}
\title{Generate a table of CIA model result summary statistsics for both Rseg and Facility models, comparing across scenarios}
\usage{
om_cia_table(
  fac_model_info = FALSE,
  rseg_model_info = FALSE,
  runid.list = c("runid_6011", "runid_6012"),
  fac.metric.list = c("wd_mgd", "ps_mgd", "unmet30_mgd"),
  rseg.metric.list = c("Qout", "Qbaseline", "l30_Qout", "l90_Qout",
    "consumptive_use_frac", "wd_cumulative_mgd", "ps_cumulative_mgd"),
  site = "http://deq1.bse.vt.edu:81/d.dh",
  site_base = "http://deq1.bse.vt.edu:81"
)
}
\arguments{
\item{fac_model_info}{facility json list}

\item{rseg_model_info}{riverseg json list}

\item{runid.list}{list of runids of interest}

\item{fac.metric.list}{list of facility metrics of interest}

\item{rseg.metric.list}{list of riverseg metrics of interest}

\item{site}{vhydro url}

\item{site_base}{base vahydro url}
}
\value{
stats.df dataframe of summary stats
}
\description{
Generate a table of CIA model result summary statistsics for both Rseg and Facility models, comparing across scenarios
}
\examples{
NA
}
\seealso{
NA
}
