% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_demand_table.R
\name{om_demand_table}
\alias{om_demand_table}
\title{Generate a table of model demand inputs, comparing across scenarios}
\usage{
om_demand_table(model_info = FALSE, runid.list = c("runid_6011", "runid_6012"))
}
\arguments{
\item{model_info}{model json list}

\item{runid.list}{list of runids of interest}
}
\value{
demands.df dataframe of scenario demands
}
\description{
Generate a table of model demand inputs, comparing across scenarios
}
\examples{
NA
}
\seealso{
NA
}
