% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_get_rundata.R
\name{om_get_rundata}
\alias{om_get_rundata}
\title{The base class for executable equation based meta-model components.}
\usage{
om_get_rundata(
  elid,
  runid,
  site = "http://deq2.bse.vt.edu",
  cached = FALSE,
  hydrowarmup = TRUE
)
}
\arguments{
\item{elid}{integer OM model element id}

\item{runid}{integer run id}

\item{site}{URL of om server}

\item{cached}{boolean - if TRUE will use recently stored local copy}

\item{hydrowarmup}{boolean - if TRUE will trim beginning of model time frame}
}
\value{
reference class of type openmi.om.equation
}
\description{
The base class for executable equation based meta-model components.
}
\examples{
NA
}
\seealso{
NA
}
