% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_vahydro_metric_grid.R
\name{om_vahydro_metric_grid}
\alias{om_vahydro_metric_grid}
\title{The base class for executable equation based meta-model components.}
\usage{
om_vahydro_metric_grid(metric, runids, featureid = "all",
  entity_type = "dh_feature", bundle = "watershed",
  ftype = "vahydro", model_version = "vahydro-1.0",
  base_url = "http://deq2.bse.vt.edu/d.dh/entity-model-prop-level-export")
}
\value{
reference class of type openmi.om.equation
}
\description{
The base class for executable equation based meta-model components.
}
