% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_vahydro_metric_grid.R
\name{om_vahydro_metric_grid}
\alias{om_vahydro_metric_grid}
\title{The base class for executable equation based meta-model components.}
\usage{
om_vahydro_metric_grid(
  metric,
  runids,
  featureid = "all",
  entity_type = "dh_feature",
  bundle = "watershed",
  ftype = "vahydro",
  model_version = "vahydro-1.0",
  base_url = "http://deq2.bse.vt.edu/d.dh/entity-model-prop-level-export"
)
}
\arguments{
\item{metric}{character deprecated in favor of runids array}

\item{runids}{dataframe with model and run info}

\item{featureid}{integer or 'all'}

\item{entity_type}{character}

\item{bundle}{character}

\item{ftype}{character feature type}

\item{model_version}{character default 'vahydro-1.0'}

\item{base_url}{character deprecated to be replaced by global datasource}
}
\value{
reference class of type openmi.om.equation
}
\description{
The base class for executable equation based meta-model components.
}
\examples{
NA
}
\seealso{
NA
}
