% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/om_vahydro_metric_grid.R
\name{om_vahydro_metric_grid_sql}
\alias{om_vahydro_metric_grid_sql}
\title{Helper function to generate a metric grid via SQL
featureid,entity_type,bundle,ftype,model_version, runid, metric}
\usage{
om_vahydro_metric_grid_sql(
  featureid,
  entity_type,
  bundle,
  ftype,
  model_version,
  runid,
  metric
)
}
\arguments{
\item{featureid}{integer or 'all'}

\item{entity_type}{character}

\item{bundle}{character}

\item{ftype}{character feature type}

\item{model_version}{character default 'vahydro-1.0'}

\item{runid}{run info}

\item{metric}{character deprecated in favor of runids array}
}
\value{
string with executable SQL query
}
\description{
Helper function to generate a metric grid via SQL
featureid,entity_type,bundle,ftype,model_version, runid, metric
}
\examples{
NA
}
\seealso{
NA
}
