% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openmi.om.matrix.R
\name{openmi.om.matrix}
\alias{openmi.om.matrix}
\title{The base class for matrix/lokup table meta-model components.}
\value{
R6 class of type openmi.om.matrix
}
\description{
The base class for matrix/lokup table meta-model components.

The base class for matrix/lokup table meta-model components.
}
\seealso{
openmi.om.base
}
\section{Super classes}{
\code{\link[hydrotools:openmi.om.base]{hydrotools::openmi.om.base}} -> \code{\link[hydrotools:openmi.om.linkableComponent]{hydrotools::openmi.om.linkableComponent}} -> \code{openmi.om.matrix}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{datamatrix}}{holds the actual table}

\item{\code{colindex}}{holds the column lookup variable name (ustabe keycol1)}

\item{\code{rowindex}}{holds the column lookup variable name (ustabe keycol2)}

\item{\code{coltype}}{holds the column lookup type (ustabe lutype1)}

\item{\code{rowtype}}{holds the row lookup type (ustabe lutype2)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-openmi.om.matrix-set_prop}{\code{openmi.om.matrix$set_prop()}}
\item \href{#method-openmi.om.matrix-parse_class_specific}{\code{openmi.om.matrix$parse_class_specific()}}
\item \href{#method-openmi.om.matrix-update}{\code{openmi.om.matrix$update()}}
\item \href{#method-openmi.om.matrix-findMatch}{\code{openmi.om.matrix$findMatch()}}
\item \href{#method-openmi.om.matrix-exactMatch}{\code{openmi.om.matrix$exactMatch()}}
\item \href{#method-openmi.om.matrix-interpolate}{\code{openmi.om.matrix$interpolate()}}
\item \href{#method-openmi.om.matrix-stairStep}{\code{openmi.om.matrix$stairStep()}}
\item \href{#method-openmi.om.matrix-closest}{\code{openmi.om.matrix$closest()}}
\item \href{#method-openmi.om.matrix-window}{\code{openmi.om.matrix$window()}}
\item \href{#method-openmi.om.matrix-init}{\code{openmi.om.matrix$init()}}
\item \href{#method-openmi.om.matrix-clone}{\code{openmi.om.matrix$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="addComponent"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-addComponent'><code>hydrotools::openmi.om.base$addComponent()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="addInput"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-addInput'><code>hydrotools::openmi.om.base$addInput()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="add_component"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-add_component'><code>hydrotools::openmi.om.base$add_component()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="asJSON"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-asJSON'><code>hydrotools::openmi.om.base$asJSON()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="finish"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-finish'><code>hydrotools::openmi.om.base$finish()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="getInputs"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-getInputs'><code>hydrotools::openmi.om.base$getInputs()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="getValue"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-getValue'><code>hydrotools::openmi.om.base$getValue()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="get_component_id"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-get_component_id'><code>hydrotools::openmi.om.base$get_component_id()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="initialize"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-initialize'><code>hydrotools::openmi.om.base$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="logState"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-logState'><code>hydrotools::openmi.om.base$logState()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="log_debug"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-log_debug'><code>hydrotools::openmi.om.base$log_debug()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="orderOperations"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-orderOperations'><code>hydrotools::openmi.om.base$orderOperations()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="parse_openmi"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-parse_openmi'><code>hydrotools::openmi.om.base$parse_openmi()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="prepare"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-prepare'><code>hydrotools::openmi.om.base$prepare()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="set_sub_prop"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-set_sub_prop'><code>hydrotools::openmi.om.base$set_sub_prop()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="set_vars"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-set_vars'><code>hydrotools::openmi.om.base$set_vars()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="settable"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-settable'><code>hydrotools::openmi.om.base$settable()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="step"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-step'><code>hydrotools::openmi.om.base$step()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="stepChildren"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-stepChildren'><code>hydrotools::openmi.om.base$stepChildren()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="openmi.om.base" data-id="validate"><a href='../../hydrotools/html/openmi.om.base.html#method-openmi.om.base-validate'><code>hydrotools::openmi.om.base$validate()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.matrix-set_prop"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.matrix-set_prop}{}}}
\subsection{Method \code{set_prop()}}{
Set propvalues and parse matrix into datamatrix field, if
 applicable
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.matrix$set_prop(propname, propvalue, format = "raw")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{propname}}{which attribute}

\item{\code{propvalue}}{what value}

\item{\code{format}}{of propvalue}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NA
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.matrix-parse_class_specific"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.matrix-parse_class_specific}{}}}
\subsection{Method \code{parse_class_specific()}}{
Custom parser
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.matrix$parse_class_specific(propvalue)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{propvalue}}{from some custom classformat reading implementation}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
a settable data value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.matrix-update"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.matrix-update}{}}}
\subsection{Method \code{update()}}{
update does the lookup QC
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.matrix$update()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.matrix-findMatch"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.matrix-findMatch}{}}}
\subsection{Method \code{findMatch()}}{
findMatch looks into a single dimension table using the
  method ixtype provided by user
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.matrix$findMatch(dm, ixval, ixcol, ixtype = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dm}}{array to search}

\item{\code{ixval}}{key value to search for matches}

\item{\code{ixcol}}{Column to search for}

\item{\code{ixtype}}{what kind of lookup to perform?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
matching value (with interpolation if ixtype allows it)
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.matrix-exactMatch"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.matrix-exactMatch}{}}}
\subsection{Method \code{exactMatch()}}{
exactMatch looks for keys and returns the appropriate row of
  the matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.matrix$exactMatch(dm, ixval, ixcol)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dm}}{matrix to search}

\item{\code{ixval}}{key value to search for matches}

\item{\code{ixcol}}{Column to search for}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
matching value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.matrix-interpolate"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.matrix-interpolate}{}}}
\subsection{Method \code{interpolate()}}{
interpolate searches by key and calculates if no exact match
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.matrix$interpolate(dm, ixval, ixcol)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dm}}{matrix to search}

\item{\code{ixval}}{key value to search for matches}

\item{\code{ixcol}}{Column to search for}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
interpolated value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.matrix-stairStep"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.matrix-stairStep}{}}}
\subsection{Method \code{stairStep()}}{
stairStep searches by key and select closest previous (i.e.
  without exceeding)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.matrix$stairStep(dm, ixval, ixcol)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dm}}{matrix to search}

\item{\code{ixval}}{key value to search for matches}

\item{\code{ixcol}}{Column to search for}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
closest value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.matrix-closest"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.matrix-closest}{}}}
\subsection{Method \code{closest()}}{
closest select closest value, similar to stair step but
  allows exceeding
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.matrix$closest(dm, ixval, ixcol)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dm}}{array to search}

\item{\code{ixval}}{key value to search for matches}

\item{\code{ixcol}}{Column to search for}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
closest value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.matrix-window"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.matrix-window}{}}}
\subsection{Method \code{window()}}{
window does what???
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.matrix$window(dm, ixval, ixoff)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dm}}{array to search}

\item{\code{ixval}}{key to search for}

\item{\code{ixoff}}{index offset}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
calculated value
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.matrix-init"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.matrix-init}{}}}
\subsection{Method \code{init()}}{
init sets up for viewing or run and parses key matrix
  variables into the relevant fields on openmi.om.matrix
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.matrix$init()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
NULL, but returns an instance of this matrix
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-openmi.om.matrix-clone"></a>}}
\if{latex}{\out{\hypertarget{method-openmi.om.matrix-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{openmi.om.matrix$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
