% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RomDataSource.R
\name{RomDataSource}
\alias{RomDataSource}
\title{Global object store holds references to all objects of a given domain
and pointers to remote data source if applicable.
if remote data source is set, all data can be synched back to it
no mixing of data sources is allowed}
\value{
reference class of type openmi.om.base.
}
\description{
Data Source containing tables and methods of Model data objects
}
\details{
Provides integrated, queryable universe of features, observations and meta data
}
\examples{
NA
}
\seealso{
NA
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{base_url}}{URL of some RESTful repository}

\item{\code{timeline}}{for default time series data}

\item{\code{tsvalues}}{table of time series data}

\item{\code{props}}{table of object properties (can be contained by objects)}

\item{\code{features}}{table of physical features}

\item{\code{var_defs}}{table of variable definitions}

\item{\code{admin_features}}{table of adminreg features}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RomDataSource$new()}}
\item \href{#method-get_token}{\code{RomDataSource$get_token()}}
\item \href{#method-clone}{\code{RomDataSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomDataSource$new(base_url)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{base_url}}{URL of some RESTful repository}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
object instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_token"></a>}}
\if{latex}{\out{\hypertarget{method-get_token}{}}}
\subsection{Method \code{get_token()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomDataSource$get_token()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing sets internal private token
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomDataSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
