% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RomDataSource.R
\name{RomDataSource}
\alias{RomDataSource}
\title{Global object store holds references to all objects of a given domain
and pointers to remote data source if applicable.
if remote data source is set, all data can be synched back to it
no mixing of data sources is allowed}
\value{
reference class of type openmi.om.base.
}
\description{
Data Source containing tables and methods of Model data objects
}
\details{
Provides integrated, queryable universe of features, observations and meta data
}
\examples{
NA
}
\seealso{
NA
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{site}}{URL of some RESTful repository}

\item{\code{username}}{to connect to RESTful repository}

\item{\code{timeline}}{for default time series data}

\item{\code{tsvalues}}{table of time series data}

\item{\code{props}}{table of object properties (can be contained by objects)}

\item{\code{features}}{table of physical features}

\item{\code{var_defs}}{table of variable definitions}

\item{\code{admin_features}}{table of adminreg features}

\item{\code{ts_cache}}{list of ts objects instantiated}

\item{\code{feature_cache}}{list of feature objects instantiated}

\item{\code{prop_cache}}{list of prop objects instantiated}

\item{\code{var_cache}}{list of var objects instantiated}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{username}}{to connect to RESTful repository}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RomDataSource$new()}}
\item \href{#method-get_token}{\code{RomDataSource$get_token()}}
\item \href{#method-get_vardef}{\code{RomDataSource$get_vardef()}}
\item \href{#method-get_props}{\code{RomDataSource$get_props()}}
\item \href{#method-get_ts}{\code{RomDataSource$get_ts()}}
\item \href{#method-auth_read}{\code{RomDataSource$auth_read()}}
\item \href{#method-set_ts}{\code{RomDataSource$set_ts()}}
\item \href{#method-set_vardef}{\code{RomDataSource$set_vardef()}}
\item \href{#method-post}{\code{RomDataSource$post()}}
\item \href{#method-clone}{\code{RomDataSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomDataSource$new(site, rest_uname = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{site}}{URL of some RESTful repository}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
object instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_token"></a>}}
\if{latex}{\out{\hypertarget{method-get_token}{}}}
\subsection{Method \code{get_token()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomDataSource$get_token()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing sets internal private token
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_vardef"></a>}}
\if{latex}{\out{\hypertarget{method-get_vardef}{}}}
\subsection{Method \code{get_vardef()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomDataSource$get_vardef(varkey, force_update = FALSE, debug = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{varkey}}{= variable key}

\item{\code{force_update}}{Check remote repository for new info?}

\item{\code{debug}}{show info}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing sets internal private token
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_props"></a>}}
\if{latex}{\out{\hypertarget{method-get_props}{}}}
\subsection{Method \code{get_props()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomDataSource$get_props(config, return_type = "df", force_refresh = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{= list(entity_type, featureid, tid = NULL, varid = NULL, tstime = NULL, tsendtime = NULL, tscode = NULL, tlid = NULL) timeline ID (not yet used)}

\item{\code{return_type}}{'df' (data.frame) or 'object'}

\item{\code{force_refresh}}{if this ds has a remote source, whether to pull anew}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing sets internal private token
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ts"></a>}}
\if{latex}{\out{\hypertarget{method-get_ts}{}}}
\subsection{Method \code{get_ts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomDataSource$get_ts(config, return_type = "data.frame", force_refresh = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{= list(entity_type, featureid, tid = NULL, varid = NULL, tstime = NULL, tsendtime = NULL, tscode = NULL, tlid = NULL) timeline ID (not yet used)}

\item{\code{return_type}}{'df' (data.frame) or 'object'}

\item{\code{force_refresh}}{if this ds has a remote source, whether to pull anew}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing sets internal private token
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-auth_read"></a>}}
\if{latex}{\out{\hypertarget{method-auth_read}{}}}
\subsection{Method \code{auth_read()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomDataSource$auth_read(
  uri,
  content_type = "text/csv",
  delim = ",",
  enc = "xml"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{uri}}{remote address to retrieve data}

\item{\code{content_type}}{http content-type}

\item{\code{delim}}{delimiter}

\item{\code{enc}}{encoding}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
result of web request
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_ts"></a>}}
\if{latex}{\out{\hypertarget{method-set_ts}{}}}
\subsection{Method \code{set_ts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomDataSource$set_ts(ts)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ts}}{= list(entity_type, featureid, tid = NULL, varid = NULL, tstime = NULL, tsendtime = NULL, tscode = NULL, tlid = NULL) timeline ID (not yet used)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing sets internal private token
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_vardef"></a>}}
\if{latex}{\out{\hypertarget{method-set_vardef}{}}}
\subsection{Method \code{set_vardef()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomDataSource$set_vardef(var_def)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{var_def}}{= list(varid, varkey, varname, varunits, varcode,...)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
local df index?
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-post"></a>}}
\if{latex}{\out{\hypertarget{method-post}{}}}
\subsection{Method \code{post()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomDataSource$post(entity_type, pk, config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{entity_type}}{= dh_feature, dh_properties, ...}

\item{\code{pk}}{= primary key column name, e.g. hydroid, pid, ...}

\item{\code{config}}{= contents of record to post in list(pid, propname, propvalue, ...)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
local df index?
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomDataSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
