% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RomDataSource.R
\name{RomDataSource}
\alias{RomDataSource}
\title{Global object store holds references to all objects of a given domain
and pointers to remote data source if applicable.
if remote data source is set, all data can be synched back to it
no mixing of data sources is allowed}
\value{
reference class of type openmi.om.base.
}
\description{
Data Source containing tables and methods of Model data objects
}
\details{
Provides integrated, queryable universe of features, observations and meta data
}
\examples{
NA
}
\seealso{
NA
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{site}}{URL of some RESTful repository}

\item{\code{timeline}}{for default time series data}

\item{\code{tsvalues}}{table of time series data}

\item{\code{props}}{table of object properties (can be contained by objects)}

\item{\code{features}}{table of physical features}

\item{\code{var_defs}}{table of variable definitions}

\item{\code{admin_features}}{table of adminreg features}

\item{\code{ts_cache}}{list of ts objects instantiated}

\item{\code{feature_cache}}{list of feature objects instantiated}

\item{\code{prop_cache}}{list of prop objects instantiated}

\item{\code{var_cache}}{list of var objects instantiated}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{RomDataSource$new()}}
\item \href{#method-get_token}{\code{RomDataSource$get_token()}}
\item \href{#method-get_ts}{\code{RomDataSource$get_ts()}}
\item \href{#method-clone}{\code{RomDataSource$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomDataSource$new(site)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{site}}{URL of some RESTful repository}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
object instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_token"></a>}}
\if{latex}{\out{\hypertarget{method-get_token}{}}}
\subsection{Method \code{get_token()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomDataSource$get_token()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
nothing sets internal private token
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_ts"></a>}}
\if{latex}{\out{\hypertarget{method-get_ts}{}}}
\subsection{Method \code{get_ts()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomDataSource$get_ts(config, return_type = "list", force_refresh = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{= list(entity_type, featureid, tid = NULL, varid = NULL, tstime = NULL, tsendtime = NULL, tscode = NULL, tlid = NULL) timeline ID (not yet used)}

\item{\code{return_type}}{'list' or 'object'}

\item{\code{force_refresh}}{if this ds has a remote source, whether to pull anew}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
nothing sets internal private token
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomDataSource$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
