% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RomEntity.R
\name{RomEntity}
\alias{RomEntity}
\title{Base entity data object}
\value{
reference class of type openmi.om.base.

an updated config if necessary or FALSE if it fails

loads the varid

the data from the remote connection

nothing, but loads the objects plugin
}
\description{
Object for storing a single entity with attribute and timeseries related
}
\details{
Has standard methods for managing data and meta data
}
\examples{
NA
}
\seealso{
NA
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{what is it called}

\item{\code{base_entity_type}}{kind of entity}

\item{\code{pk_name}}{the name of this entity's pk column}

\item{\code{entity_id}}{unique ID of entity}

\item{\code{sql_select_from}}{syntax to use to select via an odbc or other SQL based datasource}

\item{\code{has_vardef}}{is pluggable?}

\item{\code{varid}}{(optional) integer field for pluggable entities}

\item{\code{vardef}}{(optional) full RomVariableDefinition}

\item{\code{plugin}}{(optional) instance of dHVariablePlugin class}

\item{\code{datasource}}{RomDataSource}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RomEntity-get_id}{\code{RomEntity$get_id()}}
\item \href{#method-RomEntity-propvalues}{\code{RomEntity$propvalues()}}
\item \href{#method-RomEntity-tsvalues}{\code{RomEntity$tsvalues()}}
\item \href{#method-RomEntity-new}{\code{RomEntity$new()}}
\item \href{#method-RomEntity-handle_config}{\code{RomEntity$handle_config()}}
\item \href{#method-RomEntity-insure_varid}{\code{RomEntity$insure_varid()}}
\item \href{#method-RomEntity-get_vardef}{\code{RomEntity$get_vardef()}}
\item \href{#method-RomEntity-load_data}{\code{RomEntity$load_data()}}
\item \href{#method-RomEntity-load_plugin}{\code{RomEntity$load_plugin()}}
\item \href{#method-RomEntity-to_list}{\code{RomEntity$to_list()}}
\item \href{#method-RomEntity-save}{\code{RomEntity$save()}}
\item \href{#method-RomEntity-delete}{\code{RomEntity$delete()}}
\item \href{#method-RomEntity-clone}{\code{RomEntity$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomEntity-get_id"></a>}}
\if{latex}{\out{\hypertarget{method-RomEntity-get_id}{}}}
\subsection{Method \code{get_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomEntity$get_id()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
get_id the unique id of this entity alias to remote pkid, subclassed as function
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomEntity-propvalues"></a>}}
\if{latex}{\out{\hypertarget{method-RomEntity-propvalues}{}}}
\subsection{Method \code{propvalues()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomEntity$propvalues(propname = NULL, varid = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{propname}}{optional name to filter}

\item{\code{varid}}{option variable to filter}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
propvalues unique properties of this entity
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomEntity-tsvalues"></a>}}
\if{latex}{\out{\hypertarget{method-RomEntity-tsvalues}{}}}
\subsection{Method \code{tsvalues()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomEntity$tsvalues(varkey = NULL, tstime = NULL, tsendtime = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{varkey}}{option variable to filter}

\item{\code{tstime}}{timespan begin}

\item{\code{tsendtime}}{timespan end}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
tsvalues unique timeseries records for this entity
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomEntity-new"></a>}}
\if{latex}{\out{\hypertarget{method-RomEntity-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomEntity$new(datasource = NULL, config = list(), load_remote = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datasource}}{RESTful repository (optional)}

\item{\code{config}}{list of attributes to set, see also: to_list() for format}

\item{\code{load_remote}}{automatically query REST data source for matches?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
object instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomEntity-handle_config"></a>}}
\if{latex}{\out{\hypertarget{method-RomEntity-handle_config}{}}}
\subsection{Method \code{handle_config()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomEntity$handle_config(config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomEntity-insure_varid"></a>}}
\if{latex}{\out{\hypertarget{method-RomEntity-insure_varid}{}}}
\subsection{Method \code{insure_varid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomEntity$insure_varid(config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomEntity-get_vardef"></a>}}
\if{latex}{\out{\hypertarget{method-RomEntity-get_vardef}{}}}
\subsection{Method \code{get_vardef()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomEntity$get_vardef(config = FALSE, refresh = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{list of attributes to set, see also: to_list() for format}

\item{\code{refresh}}{automatically refresh var info?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomEntity-load_data"></a>}}
\if{latex}{\out{\hypertarget{method-RomEntity-load_data}{}}}
\subsection{Method \code{load_data()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomEntity$load_data(config, load_remote)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{}

\item{\code{load_remote}}{automatically query remote data source for matches?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomEntity-load_plugin"></a>}}
\if{latex}{\out{\hypertarget{method-RomEntity-load_plugin}{}}}
\subsection{Method \code{load_plugin()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomEntity$load_plugin()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomEntity-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-RomEntity-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomEntity$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of object attributes suitable for input to new() and from_list() methods
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomEntity-save"></a>}}
\if{latex}{\out{\hypertarget{method-RomEntity-save}{}}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomEntity$save(push_remote = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{push_remote}}{update locally only or push to remote database}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomEntity-delete"></a>}}
\if{latex}{\out{\hypertarget{method-RomEntity-delete}{}}}
\subsection{Method \code{delete()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomEntity$delete(delete_remote = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{delete_remote}}{update locally only or push to remote database}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomEntity-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RomEntity-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomEntity$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
