% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RomFeature.R
\name{RomFeature}
\alias{RomFeature}
\title{Feature entity data object}
\value{
feature class of type RomFeature
}
\description{
Object for storing a single feature with attribute and timeseries related
}
\details{
Has standard methods for managing data and meta data
}
\examples{
NA
}
\seealso{
NA
}
\section{Super class}{
\code{\link[hydrotools:RomEntity]{hydrotools::RomEntity}} -> \code{RomFeature}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{base_entity_type}}{kind of entity}

\item{\code{pk_name}}{the name of this entity's pk column}

\item{\code{name}}{of this entity}

\item{\code{hydrocode}}{alpha code for this entity from original dataset}

\item{\code{ftype}}{feature type}

\item{\code{hydroid}}{unique ID (integer)}

\item{\code{bundle}}{main content type, i.e. facility, well, intake, ...}

\item{\code{fstatus}}{entity status}

\item{\code{description}}{notes field}

\item{\code{mps}}{linked features}

\item{\code{geom}}{feature geometry WKT}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RomFeature-new}{\code{RomFeature$new()}}
\item \href{#method-RomFeature-get_id}{\code{RomFeature$get_id()}}
\item \href{#method-RomFeature-to_list}{\code{RomFeature$to_list()}}
\item \href{#method-RomFeature-from_list}{\code{RomFeature$from_list()}}
\item \href{#method-RomFeature-add_mp}{\code{RomFeature$add_mp()}}
\item \href{#method-RomFeature-save}{\code{RomFeature$save()}}
\item \href{#method-RomFeature-clone}{\code{RomFeature$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="RomEntity" data-id="handle_config"><a href='../../hydrotools/html/RomEntity.html#method-RomEntity-handle_config'><code>hydrotools::RomEntity$handle_config()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="RomEntity" data-id="load_data"><a href='../../hydrotools/html/RomEntity.html#method-RomEntity-load_data'><code>hydrotools::RomEntity$load_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="RomEntity" data-id="propvalues"><a href='../../hydrotools/html/RomEntity.html#method-RomEntity-propvalues'><code>hydrotools::RomEntity$propvalues()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="RomEntity" data-id="tsvalues"><a href='../../hydrotools/html/RomEntity.html#method-RomEntity-tsvalues'><code>hydrotools::RomEntity$tsvalues()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomFeature-new"></a>}}
\if{latex}{\out{\hypertarget{method-RomFeature-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomFeature$new(datasource = NULL, config, load_remote = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{datasource}}{RESTful repository object}

\item{\code{config}}{list of attributes to set, see also: to_list() for format}

\item{\code{load_remote}}{automatically query REST data source for matches?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
object instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomFeature-get_id"></a>}}
\if{latex}{\out{\hypertarget{method-RomFeature-get_id}{}}}
\subsection{Method \code{get_id()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomFeature$get_id()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
get_id the unique id of this entity alias to remote pkid, subclassed as function
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomFeature-to_list"></a>}}
\if{latex}{\out{\hypertarget{method-RomFeature-to_list}{}}}
\subsection{Method \code{to_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomFeature$to_list()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of object attributes suitable for input to new() and from_list() methods
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomFeature-from_list"></a>}}
\if{latex}{\out{\hypertarget{method-RomFeature-from_list}{}}}
\subsection{Method \code{from_list()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomFeature$from_list(config)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{list of attributes to set, see also: to_list() for format}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomFeature-add_mp"></a>}}
\if{latex}{\out{\hypertarget{method-RomFeature-add_mp}{}}}
\subsection{Method \code{add_mp()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomFeature$add_mp(thismp)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{thismp}}{mp entity}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
add a connected MP. TBD
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomFeature-save"></a>}}
\if{latex}{\out{\hypertarget{method-RomFeature-save}{}}}
\subsection{Method \code{save()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomFeature$save(push_remote = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{push_remote}}{update locally only or push to remote database}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomFeature-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RomFeature-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomFeature$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
