% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RomMetricNetwork.R
\name{RomMetricNetwork}
\alias{RomMetricNetwork}
\title{RomMetricNetwork}
\value{
R6 Object of class RomMetricNetwork
}
\description{
Utility class for examining, querying, and manipulating
  network-style data
}
\details{
Network Data Object

This R6 object has standard methods for managing the data and meta
  data related to network-style data like river networks or model metric
  returned from \code{om_vahydro_metric_grid()}.
  This object can query data from \code{om_vahydro_metric_grid()} or take in
  a data frame and help the user compile a network graph from source ids or
  riversegments. Methods on this object allow for cumulative calculations
  across the network, downstream/upstream extraction, and diagram creation.
}
\examples{
\dontrun{
self <- RomMetricNetwork$new(
 ds = ds,
 datasource = "om",
 metrics = c('Qout','wd_cumulative_mgd','l90_Qout', 'Qavailable_90_mgd'),
 runids = c("runid_400", "runid_600"),
 featureid = 'all',
 bundle = 'all',
 ftype = 'all',
 model_version = 'vahydro-1.0',
 force_unique_src = FALSE
)
View(self$network_data)
upstream_nodes <- self$get_node_relation(value = "riverseg")
self$set_numeric_nodes()
upstream_nodes_num <- self$get_node_relation(value = "riverseg")

self$network_data$newdist <- 1
cumulative_sum <- self$calc_cumulative("newdist")
all(lengths(cumulative_sum$upstream_nodes) == cumulative_sum$cumultive_sum )

wshd <- self$extact_node_network(src_node = self$network_data$src_node[1])
wshd <- self$extact_node_network(src_node = self$network_data$src_node[1],
                                direction = "downstream")

netSF <- self$get_spatial_network(value_col = "Qavailable_90_mgd_runid_600",
value_label = "WA_MGD")
plot(netSF$dh_geofield,col = netSF$color)

self$render_force_diagram()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ds}}{RomDataSource often provided in DEQ config files and features a
local data storage or a connection to a database/RESTful service}

\item{\code{datasource}}{Either 'om' or a data frame input inidicating how this
object instance should derive the metric data. If 'om' is selected,
\code{om_vahydro_metric_grid()} will be used to get model metrics and
methods will assume a riverseg column is present that contains watersheds
IDs in the form JU3_6690_7280 where 6690 identifies this segment and 7280
is the segment (source node) to which it flows (destination node)}

\item{\code{entity_type}}{character (single-length) to represent the table to query
for models, defaults to \code{dh_feature_fielded}}

\item{\code{entity_id_col}}{Character; defaults to 'hydroid' and represents the primary key
field of the entity type table (defaults dh_feature_fielded)}

\item{\code{dest_node_col}}{If datasource is a dataframe, which field identifies the
destination node that this node flows to?}

\item{\code{src_node_col}}{If datasource is a dataframe, which field is the source
node that identifies this node?}

\item{\code{network_data}}{A data frame of data with a source and destination
node field that identifies how data travels upstream and downstream
within the network represented by the data frame}

\item{\code{network_graph}}{An \code{igraph} style object from the package of
the same name with vertices defined by the source node column}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RomMetricNetwork-new}{\code{RomMetricNetwork$new()}}
\item \href{#method-RomMetricNetwork-get_om_rseg_node}{\code{RomMetricNetwork$get_om_rseg_node()}}
\item \href{#method-RomMetricNetwork-force_unique_src_ids}{\code{RomMetricNetwork$force_unique_src_ids()}}
\item \href{#method-RomMetricNetwork-set_numeric_nodes}{\code{RomMetricNetwork$set_numeric_nodes()}}
\item \href{#method-RomMetricNetwork-get_node_relation}{\code{RomMetricNetwork$get_node_relation()}}
\item \href{#method-RomMetricNetwork-extact_node_network}{\code{RomMetricNetwork$extact_node_network()}}
\item \href{#method-RomMetricNetwork-calc_cumulative}{\code{RomMetricNetwork$calc_cumulative()}}
\item \href{#method-RomMetricNetwork-get_spatial_network}{\code{RomMetricNetwork$get_spatial_network()}}
\item \href{#method-RomMetricNetwork-render_force_diagram}{\code{RomMetricNetwork$render_force_diagram()}}
\item \href{#method-RomMetricNetwork-handle_config}{\code{RomMetricNetwork$handle_config()}}
\item \href{#method-RomMetricNetwork-handle_config_item}{\code{RomMetricNetwork$handle_config_item()}}
\item \href{#method-RomMetricNetwork-clone}{\code{RomMetricNetwork$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomMetricNetwork-new"></a>}}
\if{latex}{\out{\hypertarget{method-RomMetricNetwork-new}{}}}
\subsection{Method \code{new()}}{
Initialize a RomMetricNetwork object by loading in either data from
\code{om_vahydro_metric_grid()} or by loading in a source column and
defining the approriate source and destination nodes. If datasource 'om'
is used, this method will apply \code{get_om_rseg_node()} to identify
potential source and destination node IDs and a bundle and ftype will be
joined in by featureid. The user will be warned if duplicate nodes are
encountered and it will be recommended to use
\code{force_unique_src_ids()} if the user wishes to get a unique network
only
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomMetricNetwork$new(
  ds = FALSE,
  datasource = self$datasource,
  metrics = NA,
  runids = NA,
  featureid = "all",
  bundle = "all",
  ftype = "all",
  model_version = "vahydro-1.0",
  dest_node_col = self$dest_node_col,
  src_node_col = self$src_node_col,
  entity_type = self$entity_type,
  entity_id_col = self$entity_id_col,
  force_unique_src = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ds}}{datasource object of class \code{RomDataSource} often created in DEQ
config files}

\item{\code{datasource}}{Either 'om' or a data frame input inidicating how this
object instance should derive the metric data. If 'om' is selected,
\code{om_vahydro_metric_grid()} will be used to get model metrics and
methods will assume a riverseg column is present that contains watersheds
IDs in the form JU3_6690_7280 where 6690 identifies this segment and 7280
is the segment (source node) to which it flows (destination node)}

\item{\code{metrics}}{character vector of requested metrics}

\item{\code{runids}}{Either a character vector of model run ids (in the form
'runid_600', for example) or a dataframe with model_version, runid, metric,
and runlabel specified. The data frame is generally deprecated in favor of
using the individual arguments on this function.}

\item{\code{featureid}}{integer or 'all', representing which specific features to
query for models}

\item{\code{bundle}}{character vector representing feature types or 'all'
(default). Only these features will be queried for models of
\code{model_version}. May be 'all' to query all bundle types. Defaults
to 'watershed'}

\item{\code{ftype}}{character vector representing feature types or 'all'
(default), i.e. the bundle subclass. Only these features will be
queried for models of \code{model_version}. May be 'all' to query all
ftypes.}

\item{\code{model_version}}{character vector with default 'vahydro-1.0'. This will
be used to determine which models to find on \code{entity_type}}

\item{\code{dest_node_col}}{If datasource is a dataframe, which field identifies the
destination node that this node flows to?}

\item{\code{src_node_col}}{If datasource is a dataframe, which field is the source
node that identifies this node?}

\item{\code{entity_type}}{character (single-length) to represent the table to query
for models, defaults to \code{dh_feature_fielded}}

\item{\code{entity_id_col}}{Character; defaults to 'hydroid' and represents the primary key
field of the entity type table (defaults dh_feature_fielded)}

\item{\code{force_unique_src}}{Logical, defaults to FALSE. If non-unique source ids
are found, should this object append these ids to force all nodes to be
unique? In the instance of om river segments, all data should have unique
IDs but model errors can, rarely, interfere with this and this setting may
be useful.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
object instance, with fields populated by user value and
  network_data and network_graph populated
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomMetricNetwork-get_om_rseg_node"></a>}}
\if{latex}{\out{\hypertarget{method-RomMetricNetwork-get_om_rseg_node}{}}}
\subsection{Method \code{get_om_rseg_node()}}{
Take in data from \code{om_vahydro_metric_grid()} or some equivalent data
frame with river segments, hydrocode, property names, and feature ids and
extract id columns to use in this object to define source and destination
nodes. This method uses bundle to detect non-river segments to assign
proper source ids (e.g. facility names used in place of river segment
source id on a facility node)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomMetricNetwork$get_om_rseg_node(
  df,
  om_rseg_col = "riverseg",
  om_hydrocode_col = "hydrocode",
  om_propname_col = "propname",
  om_featureid_col = "featureid",
  ds = self$ds
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{df}}{A data frame of potential network data with an OM style rseg
that can be parsed to source and destination nodes, with non-watersheds
characterized by propname}

\item{\code{om_rseg_col}}{Character, default to "riverseg". The field containing
riversegments akin to JL3_6680_6970}

\item{\code{om_hydrocode_col}}{Character, default to "hydrocode". The field that
contains the hydrocode of the entity table.}

\item{\code{om_propname_col}}{Chacater, default to propname. The name of the
node. Used to name source node IDs for non-watersheds.}

\item{\code{om_featureid_col}}{Character, default is "featureid". The field in
df that contains the primary key id relating entries to the entity
table}

\item{\code{ds}}{A \code{RomDataSource} object, defaulting to that stored on
this object via \code{initialize()}}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list with src_nodes defining the source nodes and dest_nodes
  defining the destination of each node
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomMetricNetwork-force_unique_src_ids"></a>}}
\if{latex}{\out{\hypertarget{method-RomMetricNetwork-force_unique_src_ids}{}}}
\subsection{Method \code{force_unique_src_ids()}}{
Ensures all nodes ids in src_node_col are unique by appending an
iterator on duplicates such that
\code{c("nodeA","nodeA","nodeB","nodeB","nodeA")} becomes
\code{c("nodeA","nodeA1","nodeB","nodeB1","nodeA2")}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomMetricNetwork$force_unique_src_ids()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Nothing, but sets network_data to have unique source nodes by
  appending an iterator on the ID to signify a duplicate
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomMetricNetwork-set_numeric_nodes"></a>}}
\if{latex}{\out{\hypertarget{method-RomMetricNetwork-set_numeric_nodes}{}}}
\subsection{Method \code{set_numeric_nodes()}}{
Converts source and destination node columns in network_data and
network_graph to be numeric integers such that each unique node is given
a unique integer value. Can optionally be run to force unique source IDs.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomMetricNetwork$set_numeric_nodes(force_unique_src_ids = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{force_unique_src_ids}}{Logical, default FALSE. Should
\code{force_unique_src_ids()} method be used to ensure unique IDs?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing, but sets network_data and network_graph to use the now
  integer source IDs.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomMetricNetwork-get_node_relation"></a>}}
\if{latex}{\out{\hypertarget{method-RomMetricNetwork-get_node_relation}{}}}
\subsection{Method \code{get_node_relation()}}{
Creates a list of all nodes upstream or downstream of target source node
and can optionally return aliases for node IDs
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomMetricNetwork$get_node_relation(
  direction = "upstream",
  value = NA,
  src_nodes = "all"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{direction}}{Character, default "upstream". Either "upstream" or
"downstream" and determines if node searching should look for all
upstream or downstream nodes from target source nodes.}

\item{\code{value}}{Character, default NA. A field in network_data containing
node aliases to return. For instance, may be 'riverseg' for OM data
from \code{om_vahydro_metric_grid()} to return upstream/downstream
riversegments (rather than just source segment id e.g. to return
JL3_6680_7280 instead of 6680)}

\item{\code{src_nodes}}{Character, defaults to "all". Either "all" or a
character vector of source node IDs. Related nodes will only be
returned for these target source nodes.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of the nodes or value in target direction from all or
  src_nodes
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomMetricNetwork-extact_node_network"></a>}}
\if{latex}{\out{\hypertarget{method-RomMetricNetwork-extact_node_network}{}}}
\subsection{Method \code{extact_node_network()}}{
A subset of network_data that contains all data from nodes in target
direction of source node (with optional destination node)
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomMetricNetwork$extact_node_network(
  src_node,
  dest_node = NA,
  direction = "upstream"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{src_node}}{Character. The starting node to look for related nodes
and data in direction. This may optionally require a dest_node.}

\item{\code{dest_node}}{Character. Default NA. The dest_node of src_node in case
of non-unique source IDs}

\item{\code{direction}}{Character. Default "upstream". Either "upstream" or
"downstream" and represent the direction along the network with which
to search for related data.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame of the extracted node network from src_node (with
  dest_node) in direction
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomMetricNetwork-calc_cumulative"></a>}}
\if{latex}{\out{\hypertarget{method-RomMetricNetwork-calc_cumulative}{}}}
\subsection{Method \code{calc_cumulative()}}{
Calculate the cumulative sum of a value defined by value_col from all
upstream nodes of each node (defined by \code{get_node_relation()})
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomMetricNetwork$calc_cumulative(value_col, out_col_name = "cumultive_sum")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value_col}}{Character. The value in network_data to sum across
upstream nodes}

\item{\code{out_col_name}}{Character. Default "cumulative_sum" and represents
where output network cumulative sum should be stored.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame of network_data now containing a field named by
  out_col_name for the cumulative sum of the values in value_col
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomMetricNetwork-get_spatial_network"></a>}}
\if{latex}{\out{\hypertarget{method-RomMetricNetwork-get_spatial_network}{}}}
\subsection{Method \code{get_spatial_network()}}{
Get a spatial data frame by joining in entity spatial data defined by
pkid_col with optional color and metadata columns calculated from
value_col to allow for rapid plotting.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomMetricNetwork$get_spatial_network(
  pkid_col = "featureid",
  value_col = NA,
  value_label = "value"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{pkid_col}}{Character. Default "feature_id" and represents
how to join entity table onto data to retrieve spatial data}

\item{\code{value_col}}{Character. The value in network_data to create color and
metadata columns for.}

\item{\code{value_label}}{Character. A metadata label defining value_col.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A SF data frame containing geometries and optional color and
  metadata columns defining values in value_col
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomMetricNetwork-render_force_diagram"></a>}}
\if{latex}{\out{\hypertarget{method-RomMetricNetwork-render_force_diagram}{}}}
\subsection{Method \code{render_force_diagram()}}{
Get an interactive force network diagram of network_data with IDs defined
by name_col, colors by group_col, node size by size_col, and a
default_size for nodes with no size_col value
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomMetricNetwork$render_force_diagram(
  name_col = NA,
  group_col = NA,
  size_col = NA,
  default_size = 5
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{name_col}}{Character. A field in network_data that can characterize
node names.}

\item{\code{group_col}}{Character. A field in network_data that can be used to
color the data by broad groups e.g. like bundle from OM data.}

\item{\code{size_col}}{Character. A numeric field in network_data from which the
size of each node may be calculated.}

\item{\code{default_size}}{Integer. Default 5. Number of pixels each node should
be sized to by default if no size_col value provided.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A \code{networkD3} interactive (JS) force diagram of the network
  data with size and colors broadly controlled by user.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomMetricNetwork-handle_config"></a>}}
\if{latex}{\out{\hypertarget{method-RomMetricNetwork-handle_config}{}}}
\subsection{Method \code{handle_config()}}{
Handles the config passed in initialize to set all fields on the object
by calling \code{self$handle_config_item()}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomMetricNetwork$handle_config(config = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{A named list of public fields with values to set}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomMetricNetwork-handle_config_item"></a>}}
\if{latex}{\out{\hypertarget{method-RomMetricNetwork-handle_config_item}{}}}
\subsection{Method \code{handle_config_item()}}{
For a given item in the user config, check to see if the name of the
config list item is a field on this object. If so, set it to the value of
the config_item
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomMetricNetwork$handle_config_item(config_item, config_item_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config_item}}{A value to set on the field in config_item_name, if it
exists on this object}

\item{\code{config_item_name}}{A field to set with the value config_item, if the
field exists on this object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The target field, may be NULL if it does not exist
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RomMetricNetwork-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RomMetricNetwork-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RomMetricNetwork$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
