% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaterGageBase.R
\name{WaterGageBase}
\alias{WaterGageBase}
\title{WaterGageBase}
\value{
R6 Object of class WaterGageBase
}
\description{
Utility class for compiling query info for stream gage analysis
}
\details{
Water Gage Data Information Object

This R6 object has standard methods for managing the data and meta
  data related to stream gage data. On initialize, this object sets fields
  relevant for stream flow data querying and analysis. This object can retrieve
  spatial data about a gage, but is not used to query streamflow itself as it
  is instead inherited by subclasses like WaterGageDaily that can query data
  instead
}
\examples{
\dontrun{
 #Spatial Info:
SA_Ashland <- WaterGageBase$new(config = list(gage_id = "01672500", ds_in = ds,
                                end_date = "2021-11-11",
                                approval_status = "Approved"))
 SA_Ashland$load_sf_da()
 SA_Ashland$drainage_area
 #If data source provided:
 SA_Ashland$load_wshd_feat()
}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ds}}{RomDataSource often provided in DEQ config files and features a
local data storage or a connection to a database/RESTful service}

\item{\code{data_source}}{May be either USGS or the path to gage data. If a file
path is used, both the flow_col and date_col fields should be populated}

\item{\code{gage_id}}{The USGS ID of the gage}

\item{\code{date_col}}{The column containing "date" R class date times for the
date_source}

\item{\code{flow_col}}{The column containing the flow values from the data_source}

\item{\code{drainage_area}}{The drainage area of the gage, populated on
initialize if possible}

\item{\code{gage_data_sf}}{A simple feature of the gage data}

\item{\code{gage_feature}}{a VA Hydro feature of the gage if available from ds}

\item{\code{start_date}}{The first date to include in the data}

\item{\code{end_date}}{The last date to include in the data}

\item{\code{parameter_code}}{Which USGS paramter to analyze? Defaults to 00060
for discharge}

\item{\code{approval_status}}{Either 'approved', 'provisional', or 'all'.
Determines what kind of data should be retrieved from USGS: all data or
only those marked approved or provisional?}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WaterGageBase-new}{\code{WaterGageBase$new()}}
\item \href{#method-WaterGageBase-handle_config}{\code{WaterGageBase$handle_config()}}
\item \href{#method-WaterGageBase-handle_config_item}{\code{WaterGageBase$handle_config_item()}}
\item \href{#method-WaterGageBase-load_sf_da}{\code{WaterGageBase$load_sf_da()}}
\item \href{#method-WaterGageBase-load_wshd_feat}{\code{WaterGageBase$load_wshd_feat()}}
\item \href{#method-WaterGageBase-clone}{\code{WaterGageBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaterGageBase-new"></a>}}
\if{latex}{\out{\hypertarget{method-WaterGageBase-new}{}}}
\subsection{Method \code{new()}}{
Initialize a WaterGageBase() instance populating all fields passed to
object by the named list config. Only valid public fields are populated.
The object will iterate over config to set fields.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterGageBase$new(config = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{A named list of public fields with values to set}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
object instance, with fields populated by values in config
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaterGageBase-handle_config"></a>}}
\if{latex}{\out{\hypertarget{method-WaterGageBase-handle_config}{}}}
\subsection{Method \code{handle_config()}}{
Handles the config passed in initialize to set all fields on the object
by calling \code{self$handle_config_item()}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterGageBase$handle_config(config = list())}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config}}{A named list of public fields with values to set}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
NULL
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaterGageBase-handle_config_item"></a>}}
\if{latex}{\out{\hypertarget{method-WaterGageBase-handle_config_item}{}}}
\subsection{Method \code{handle_config_item()}}{
For a given item in the user config, check to see if the name of the
config list item is a field on this object. If so, set it to the value of
the config_item
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterGageBase$handle_config_item(config_item, config_item_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config_item}}{A value to set on the field in config_item_name, if it
exists on this object}

\item{\code{config_item_name}}{A field to set with the value config_item, if the
field exists on this object}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The target field, may be NULL if it does not exist
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaterGageBase-load_sf_da"></a>}}
\if{latex}{\out{\hypertarget{method-WaterGageBase-load_sf_da}{}}}
\subsection{Method \code{load_sf_da()}}{
If a valid USGS gage_id is set on object, use
\code{dataRetrieval::read_waterdata_monitoring_location()} or
\code{dataRetrieval::readNWISsite()} to get an sf data frame with
additional details about the gage and use it to set the
\code{gage_data_sf} and \code{drainage_area} fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterGageBase$load_sf_da()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Nothing, but sets \code{gage_data_sf} and \code{drainage_area}
  fields on object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaterGageBase-load_wshd_feat"></a>}}
\if{latex}{\out{\hypertarget{method-WaterGageBase-load_wshd_feat}{}}}
\subsection{Method \code{load_wshd_feat()}}{
If a valid RomDataSource has been provided and the gage_id field has been
set, query the ds field for the corresponding USGS watershed feature
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterGageBase$load_wshd_feat()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The RomFeature if found or NA. Also sets \code{gage_feature} field on the object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaterGageBase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WaterGageBase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterGageBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
