% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaterGageBase.R
\name{WaterGageBase}
\alias{WaterGageBase}
\title{WaterGageBase}
\value{
R6 Object of class WaterGageBase
}
\description{
Utility class for interacting with water gage data from USGS or
  other source
}
\details{
Water Gage Data Object

This R6 object has standard methods for managing the data and meta
  data related to stream gage data. On initialize, the object can load in USGS
  gage data, a csv, or data provided by user and then offers methods to bring
  in USGS geometries, OWS watersheds, calculate VPDES critical low flows, and
  to provide basic drought forecasting capabilities.
}
\examples{
\dontrun{
NA_HartCorner <- WaterGageBase$new(data_source = "USGS", gage_id = "01671020",
                                    ds_in = NA, get_sf = FALSE)
NA_HartCorner$load_sf_da()
NA_HartCorner$comp_xQy()
NA_HartCorner$drainage_area
NA_HartCorner$low_flows
#If data source provided:
#NA_HartCorner$load_wshd_feat()
}
}
\seealso{
[xQy()]

[om_flow_table()]
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ds}}{RomDataSource often provided in DEQ config files and features a
local data storage or a connection to a database/RESTful service}

\item{\code{data_source}}{May be either USGS or the path to gage data. If a file
path is used, both the flow_col and date_col fields should be populated}

\item{\code{gage_id}}{The USGS ID of the gage}

\item{\code{date_col}}{The column containing "date" R class date times for the
date_source}

\item{\code{flow_col}}{The column containing the flow values from the data_source}

\item{\code{drainage_area}}{The drainage area of the gage, populated on
initialize if possible}

\item{\code{gage_data}}{The gage data from data_source}

\item{\code{gage_data_sf}}{A simple feature of the gage data}

\item{\code{gage_feature}}{a VA Hydro feature of the gage if available from ds}

\item{\code{statistic_id}}{The kind of stream data to pull}

\item{\code{parameter_code}}{Which USGS paramter to analyze? Defaults to 00060
for discharge}

\item{\code{usgs_properties}}{The fields to pull from USGS}

\item{\code{low_flows}}{Critical low flows calculated by hydrotools::xQy() via
\code{self$comp_xQy()}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-waterGageBase-new}{\code{WaterGageBase$new()}}
\item \href{#method-waterGageBase-get_gage_data}{\code{WaterGageBase$get_gage_data()}}
\item \href{#method-waterGageBase-get_gage_data_old}{\code{WaterGageBase$get_gage_data_old()}}
\item \href{#method-waterGageBase-load_sf_da}{\code{WaterGageBase$load_sf_da()}}
\item \href{#method-waterGageBase-load_wshd_feat}{\code{WaterGageBase$load_wshd_feat()}}
\item \href{#method-waterGageBase-comp_xQy}{\code{WaterGageBase$comp_xQy()}}
\item \href{#method-waterGageBase-nep_table}{\code{WaterGageBase$nep_table()}}
\item \href{#method-waterGageBase-baseflow_analysis}{\code{WaterGageBase$baseflow_analysis()}}
\item \href{#method-waterGageBase-baseflow_forecast}{\code{WaterGageBase$baseflow_forecast()}}
\item \href{#method-waterGageBase-join_precip_data}{\code{WaterGageBase$join_precip_data()}}
\item \href{#method-waterGageBase-clone}{\code{WaterGageBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-waterGageBase-new"></a>}}
\if{latex}{\out{\hypertarget{method-waterGageBase-new}{}}}
\subsection{Method \code{new()}}{
Initialize a WaterGageBase() instance, returning an R6 object that is
now populated with gage data in its fields either by querying USGS or
through user upload of data. This object offers various standard methods
to analyze USGS gage data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterGageBase$new(data_source, gage_id = NA, flow_col, date_col, ds_in = NA)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_source}}{"USGS", a file path, or a data frame of stream gage data that
has columns as named in \code{flow_col} and \code{date_col}. If using USGS,
\code{flow_col} and \code{date_col} are not used as data will be pulled
directly from USGS.}

\item{\code{gage_id}}{The gage ID. This should be a USGS site number if \code{data_source}
is "USGS" such as 01671020}

\item{\code{flow_col}}{The column identifying the data of interest in the gage data.
Only required if \code{data_source} is a file path or a data frame}

\item{\code{date_col}}{The column identifying the dates in the gage data. Only
required if \code{data_source} is a file path or a data frame}

\item{\code{ds_in}}{An optional RomDataSource to allow for querying of additional
information. May be provided by OWS config files.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
object instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-waterGageBase-get_gage_data"></a>}}
\if{latex}{\out{\hypertarget{method-waterGageBase-get_gage_data}{}}}
\subsection{Method \code{get_gage_data()}}{
Using the new \code{dataRetrieval::read_waterdata_daily()}, query USGS
for the \code{self$gage_id} based on the set \code{self$paramter_code}
and \code{self$statistic_id}. Fields returned can be controlled by
\code{usgs_properties} field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterGageBase$get_gage_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Nothing, but sets \code{flow_col}, \code{date_col}, and
  \code{gage_data} fields on object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-waterGageBase-get_gage_data_old"></a>}}
\if{latex}{\out{\hypertarget{method-waterGageBase-get_gage_data_old}{}}}
\subsection{Method \code{get_gage_data_old()}}{
Using the deprecated \code{dataRetrieval::readNWISdv()}, query USGS
for the \code{self$gage_id} based on the set \code{self$paramter_code}
and \code{self$statistic_id}. Fields are renamed via
\code{dataRetrieval::renameNWISColumns()}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterGageBase$get_gage_data_old()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Nothing, but sets \code{flow_col}, \code{date_col}, and
  \code{gage_data} fields on object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-waterGageBase-load_sf_da"></a>}}
\if{latex}{\out{\hypertarget{method-waterGageBase-load_sf_da}{}}}
\subsection{Method \code{load_sf_da()}}{
If a valid USGS gage_id is set on object, use
\code{dataRetrieval::read_waterdata_monitoring_location()} or
\code{dataRetrieval::readNWISsite()} to get an sf data frame with
additional details about the gage and use it to set the
\code{gage_data_sf} and \code{drainage_area} fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterGageBase$load_sf_da()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Nothing, but sets \code{gage_data_sf} and \code{drainage_area}
  fields on object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-waterGageBase-load_wshd_feat"></a>}}
\if{latex}{\out{\hypertarget{method-waterGageBase-load_wshd_feat}{}}}
\subsection{Method \code{load_wshd_feat()}}{
If a valid RomDataSource has been provided and the gage_id field has been
set, query the ds field for the corresponding USGS watershed feature
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterGageBase$load_wshd_feat()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The RomFeature if found or NA. Also sets \code{gage_feature} field on the object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-waterGageBase-comp_xQy"></a>}}
\if{latex}{\out{\hypertarget{method-waterGageBase-comp_xQy}{}}}
\subsection{Method \code{comp_xQy()}}{
Calculate critical low flows like the 1Q10, 7Q10, 30Q10,
30Q5, and harmonic mean using the VPDES standard
\code{hydrotools::xQy()}. Custom parameters may be entered to customize
the low flow calculation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterGageBase$comp_xQy(
  AYS = "04-01",
  AYE = "03-31",
  startYear = NULL,
  endYear = NULL,
  x = 7,
  y = 10
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AYS}}{The start of the analysis season in the format MM-DD. For
instance, to calculate low flows across the water year, AYS = '10-01'}

\item{\code{AYE}}{The end of the analysis season in the format MM-DD. For instance,
to calculate low flows across the water year, AYE = '09-31'}

\item{\code{startYear}}{Any calendar year prior to this year will not be included.
User can leave this as NULL (default) to include all data.}

\item{\code{endYear}}{Any calendar year after this year will not be included. User
can leave this as NULL (default) to include all data.}

\item{\code{x}}{User may optionally calculate a custom xQy flow. 1Q10, 7Q10, 30Q10,
90Q10, and 30Q5 are included in the results already. This input serves as
the averaging period.}

\item{\code{y}}{User may optionally calculate a custom xQy flow. 1Q10, 7Q10, 30Q10,
90Q10, and 30Q5 are included in the results already. This input serves as
the return period}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Sets the low_flow field on the object but also returns the list
  from \code{hydrotools::xQy()}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-waterGageBase-nep_table"></a>}}
\if{latex}{\out{\hypertarget{method-waterGageBase-nep_table}{}}}
\subsection{Method \code{nep_table()}}{
Calculate non-exceedance flows for standard OWS percentiles in each month
of the year using \code{hydrotools::om_flow_table()}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterGageBase$nep_table()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A non-exceedance from table from \code{hydrotools::om_flow_table()}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-waterGageBase-baseflow_analysis"></a>}}
\if{latex}{\out{\hypertarget{method-waterGageBase-baseflow_analysis}{}}}
\subsection{Method \code{baseflow_analysis()}}{
Derive relevant baseflow recession parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterGageBase$baseflow_analysis()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-waterGageBase-baseflow_forecast"></a>}}
\if{latex}{\out{\hypertarget{method-waterGageBase-baseflow_forecast}{}}}
\subsection{Method \code{baseflow_forecast()}}{
Using the regression derived from baseflow_analysis method,
  what may flows look like if no rain occurs for the next ndays?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterGageBase$baseflow_forecast()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-waterGageBase-join_precip_data"></a>}}
\if{latex}{\out{\hypertarget{method-waterGageBase-join_precip_data}{}}}
\subsection{Method \code{join_precip_data()}}{
Use the watershed feature to find precipitation from 1980
  onward using \code{RomFeature$get_raster_ts()}. Join this data to the
  stream data and return to the user. Inputs are further described in
  \code{RomFeature}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterGageBase$join_precip_data(
  metric = "mean",
  aggregateData = FALSE,
  varkey = "prism_mod_daily",
  touched = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metric}}{default mean, which aspect of the ST_SummaryStats() to return}

\item{\code{aggregateData}}{Should data be aggregated and if so, how? FALSE will
return 1 record for every date. Otheriwse, can use basic aggregate
functions like mean, min, max to summarize data}

\item{\code{varkey}}{What variable to retrieve? This should be a varkey that has
been used to create raster summaries in dh_timeseries_weather. Relevant
meteorology varkeys include 'prism_mod_daily', 'daymet_mod_daily',
'nldas2_obs_hourly', 'nldas2_precip_hourly_tiled_16x16',
'nldas2_precip_daily', 'amalgamate_simple_lm', 'amalgamate_storm_vol'}

\item{\code{touched}}{Defaults to FALSE. Should PostGIS ST_Clip() use the
touched argument to return all raster cells touched by the feature? Or,
if FALSE, should only pixels that have centroids within the feature be
considered?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The gage_data data frame now with a column of precip_mm joined
    on
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-waterGageBase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-waterGageBase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterGageBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
