% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaterGageBase.R
\name{WaterGageBase}
\alias{WaterGageBase}
\title{WaterGageBase}
\value{
R6 Object of class WaterGageBase
}
\description{
Utility class for interacting with water gage data from USGS or
  other source
}
\details{
Water Gage Data Object

This R6 object has standard methods for managing the data and meta
  data related to stream gage data. On initialize, the object can load in USGS
  gage data, a csv, or data provided by user and then offers methods to bring
  in USGS geometries, OWS watersheds, calculate VPDES critical low flows, and
  to provide basic drought forecasting capabilities.
}
\examples{
\dontrun{
 #Spatial Info:
 SA_Ashland$load_sf_da()
 SA_Ashland$drainage_area
 #Flow metrics:
 SA_Ashland$comp_xQy()
 SA_Ashland$low_flows
 SA_Ashland$nep_table()
 SA_Ashland$days_below(75)
 #Plots:
 SA_Ashland$plot_rechare_context(2026,c(12,1,2))
 SA_Ashland$plot_low_flows("boxplot")
 #If data source provided:
 SA_Ashland$load_wshd_feat()
 SA_Ashland$plot_precip_data(targetYear = 2022,include_months = c(10:12,1:4))
 #All plots:
 SA_Ashland$plot_all()
}
}
\seealso{
[xQy()]

[om_flow_table()]
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ds}}{RomDataSource often provided in DEQ config files and features a
local data storage or a connection to a database/RESTful service}

\item{\code{data_source}}{May be either USGS or the path to gage data. If a file
path is used, both the flow_col and date_col fields should be populated}

\item{\code{gage_id}}{The USGS ID of the gage}

\item{\code{date_col}}{The column containing "date" R class date times for the
date_source}

\item{\code{flow_col}}{The column containing the flow values from the data_source}

\item{\code{drainage_area}}{The drainage area of the gage, populated on
initialize if possible}

\item{\code{gage_data}}{The gage data from data_source}

\item{\code{gage_data_sf}}{A simple feature of the gage data}

\item{\code{gage_feature}}{a VA Hydro feature of the gage if available from ds}

\item{\code{statistic_id}}{The kind of stream data to pull}

\item{\code{parameter_code}}{Which USGS paramter to analyze? Defaults to 00060
for discharge}

\item{\code{usgs_properties}}{The fields to pull from USGS}

\item{\code{approval_status}}{Either 'approved', 'provisional', or 'all'.
Determines what kind of data should be retrieved from USGS: all data or
only those marked approved or provisional?}

\item{\code{start_date}}{The first date to include in the data}

\item{\code{end_date}}{The last date to include in the data}

\item{\code{low_flows}}{Critical low flows calculated by hydrotools::xQy() via
\code{self$comp_xQy()}}

\item{\code{precip_data}}{A daily precipitation data frame}

\item{\code{plots}}{A container to hold plots generated by methods on this object}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-waterGageBase-new}{\code{WaterGageBase$new()}}
\item \href{#method-waterGageBase-get_gage_data}{\code{WaterGageBase$get_gage_data()}}
\item \href{#method-waterGageBase-get_gage_data_old}{\code{WaterGageBase$get_gage_data_old()}}
\item \href{#method-waterGageBase-load_sf_da}{\code{WaterGageBase$load_sf_da()}}
\item \href{#method-waterGageBase-load_wshd_feat}{\code{WaterGageBase$load_wshd_feat()}}
\item \href{#method-waterGageBase-comp_xQy}{\code{WaterGageBase$comp_xQy()}}
\item \href{#method-waterGageBase-nep_table}{\code{WaterGageBase$nep_table()}}
\item \href{#method-waterGageBase-baseflow_analysis}{\code{WaterGageBase$baseflow_analysis()}}
\item \href{#method-waterGageBase-baseflow_forecast}{\code{WaterGageBase$baseflow_forecast()}}
\item \href{#method-waterGageBase-days_below}{\code{WaterGageBase$days_below()}}
\item \href{#method-waterGageBase-join_precip_data}{\code{WaterGageBase$join_precip_data()}}
\item \href{#method-waterGageBase-plot_rechare_context}{\code{WaterGageBase$plot_rechare_context()}}
\item \href{#method-waterGageBase-plot_precip_data}{\code{WaterGageBase$plot_precip_data()}}
\item \href{#method-waterGageBase-plot_low_flows}{\code{WaterGageBase$plot_low_flows()}}
\item \href{#method-waterGageBase-plot_all}{\code{WaterGageBase$plot_all()}}
\item \href{#method-waterGageBase-clone}{\code{WaterGageBase$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-waterGageBase-new"></a>}}
\if{latex}{\out{\hypertarget{method-waterGageBase-new}{}}}
\subsection{Method \code{new()}}{
Initialize a WaterGageBase() instance, returning an R6 object that is
now populated with gage data in its fields either by querying USGS or
through user upload of data. This object offers various standard methods
to analyze USGS gage data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterGageBase$new(
  data_source = "USGS",
  gage_id = NA,
  flow_col,
  date_col,
  start_date = self$start_date,
  end_date = self$end_date,
  approval_status = self$approval_status,
  ds_in = NA
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data_source}}{"USGS" (default), a file path, or a data frame of stream gage data that
has columns as named in \code{flow_col} and \code{date_col}. If using USGS,
\code{flow_col} and \code{date_col} are not used as data will be pulled
directly from USGS.}

\item{\code{gage_id}}{The gage ID. This should be a USGS site number if \code{data_source}
is "USGS" such as 01671020}

\item{\code{flow_col}}{The column identifying the data of interest in the gage data.
Only required if \code{data_source} is a file path or a data frame}

\item{\code{date_col}}{The column identifying the dates in the gage data. Only
required if \code{data_source} is a file path or a data frame}

\item{\code{start_date}}{The first date to include in data get}

\item{\code{end_date}}{The last date to include in data get}

\item{\code{approval_status}}{'all', 'approved', or 'provisional'. What kind of
data should be used from USGS? Defaults to that set in the field on
this object, which itself defaults to 'all'}

\item{\code{ds_in}}{An optional RomDataSource to allow for querying of additional
information. May be provided by OWS config files.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
object instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-waterGageBase-get_gage_data"></a>}}
\if{latex}{\out{\hypertarget{method-waterGageBase-get_gage_data}{}}}
\subsection{Method \code{get_gage_data()}}{
Using the new \code{dataRetrieval::read_waterdata_daily()}, query USGS
for the \code{self$gage_id} based on the set \code{self$paramter_code}
and \code{self$statistic_id}. Fields returned can be controlled by
\code{usgs_properties} field
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterGageBase$get_gage_data(start_date, end_date, approval_status)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start_date}}{The first date to include in data get}

\item{\code{end_date}}{The last date to include in data get}

\item{\code{approval_status}}{'all', 'approved', or 'provisional'. What kind of
data should be used from USGS? Defaults to that set in the field on
this object, which itself defaults to 'all'}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing, but sets \code{flow_col}, \code{date_col}, and
  \code{gage_data} fields on object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-waterGageBase-get_gage_data_old"></a>}}
\if{latex}{\out{\hypertarget{method-waterGageBase-get_gage_data_old}{}}}
\subsection{Method \code{get_gage_data_old()}}{
Using the deprecated \code{dataRetrieval::readNWISdv()}, query USGS
for the \code{self$gage_id} based on the set \code{self$paramter_code}
and \code{self$statistic_id}. Fields are renamed via
\code{dataRetrieval::renameNWISColumns()}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterGageBase$get_gage_data_old(start_date, end_date, approval_status)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{start_date}}{The first date to include in data get}

\item{\code{end_date}}{The last date to include in data get}

\item{\code{approval_status}}{'all', 'approved', or 'provisional'. What kind of
data should be used from USGS? Defaults to that set in the field on
this object, which itself defaults to 'all'}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Nothing, but sets \code{flow_col}, \code{date_col}, and
  \code{gage_data} fields on object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-waterGageBase-load_sf_da"></a>}}
\if{latex}{\out{\hypertarget{method-waterGageBase-load_sf_da}{}}}
\subsection{Method \code{load_sf_da()}}{
If a valid USGS gage_id is set on object, use
\code{dataRetrieval::read_waterdata_monitoring_location()} or
\code{dataRetrieval::readNWISsite()} to get an sf data frame with
additional details about the gage and use it to set the
\code{gage_data_sf} and \code{drainage_area} fields
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterGageBase$load_sf_da()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Nothing, but sets \code{gage_data_sf} and \code{drainage_area}
  fields on object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-waterGageBase-load_wshd_feat"></a>}}
\if{latex}{\out{\hypertarget{method-waterGageBase-load_wshd_feat}{}}}
\subsection{Method \code{load_wshd_feat()}}{
If a valid RomDataSource has been provided and the gage_id field has been
set, query the ds field for the corresponding USGS watershed feature
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterGageBase$load_wshd_feat()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
The RomFeature if found or NA. Also sets \code{gage_feature} field on the object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-waterGageBase-comp_xQy"></a>}}
\if{latex}{\out{\hypertarget{method-waterGageBase-comp_xQy}{}}}
\subsection{Method \code{comp_xQy()}}{
Calculate critical low flows like the 1Q10, 7Q10, 30Q10,
30Q5, and harmonic mean using the VPDES standard
\code{hydrotools::xQy()}. Custom parameters may be entered to customize
the low flow calculation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterGageBase$comp_xQy(
  AYS = "04-01",
  AYE = "03-31",
  startYear = NULL,
  endYear = NULL,
  x = 7,
  y = 10
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{AYS}}{The start of the analysis season in the format MM-DD. For
instance, to calculate low flows across the water year, AYS = '10-01'}

\item{\code{AYE}}{The end of the analysis season in the format MM-DD. For instance,
to calculate low flows across the water year, AYE = '09-31'}

\item{\code{startYear}}{Any calendar year prior to this year will not be included.
User can leave this as NULL (default) to include all data.}

\item{\code{endYear}}{Any calendar year after this year will not be included. User
can leave this as NULL (default) to include all data.}

\item{\code{x}}{User may optionally calculate a custom xQy flow. 1Q10, 7Q10, 30Q10,
90Q10, and 30Q5 are included in the results already. This input serves as
the averaging period.}

\item{\code{y}}{User may optionally calculate a custom xQy flow. 1Q10, 7Q10, 30Q10,
90Q10, and 30Q5 are included in the results already. This input serves as
the return period}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
Sets the low_flow field on the object but also returns the list
  from \code{hydrotools::xQy()}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-waterGageBase-nep_table"></a>}}
\if{latex}{\out{\hypertarget{method-waterGageBase-nep_table}{}}}
\subsection{Method \code{nep_table()}}{
Calculate non-exceedance flows for standard OWS percentiles in each month
of the year using \code{hydrotools::om_flow_table()}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterGageBase$nep_table()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A non-exceedance from table from \code{hydrotools::om_flow_table()}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-waterGageBase-baseflow_analysis"></a>}}
\if{latex}{\out{\hypertarget{method-waterGageBase-baseflow_analysis}{}}}
\subsection{Method \code{baseflow_analysis()}}{
Derive relevant baseflow recession parameters
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterGageBase$baseflow_analysis()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-waterGageBase-baseflow_forecast"></a>}}
\if{latex}{\out{\hypertarget{method-waterGageBase-baseflow_forecast}{}}}
\subsection{Method \code{baseflow_forecast()}}{
Using the regression derived from baseflow_analysis method,
  what may flows look like if no rain occurs for the next ndays?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterGageBase$baseflow_forecast()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-waterGageBase-days_below"></a>}}
\if{latex}{\out{\hypertarget{method-waterGageBase-days_below}{}}}
\subsection{Method \code{days_below()}}{
How many days, on average, each month are below a threshold
 value?
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterGageBase$days_below(threshold = 100)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{threshold}}{The value to assess}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A data frame with the number of days, on average, where the gage
 value fell below the user input threshold
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-waterGageBase-join_precip_data"></a>}}
\if{latex}{\out{\hypertarget{method-waterGageBase-join_precip_data}{}}}
\subsection{Method \code{join_precip_data()}}{
Use the watershed feature to find precipitation from 1980
  onward using \code{RomFeature$get_raster_ts()}. Join this data to the
  stream data and return to the user. Inputs are further described in
  \code{RomFeature}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterGageBase$join_precip_data(
  metric = "mean",
  aggregateData = FALSE,
  varkey = "prism_mod_daily",
  touched = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{metric}}{default mean, which aspect of the ST_SummaryStats() to return}

\item{\code{aggregateData}}{Should data be aggregated and if so, how? FALSE will
return 1 record for every date. Otheriwse, can use basic aggregate
functions like mean, min, max to summarize data}

\item{\code{varkey}}{What variable to retrieve? This should be a varkey that has
been used to create raster summaries in dh_timeseries_weather. Relevant
meteorology varkeys include 'prism_mod_daily', 'daymet_mod_daily',
'nldas2_obs_hourly', 'nldas2_precip_hourly_tiled_16x16',
'nldas2_precip_daily', 'amalgamate_simple_lm', 'amalgamate_storm_vol'}

\item{\code{touched}}{Defaults to FALSE. Should PostGIS ST_Clip() use the
touched argument to return all raster cells touched by the feature? Or,
if FALSE, should only pixels that have centroids within the feature be
considered?}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
The gage_data data frame now with a column of precip_mm joined
    on. Also set in the precip_data field on this object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-waterGageBase-plot_rechare_context"></a>}}
\if{latex}{\out{\hypertarget{method-waterGageBase-plot_rechare_context}{}}}
\subsection{Method \code{plot_rechare_context()}}{
Create a boxplot of the monthly metric requested by user at
  this gage (or of a user input data frame) and then plot the target year
  as highlighted points on the plot. Often used in OWS to assess
  potential winter baseflow recharge when using gage data and evaluting
  between October - March (include_months = c(10,11,12,1,2,3)).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterGageBase$plot_rechare_context(
  targetYear = as.numeric(format(Sys.Date(), "\%Y")),
  include_months = 1:12,
  AYS = 10,
  metric = mean,
  use_y_log = TRUE,
  ylab = "Monthly Mean Flow (cfs)"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{targetYear}}{What year should be highlighted on the plot? Defaults
to this year. If left as NULL, no additional point data will be included}

\item{\code{include_months}}{What months should be included in plot/analysis?}

\item{\code{AYS}}{Which month begins the analysis year? The plot will start in
this month and then proceed through the next 12 months or those
included by user in include_months}

\item{\code{metric}}{A function to pass to \code{dplyr::summarize()} to control
how data monthly data is aggregated}

\item{\code{use_y_log}}{Boolean. Should the y axis be on a log10 scale? Defaults
to TRUE}

\item{\code{ylab}}{Character. The text to display on the y-axis of the plot.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A boxplot of monthly mean flows with highlighted points
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-waterGageBase-plot_precip_data"></a>}}
\if{latex}{\out{\hypertarget{method-waterGageBase-plot_precip_data}{}}}
\subsection{Method \code{plot_precip_data()}}{
Create a boxplot of monthly sums of precipitation using the
 data provided in either the precip_data field or by join_precip_data(). 
 The boxplot is created by plot_recharge_context(). This will ONLY use
 days where the data in the input precip column is not NA
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterGageBase$plot_precip_data(
  targetYear = as.numeric(format(Sys.Date(), "\%Y")),
  include_months = 1:12,
  AYS = 10,
  metric = sum,
  precip_col = "precip_mm",
  use_y_log = FALSE,
  ylab = "Monthly Precip (mm)"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{targetYear}}{What year should be highlighted on the plot? Defaults
to this year. If left as NULL, no additional point data will be included}

\item{\code{include_months}}{What months should be included in plot/analysis?}

\item{\code{AYS}}{Which month begins the analysis year? The plot will start in
this month and then proceed through the next 12 months or those
included by user in include_months}

\item{\code{metric}}{A function to pass to \code{dplyr::summarize()} to control
how data monthly data is aggregated}

\item{\code{precip_col}}{The column to plot in the precip data frame}

\item{\code{use_y_log}}{Boolean. Should the y axis be on a log10 scale? Defaults
to FALSE}

\item{\code{ylab}}{Character. The text to display on the y-axis of the plot.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A boxplot of precipitation data, also set in the plots field
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-waterGageBase-plot_low_flows"></a>}}
\if{latex}{\out{\hypertarget{method-waterGageBase-plot_low_flows}{}}}
\subsection{Method \code{plot_low_flows()}}{
Create a line plot of the lowest x-day average flow recorded
  each analysis year as set in \code{comp_xQy()}. For instance, the plot
  will include the lowest 7-day average flow recorded each analysis year.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterGageBase$plot_low_flows(plot_type = "lineplot")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{plot_type}}{Either "lineplot" or "boxplot". plot_type will control
which kind of plot is returned to user and set in the plots field,
either a "lineplot" of the low flow time series or a "boxplot" per
average low flow metric}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A line plot of the x-day average low flow recorded each analysis
  year. Will set in the plots list as well.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-waterGageBase-plot_all"></a>}}
\if{latex}{\out{\hypertarget{method-waterGageBase-plot_all}{}}}
\subsection{Method \code{plot_all()}}{
Run all plotting methods and populate the plots field accordingly
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterGageBase$plot_all(
  recharge_targetYear = as.numeric(format(Sys.Date(), "\%Y")),
  recharge_include_months = 1:12,
  recharge_AYS = 10
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{recharge_targetYear}}{What year should be highlighted on the recharge plot?
See \code{self$plot_recharge_context()} for more information.}

\item{\code{recharge_include_months}}{What months (numeric) should be included in plot/analysis?}

\item{\code{recharge_AYS}}{Which month begins the recharge analysis year? 
See \code{self$plot_recharge_context()} for more information.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A list of all plots created by methods in this objects. Will also
  populate the plots field.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-waterGageBase-clone"></a>}}
\if{latex}{\out{\hypertarget{method-waterGageBase-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WaterGageBase$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
