% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WaterGageDaily.R
\name{WaterGageDaily}
\alias{WaterGageDaily}
\title{WaterGageDaily}
\value{
R6 Object of class WaterGageDaily
}
\description{
Utility class for interacting with water gage data from USGS or
other source
}
\details{
Water Gage Data Object

This R6 object has standard methods for managing the data and meta
data related to stream gage data. On initialize, the object can load in USGS
gage data, a csv, or data provided by user and then offers methods to bring
in USGS geometries, OWS watersheds, calculate VPDES critical low flows, and
to provide basic drought forecasting capabilities.

Using the regressions from the 2025 - 2027 HARP projects, allow
users to forecast baseflow from a given start_date. Forecasts can have
static or variable decay coefficients (AGWRC) and these may be
calculated from the regression or input directly as vectors by users.
Observed flow is joined onto the output data frame for comparison if a
"historic" forecast was run.

Calculates the IHA parameter group 1: Magnitude of montly water
conditions see \code{hydrotools::group1()}

Calculates the IHA parameter group 2: Magnitude of montly water
conditions see \code{hydrotools::group2()}

Use \code{self$baseflow_forecast()} iteratively over a list of
provided AGWRC options to create a plot of current potential forecasts.
Can return either a plot or a list with the forecast results in a data
frame and a plot as a separate object in the list

Loads data from the 2025-2027 HARP project to draw a scatter plot
of AGWRC vs Flow and plots the workflow regression
}
\examples{
\dontrun{
 #Spatial Info:
SA_Ashland <- WaterGageDaily$new(gage_id = "01672500", ds_in = ds,
                                end_date = "2021-11-11",
                                approval_status = "Approved")
 SA_Ashland$load_sf_da()
 SA_Ashland$drainage_area
 #Flow metrics:
 SA_Ashland$comp_xQy()
 SA_Ashland$low_flows
 SA_Ashland$nep_table()
 SA_Ashland$days_below(75)
 #zoos and group stats
 g1 <- SA_Ashland$group1("water", FUN = mean)
 g2 <- SA_Ashland$group2(yearType = "water", mimic.tnc = TRUE)
 myzoo <- SA_Ashland$as.zoo()
 #Plots:
 SA_Ashland$plot_rechare_context(2026,c(12,1,2))
 SA_Ashland$plot_low_flows("boxplot")
 #If data source provided:
 SA_Ashland$load_wshd_feat()
 SA_Ashland$plot_precip_data(targetYear = 2022,include_months = c(10:12,1:4))
 #All plots:
 SA_Ashland$plot_all()
}
}
\seealso{
\code{\link[=xQy]{xQy()}}

\code{\link[=om_flow_table]{om_flow_table()}}
}
\section{Super class}{
\code{\link[hydrotools:WaterGageBase]{WaterGageBase}} -> \code{WaterGageDaily}
}
\section{Public fields}{
  \if{html}{\out{<div class="r6-fields">}}
  \describe{
    \item{\code{gage_data}}{The gage data from data_source}

    \item{\code{statistic_id}}{The kind of stream data to pull}

    \item{\code{usgs_properties}}{The fields to pull from USGS}

    \item{\code{low_flows}}{Critical low flows calculated by hydrotools::xQy() via
\code{self$comp_xQy()}}

    \item{\code{precip_data}}{A daily precipitation data frame}

    \item{\code{stats_obj}}{A \code{WaterGageStats()} that contains relevant
monthly, yearly, and day-of-year statistics on the gage flow}

    \item{\code{plots}}{A container to hold plots generated by methods on this object}
  }
  \if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
  \itemize{
    \item \href{#method-WaterGageDaily-initialize}{\code{WaterGageDaily$new()}}
    \item \href{#method-WaterGageDaily-get_gage_data}{\code{WaterGageDaily$get_gage_data()}}
    \item \href{#method-WaterGageDaily-get_gage_data_old}{\code{WaterGageDaily$get_gage_data_old()}}
    \item \href{#method-WaterGageDaily-filter_data_by_date}{\code{WaterGageDaily$filter_data_by_date()}}
    \item \href{#method-WaterGageDaily-comp_xQy}{\code{WaterGageDaily$comp_xQy()}}
    \item \href{#method-WaterGageDaily-nep_table}{\code{WaterGageDaily$nep_table()}}
    \item \href{#method-WaterGageDaily-baseflow_forecast}{\code{WaterGageDaily$baseflow_forecast()}}
    \item \href{#method-WaterGageDaily-days_below}{\code{WaterGageDaily$days_below()}}
    \item \href{#method-WaterGageDaily-group1}{\code{WaterGageDaily$group1()}}
    \item \href{#method-WaterGageDaily-group2}{\code{WaterGageDaily$group2()}}
    \item \href{#method-WaterGageDaily-as.zoo}{\code{WaterGageDaily$as.zoo()}}
    \item \href{#method-WaterGageDaily-join_precip_data}{\code{WaterGageDaily$join_precip_data()}}
    \item \href{#method-WaterGageDaily-plot_baseflow_forecast}{\code{WaterGageDaily$plot_baseflow_forecast()}}
    \item \href{#method-WaterGageDaily-plot_baseflow_agwrc}{\code{WaterGageDaily$plot_baseflow_agwrc()}}
    \item \href{#method-WaterGageDaily-plot_rechare_context}{\code{WaterGageDaily$plot_rechare_context()}}
    \item \href{#method-WaterGageDaily-plot_precip_data}{\code{WaterGageDaily$plot_precip_data()}}
    \item \href{#method-WaterGageDaily-plot_low_flows}{\code{WaterGageDaily$plot_low_flows()}}
    \item \href{#method-WaterGageDaily-get_doy_percentiles}{\code{WaterGageDaily$get_doy_percentiles()}}
    \item \href{#method-WaterGageDaily-plot_percentiles}{\code{WaterGageDaily$plot_percentiles()}}
    \item \href{#method-WaterGageDaily-plot_all}{\code{WaterGageDaily$plot_all()}}
    \item \href{#method-WaterGageDaily-clone}{\code{WaterGageDaily$clone()}}
  }
}
\if{html}{\out{<details><summary>Inherited methods</summary>
<ul>
  <li><span class="pkg-link" data-pkg="hydrotools" data-topic="WaterGageBase" data-id="agwrc_fun"><a href='../../hydrotools/html/WaterGageBase.html#method-WaterGageBase-agwrc_fun'><code>WaterGageBase$agwrc_fun()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hydrotools" data-topic="WaterGageBase" data-id="baseflow_workflow_data"><a href='../../hydrotools/html/WaterGageBase.html#method-WaterGageBase-baseflow_workflow_data'><code>WaterGageBase$baseflow_workflow_data()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hydrotools" data-topic="WaterGageBase" data-id="get_model_or_scenario_props"><a href='../../hydrotools/html/WaterGageBase.html#method-WaterGageBase-get_model_or_scenario_props'><code>WaterGageBase$get_model_or_scenario_props()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hydrotools" data-topic="WaterGageBase" data-id="handle_config"><a href='../../hydrotools/html/WaterGageBase.html#method-WaterGageBase-handle_config'><code>WaterGageBase$handle_config()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hydrotools" data-topic="WaterGageBase" data-id="handle_config_item"><a href='../../hydrotools/html/WaterGageBase.html#method-WaterGageBase-handle_config_item'><code>WaterGageBase$handle_config_item()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hydrotools" data-topic="WaterGageBase" data-id="load_sf_da"><a href='../../hydrotools/html/WaterGageBase.html#method-WaterGageBase-load_sf_da'><code>WaterGageBase$load_sf_da()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hydrotools" data-topic="WaterGageBase" data-id="load_wshd_feat"><a href='../../hydrotools/html/WaterGageBase.html#method-WaterGageBase-load_wshd_feat'><code>WaterGageBase$load_wshd_feat()</code></a></span></li>
  <li><span class="pkg-link" data-pkg="hydrotools" data-topic="WaterGageBase" data-id="read_om_file"><a href='../../hydrotools/html/WaterGageBase.html#method-WaterGageBase-read_om_file'><code>WaterGageBase$read_om_file()</code></a></span></li>
</ul>
</details>}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaterGageDaily-initialize"></a>}}
\if{latex}{\out{\hypertarget{method-WaterGageDaily-initialize}{}}}
\subsection{\code{WaterGageDaily$new()}}{
  Initialize a WaterGageDaily() instance, returning an R6 object that is
now populated with gage data in its fields either by querying USGS or
through user upload of data. This object offers various standard methods
to analyze USGS gage data
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{WaterGageDaily$new(
  data_source = "USGS",
  gage_id = NA,
  flow_col,
  date_col,
  start_date = self$start_date,
  end_date = self$end_date,
  approval_status = self$approval_status,
  dataRetrieval_version = "auto",
  get_station_attr = TRUE,
  ds_in = NA
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{data_source}}{"USGS" (default), a file path, or a data frame of stream gage data that
has columns as named in \code{flow_col} and \code{date_col}. If using USGS,
\code{flow_col} and \code{date_col} are not used as data will be pulled
directly from USGS.}
      \item{\code{gage_id}}{The gage ID. This should be a USGS site number if \code{data_source}
is "USGS" such as 01671020}
      \item{\code{flow_col}}{The column identifying the data of interest in the gage data.
Only required if \code{data_source} is a file path or a data frame}
      \item{\code{date_col}}{The column identifying the dates in the gage data. Only
required if \code{data_source} is a file path or a data frame}
      \item{\code{start_date}}{The first date to include in data get}
      \item{\code{end_date}}{The last date to include in data get}
      \item{\code{approval_status}}{'all', 'approved', or 'provisional'. What kind of
data should be used from USGS? Defaults to that set in the field on
this object, which itself defaults to 'all'}
      \item{\code{dataRetrieval_version}}{Character. If data_source is "USGS", should
the legacy or modern API be used? If set to "auto", this object will
use the most appropriate API based on the installed version of
dataRetrieval. Otherwise, the user may use "legacy" or "modern" to
control which API to retrieve data from.}
      \item{\code{get_station_attr}}{Logical, default to TRUE. If TRUE,
\code{load_sf_da()} will be run to provide the station sf data frame,
drainage area, and station name. If FALSE, the method will not be run
automatically.}
      \item{\code{ds_in}}{An optional RomDataSource to allow for querying of additional
information. May be provided by OWS config files.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    object instance
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaterGageDaily-get_gage_data"></a>}}
\if{latex}{\out{\hypertarget{method-WaterGageDaily-get_gage_data}{}}}
\subsection{\code{WaterGageDaily$get_gage_data()}}{
  Using the new \code{dataRetrieval::read_waterdata_daily()}, query USGS
for the \code{self$gage_id} based on the set \code{self$paramter_code}
and \code{self$statistic_id}. Fields returned can be controlled by
\code{usgs_properties} field
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{WaterGageDaily$get_gage_data(start_date, end_date, approval_status)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{start_date}}{The first date to include in data get}
      \item{\code{end_date}}{The last date to include in data get}
      \item{\code{approval_status}}{'all', 'approved', or 'provisional'. What kind of
data should be used from USGS? Defaults to that set in the field on
this object, which itself defaults to 'all'}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Nothing, but sets \code{flow_col}, \code{date_col}, and
\code{gage_data} fields on object
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaterGageDaily-get_gage_data_old"></a>}}
\if{latex}{\out{\hypertarget{method-WaterGageDaily-get_gage_data_old}{}}}
\subsection{\code{WaterGageDaily$get_gage_data_old()}}{
  Using the deprecated \code{dataRetrieval::readNWISdv()}, query USGS
for the \code{self$gage_id} based on the set \code{self$paramter_code}
and \code{self$statistic_id}. Fields are renamed via
\code{dataRetrieval::renameNWISColumns()}
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{WaterGageDaily$get_gage_data_old(start_date, end_date, approval_status)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{start_date}}{The first date to include in data get}
      \item{\code{end_date}}{The last date to include in data get}
      \item{\code{approval_status}}{'all', 'approved', or 'provisional'. What kind of
data should be used from USGS? Defaults to that set in the field on
this object, which itself defaults to 'all'}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Nothing, but sets \code{flow_col}, \code{date_col}, and
\code{gage_data} fields on object
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaterGageDaily-filter_data_by_date"></a>}}
\if{latex}{\out{\hypertarget{method-WaterGageDaily-filter_data_by_date}{}}}
\subsection{\code{WaterGageDaily$filter_data_by_date()}}{
  Filter gage_data field by date and provide a new WaterGage object to user
filtered to these limits
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{WaterGageDaily$filter_data_by_date(start_date = "", end_date = "")}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{start_date}}{Filter the gage data to remove any data prior to this date}
      \item{\code{end_date}}{Filter the gage data to remove any data after to this date}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A WaterGageDaily object with data filtered to the start and end
date provided by the user
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaterGageDaily-comp_xQy"></a>}}
\if{latex}{\out{\hypertarget{method-WaterGageDaily-comp_xQy}{}}}
\subsection{\code{WaterGageDaily$comp_xQy()}}{
  Calculate critical low flows like the 1Q10, 7Q10, 30Q10,
30Q5, and harmonic mean using the VPDES standard
\code{hydrotools::xQy()}. Custom parameters may be entered to customize
the low flow calculation
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{WaterGageDaily$comp_xQy(
  AYS = "04-01",
  AYE = "03-31",
  startYear = NULL,
  endYear = NULL,
  x = 7,
  y = 10
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{AYS}}{The start of the analysis season in the format MM-DD. For
instance, to calculate low flows across the water year, AYS = '10-01'}
      \item{\code{AYE}}{The end of the analysis season in the format MM-DD. For instance,
to calculate low flows across the water year, AYE = '09-31'}
      \item{\code{startYear}}{Any calendar year prior to this year will not be included.
User can leave this as NULL (default) to include all data.}
      \item{\code{endYear}}{Any calendar year after this year will not be included. User
can leave this as NULL (default) to include all data.}
      \item{\code{x}}{User may optionally calculate a custom xQy flow. 1Q10, 7Q10, 30Q10,
90Q10, and 30Q5 are included in the results already. This input serves as
the averaging period.}
      \item{\code{y}}{User may optionally calculate a custom xQy flow. 1Q10, 7Q10, 30Q10,
90Q10, and 30Q5 are included in the results already. This input serves as
the return period}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    Sets the low_flow field on the object but also returns the list
from \code{hydrotools::xQy()}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaterGageDaily-nep_table"></a>}}
\if{latex}{\out{\hypertarget{method-WaterGageDaily-nep_table}{}}}
\subsection{\code{WaterGageDaily$nep_table()}}{
  Calculate non-exceedance flows for standard OWS percentiles in each month
of the year using \code{hydrotools::om_flow_table()}
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{WaterGageDaily$nep_table()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A non-exceedance from table from \code{hydrotools::om_flow_table()}
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaterGageDaily-baseflow_forecast"></a>}}
\if{latex}{\out{\hypertarget{method-WaterGageDaily-baseflow_forecast}{}}}
\subsection{\code{WaterGageDaily$baseflow_forecast()}}{
  Flow forecast
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{WaterGageDaily$baseflow_forecast(
  start_date,
  forecast_days = 0:90,
  AGWRC = "lm_constant",
  use_limits = TRUE
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{start_date}}{Character. The date to begin the forecast, i.e. the flow
on this day will be used to run forecast calculations}
      \item{\code{forecast_days}}{numeric vector, defaults to 0:90. Days where forcast
should be calculated and will use corresponding AGWRC if provided
vector. Will use a variable calculated AGWRC if AGWRC = "lm_variable"}
      \item{\code{AGWRC}}{numeric or character. The decay coefficient to regress the
flow on start_date for forecast. May be a single numeric to allow for a
constant forecast or a vector of numeric values to allow for a variable
forecast but must be of length days. Otherwise, may be "lm_constant" to
calculate a constant value from m and b or "lm_variable" to have a
variable value}
      \item{\code{use_limits}}{logical. Should the regression be limited using data in
the agwrc_lm_limit field?}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A data.frame of date, AGWRC, forecasted flow, and observed flow
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaterGageDaily-days_below"></a>}}
\if{latex}{\out{\hypertarget{method-WaterGageDaily-days_below}{}}}
\subsection{\code{WaterGageDaily$days_below()}}{
  How many days, on average, each month are below a threshold
value?
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{WaterGageDaily$days_below(threshold = 100)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{threshold}}{The value to assess}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A data frame with the number of days, on average, where the gage
value fell below the user input threshold
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaterGageDaily-group1"></a>}}
\if{latex}{\out{\hypertarget{method-WaterGageDaily-group1}{}}}
\subsection{\code{WaterGageDaily$group1()}}{
  group1 stats wrapper
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{WaterGageDaily$group1(yearType = "water", FUN = stats::median)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{yearType}}{The type of year factor to be used when determining
statistcs, yr = 'water' or yr ='calendar' for water years and calculated
years respectively}
      \item{\code{FUN}}{the function to be applied to the monthly values. Median is the
default here. This can be a character string of a function name but
should ideally be a closure}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A matrix with the monthly results of FUN over each year
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaterGageDaily-group2"></a>}}
\if{latex}{\out{\hypertarget{method-WaterGageDaily-group2}{}}}
\subsection{\code{WaterGageDaily$group2()}}{
  group2 stats wrapper
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{WaterGageDaily$group2(yearType = "water", mimic.tnc = TRUE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{yearType}}{The type of year factor to be used when determining
statistcs, yr = 'water' or yr ='calendar' for water years and calculated
years respectively}
      \item{\code{mimic.tnc}}{should the function perform the calculation like the TNC
IHA software? If mimic.tnc is TRUE, then running means will be
calculated for each year individually and will not use data from the
previous or next year, thereby}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    a data frame with the group 2 statistics for each year
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaterGageDaily-as.zoo"></a>}}
\if{latex}{\out{\hypertarget{method-WaterGageDaily-as.zoo}{}}}
\subsection{\code{WaterGageDaily$as.zoo()}}{
  convert the flow_col to a zoo based on date_col with
\code{zoo::as.zoo()}
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{WaterGageDaily$as.zoo()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A zoo where the timeseries values are those defined in gage_data
by flow_col and the indices are in date_col
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaterGageDaily-join_precip_data"></a>}}
\if{latex}{\out{\hypertarget{method-WaterGageDaily-join_precip_data}{}}}
\subsection{\code{WaterGageDaily$join_precip_data()}}{
  Use the watershed feature to find precipitation from 1980
onward using \code{RomFeature$get_raster_ts()}. Join this data to the
stream data and return to the user. Inputs are further described in
\code{RomFeature}.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{WaterGageDaily$join_precip_data(
  metric = "mean",
  aggregateData = FALSE,
  varkey = "prism_mod_daily",
  touched = FALSE
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{metric}}{default mean, which aspect of the ST_SummaryStats() to return}
      \item{\code{aggregateData}}{Should data be aggregated and if so, how? FALSE will
return 1 record for every date. Otheriwse, can use basic aggregate
functions like mean, min, max to summarize data}
      \item{\code{varkey}}{What variable to retrieve? This should be a varkey that has
been used to create raster summaries in dh_timeseries_weather. Relevant
meteorology varkeys include 'prism_mod_daily', 'daymet_mod_daily',
'nldas2_obs_hourly', 'nldas2_precip_hourly_tiled_16x16',
'nldas2_precip_daily', 'amalgamate_simple_lm', 'amalgamate_storm_vol'}
      \item{\code{touched}}{Defaults to FALSE. Should PostGIS ST_Clip() use the
touched argument to return all raster cells touched by the feature? Or,
if FALSE, should only pixels that have centroids within the feature be
considered?}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    The gage_data data frame now with a column of precip_mm joined
on. Also set in the precip_data field on this object.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaterGageDaily-plot_baseflow_forecast"></a>}}
\if{latex}{\out{\hypertarget{method-WaterGageDaily-plot_baseflow_forecast}{}}}
\subsection{\code{WaterGageDaily$plot_baseflow_forecast()}}{
  plot current flow forecasts
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{WaterGageDaily$plot_baseflow_forecast(
  start_date,
  forecast_days = 0:90,
  AGWRC = list(lm_constant = "lm_constant", lm_variable = "lm_variable"),
  return_plotly = FALSE,
  include_days_before = 30,
  use_limits = TRUE,
  return_data = FALSE
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{start_date}}{Character. The date to begin the forecast, i.e. the flow
on this day will be used to run forecast calculations}
      \item{\code{forecast_days}}{numeric vector, defaults to 0:90. Days where forcast
should be calculated and will use corresponding AGWRC if provided
vector. Will use a variable calculated AGWRC if AGWRC = "lm_variable"}
      \item{\code{AGWRC}}{A named list of numeric or character. Each named entry is the
decay coefficient to regress the flow on start_date for forecast. May be
a single numeric to allow for a constant forecast or a vector of numeric
values to allow for a variable forecast but must be of length days.
Otherwise, may be "lm_constant" to calculate a constant value from m and
b or "lm_variable" to have a variable value. The names are used for the
plot legend}
      \item{\code{return_plotly}}{logical, default to FALSE. If TRUE, a plotly object
is returned. Otherwise, a ggplot is returned}
      \item{\code{include_days_before}}{numeric, defaults to 30. How many days prior to
start_date should be included on the plot?}
      \item{\code{use_limits}}{logical. Should the regression be limited using data in
the agwrc_lm_limit field?}
      \item{\code{return_data}}{logical, default to FALSE. Should results be returned
in a list, with the plot in one object and the raw data in another?}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    either a ggplot or plotly object of forecasts or a list with a
ggplot or plotly object and a dataframe of forecast results, depending
on user input
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaterGageDaily-plot_baseflow_agwrc"></a>}}
\if{latex}{\out{\hypertarget{method-WaterGageDaily-plot_baseflow_agwrc}{}}}
\subsection{\code{WaterGageDaily$plot_baseflow_agwrc()}}{
  Scatterplot of AGWRC vs Median Flow
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{WaterGageDaily$plot_baseflow_agwrc(return_plotly = FALSE, CI = FALSE, omsite)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{return_plotly}}{logical, default to FALSE. If TRUE, a plotly object
is returned. Otherwise, a ggplot is returned}
      \item{\code{CI}}{logical. Should a confidence interval be calculated to add to
plot?}
      \item{\code{omsite}}{The URL of the base VT Apache webserver, often defined in
/var/www/R/config.R}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A ggplot scatterplot of median flow vs AGWRC
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaterGageDaily-plot_rechare_context"></a>}}
\if{latex}{\out{\hypertarget{method-WaterGageDaily-plot_rechare_context}{}}}
\subsection{\code{WaterGageDaily$plot_rechare_context()}}{
  Create a boxplot of the monthly metric requested by user at
this gage (or of a user input data frame) and then plot the target year
as highlighted points on the plot. Often used in OWS to assess
potential winter baseflow recharge when using gage data and evaluting
between October - March (include_months = c(10,11,12,1,2,3)).
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{WaterGageDaily$plot_rechare_context(
  targetYear = as.numeric(format(Sys.Date(), "\%Y")),
  include_months = 1:12,
  AYS = 10,
  metric = mean,
  use_y_log = TRUE,
  ylab = "Monthly Mean Flow (cfs)"
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{targetYear}}{What year should be highlighted on the plot? Defaults
to this year. If left as NULL, no additional point data will be included}
      \item{\code{include_months}}{What months should be included in plot/analysis?}
      \item{\code{AYS}}{Which month begins the analysis year? The plot will start in
this month and then proceed through the next 12 months or those
included by user in include_months}
      \item{\code{metric}}{A function to pass to \code{dplyr::summarize()} to control
how data monthly data is aggregated}
      \item{\code{use_y_log}}{Boolean. Should the y axis be on a log10 scale? Defaults
to TRUE}
      \item{\code{ylab}}{Character. The text to display on the y-axis of the plot.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A boxplot of monthly mean flows with highlighted points
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaterGageDaily-plot_precip_data"></a>}}
\if{latex}{\out{\hypertarget{method-WaterGageDaily-plot_precip_data}{}}}
\subsection{\code{WaterGageDaily$plot_precip_data()}}{
  Create a boxplot of monthly sums of precipitation using the
data provided in either the precip_data field or by join_precip_data().
The boxplot is created by plot_recharge_context(). This will ONLY use
days where the data in the input precip column is not NA
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{WaterGageDaily$plot_precip_data(
  targetYear = as.numeric(format(Sys.Date(), "\%Y")),
  include_months = 1:12,
  AYS = 10,
  metric = sum,
  precip_col = "precip_mm",
  use_y_log = FALSE,
  ylab = "Monthly Precip (mm)"
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{targetYear}}{What year should be highlighted on the plot? Defaults
to this year. If left as NULL, no additional point data will be included}
      \item{\code{include_months}}{What months should be included in plot/analysis?}
      \item{\code{AYS}}{Which month begins the analysis year? The plot will start in
this month and then proceed through the next 12 months or those
included by user in include_months}
      \item{\code{metric}}{A function to pass to \code{dplyr::summarize()} to control
how data monthly data is aggregated}
      \item{\code{precip_col}}{The column to plot in the precip data frame}
      \item{\code{use_y_log}}{Boolean. Should the y axis be on a log10 scale? Defaults
to FALSE}
      \item{\code{ylab}}{Character. The text to display on the y-axis of the plot.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A boxplot of precipitation data, also set in the plots field
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaterGageDaily-plot_low_flows"></a>}}
\if{latex}{\out{\hypertarget{method-WaterGageDaily-plot_low_flows}{}}}
\subsection{\code{WaterGageDaily$plot_low_flows()}}{
  Create a line plot of the lowest x-day average flow recorded
each analysis year as set in \code{comp_xQy()}. For instance, the plot
will include the lowest 7-day average flow recorded each analysis year.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{WaterGageDaily$plot_low_flows(plot_type = "lineplot")}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{plot_type}}{Either "lineplot" or "boxplot". plot_type will control
which kind of plot is returned to user and set in the plots field,
either a "lineplot" of the low flow time series or a "boxplot" per
average low flow metric}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A line plot of the x-day average low flow recorded each analysis
year. Will set in the plots list as well.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaterGageDaily-get_doy_percentiles"></a>}}
\if{latex}{\out{\hypertarget{method-WaterGageDaily-get_doy_percentiles}{}}}
\subsection{\code{WaterGageDaily$get_doy_percentiles()}}{
  Gets percentile day-of-year statistics from the USGS using
\code{WaterGageStats}
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{WaterGageDaily$get_doy_percentiles()}
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A \code{WaterGageStats} object
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaterGageDaily-plot_percentiles"></a>}}
\if{latex}{\out{\hypertarget{method-WaterGageDaily-plot_percentiles}{}}}
\subsection{\code{WaterGageDaily$plot_percentiles()}}{
  Creates a plot of the the daily historic flow percentiles,
in bands from 100, 95, 90, 75, 50, 25, 10, 5, and 0. Also plots the
daily flow, to show where the flow is in term sof historic flows of that
day. This uses \code{WaterGageStats()} to retrieve the percentiles
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{WaterGageDaily$plot_percentiles(
  start_date = (Sys.Date() - 7),
  end_date = Sys.Date()
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{start_date}}{The lower bound of the daily flows to put on the plot, in YYYY-MM-DD.
If more than#' a year is selected, the percentiles repeat (i.e. the bands for
2024-01-01 will be the #' same as the bands for 2025-01-01)}
      \item{\code{end_date}}{The upper bound of daily flows to show, in YYYY-MM-DD.. Defaults to
today's date}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A plot showing the daily flow and the percentile bands for each day
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaterGageDaily-plot_all"></a>}}
\if{latex}{\out{\hypertarget{method-WaterGageDaily-plot_all}{}}}
\subsection{\code{WaterGageDaily$plot_all()}}{
  Run all plotting methods and populate the plots field accordingly
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{WaterGageDaily$plot_all(
  recharge_targetYear = as.numeric(format(Sys.Date(), "\%Y")),
  recharge_include_months = 1:12,
  recharge_AYS = 10,
  forecast_start_date = (Sys.Date() - 15)
)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{recharge_targetYear}}{What year should be highlighted on the recharge plot?
See \code{self$plot_recharge_context()} for more information.}
      \item{\code{recharge_include_months}}{What months (numeric) should be included in plot/analysis?}
      \item{\code{recharge_AYS}}{Which month begins the recharge analysis year?
See \code{self$plot_recharge_context()} for more information.}
      \item{\code{forecast_start_date}}{Character. The date to begin the forecast, i.e. the flow
on this day will be used to run forecast calculations. See
\code{self$plot_baseflow_forecast()} for more information.}
    }
    \if{html}{\out{</div>}}
  }
  \subsection{Returns}{
    A list of all plots created by methods in this objects. Will also
populate the plots field.
  }
}

\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WaterGageDaily-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WaterGageDaily-clone}{}}}
\subsection{\code{WaterGageDaily$clone()}}{
  The objects of this class are cloneable with this method.
  \subsection{Usage}{
    \if{html}{\out{<div class="r">}}
    \preformatted{WaterGageDaily$clone(deep = FALSE)}
    \if{html}{\out{</div>}}
  }
  \subsection{Arguments}{
    \if{html}{\out{<div class="arguments">}}
    \describe{
      \item{\code{deep}}{Whether to make a deep clone.}
    }
    \if{html}{\out{</div>}}
  }
}

}
