% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_handlers.R
\name{WatershedModelNode}
\alias{WatershedModelNode}
\title{WatershedModelNode}
\value{
R6Class of type WatershedModelNode
}
\description{
Utility class for interacting with a watershed feature/model combo
}
\details{
Watershed Model Node data object

Has standard methods for managing data and meta data
}
\examples{
NA
}
\seealso{
NA
}
\section{Super class}{
\code{\link[hydrotools:ModelElementBase]{hydrotools::ModelElementBase}} -> \code{WatershedModelNode}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{ds}}{RomDataSource}

\item{\code{ds_om}}{RomDataSource points to legacy model db}

\item{\code{hydroid}}{unique ID from dh_feature database}

\item{\code{hydrocode}}{non-numeric identifier may reference ext source db}

\item{\code{bundle}}{type of feature}

\item{\code{ftype}}{sub-type of feature}

\item{\code{feature}}{a RomFeature entity}

\item{\code{pid}}{integer identifier}

\item{\code{version}}{descriptive code for model}

\item{\code{prop}}{as RomProperty}

\item{\code{json}}{as list}

\item{\code{elementid}}{as integer}

\item{\code{rocode}}{describes runoff type default = 'cbp6_lrseg'}

\item{\code{name}}{what is it called}

\item{\code{riverseg}}{we using code names}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WatershedModelNode-new}{\code{WatershedModelNode$new()}}
\item \href{#method-WatershedModelNode-get_runoff_elids}{\code{WatershedModelNode$get_runoff_elids()}}
\item \href{#method-WatershedModelNode-nearest_gage}{\code{WatershedModelNode$nearest_gage()}}
\item \href{#method-WatershedModelNode-render_cia_detailed}{\code{WatershedModelNode$render_cia_detailed()}}
\item \href{#method-WatershedModelNode-render_cia_brief}{\code{WatershedModelNode$render_cia_brief()}}
\item \href{#method-WatershedModelNode-clone}{\code{WatershedModelNode$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="ModelElementBase" data-id="get_feature"><a href='../../hydrotools/html/ModelElementBase.html#method-ModelElementBase-get_feature'><code>hydrotools::ModelElementBase$get_feature()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="ModelElementBase" data-id="get_json_model"><a href='../../hydrotools/html/ModelElementBase.html#method-ModelElementBase-get_json_model'><code>hydrotools::ModelElementBase$get_json_model()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="ModelElementBase" data-id="get_model"><a href='../../hydrotools/html/ModelElementBase.html#method-ModelElementBase-get_model'><code>hydrotools::ModelElementBase$get_model()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="ModelElementBase" data-id="get_model_pid"><a href='../../hydrotools/html/ModelElementBase.html#method-ModelElementBase-get_model_pid'><code>hydrotools::ModelElementBase$get_model_pid()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="ModelElementBase" data-id="get_run_data"><a href='../../hydrotools/html/ModelElementBase.html#method-ModelElementBase-get_run_data'><code>hydrotools::ModelElementBase$get_run_data()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="ModelElementBase" data-id="handle_config"><a href='../../hydrotools/html/ModelElementBase.html#method-ModelElementBase-handle_config'><code>hydrotools::ModelElementBase$handle_config()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="ModelElementBase" data-id="load_json_model"><a href='../../hydrotools/html/ModelElementBase.html#method-ModelElementBase-load_json_model'><code>hydrotools::ModelElementBase$load_json_model()</code></a></span></li>
<li><span class="pkg-link" data-pkg="hydrotools" data-topic="ModelElementBase" data-id="om_element_tree"><a href='../../hydrotools/html/ModelElementBase.html#method-ModelElementBase-om_element_tree'><code>hydrotools::ModelElementBase$om_element_tree()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WatershedModelNode-new"></a>}}
\if{latex}{\out{\hypertarget{method-WatershedModelNode-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WatershedModelNode$new(ds, config = list(), ds_om = NA, site = omsite)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{ds}}{RomDataSource object mandatory}

\item{\code{config}}{list of attributes to set, see also: to_list() for format}

\item{\code{ds_om}}{RomDataSource object pointing to OM (optional)}

\item{\code{site}}{URI base for the legacy model scripts assumes global var omsite exists for default}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
object instance
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WatershedModelNode-get_runoff_elids"></a>}}
\if{latex}{\out{\hypertarget{method-WatershedModelNode-get_runoff_elids}{}}}
\subsection{Method \code{get_runoff_elids()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WatershedModelNode$get_runoff_elids()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
list of elementids corresponding to runoff props uses rocode prop to filter
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WatershedModelNode-nearest_gage"></a>}}
\if{latex}{\out{\hypertarget{method-WatershedModelNode-nearest_gage}{}}}
\subsection{Method \code{nearest_gage()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WatershedModelNode$nearest_gage()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
data frame with the closest gage
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WatershedModelNode-render_cia_detailed"></a>}}
\if{latex}{\out{\hypertarget{method-WatershedModelNode-render_cia_detailed}{}}}
\subsection{Method \code{render_cia_detailed()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WatershedModelNode$render_cia_detailed(runids, export_path, github_location)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{runids}}{runids to summarize}

\item{\code{export_path}}{where to store the file}

\item{\code{github_location}}{library to load rendering code}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
file path of rendered output
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WatershedModelNode-render_cia_brief"></a>}}
\if{latex}{\out{\hypertarget{method-WatershedModelNode-render_cia_brief}{}}}
\subsection{Method \code{render_cia_brief()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WatershedModelNode$render_cia_brief(
  runid,
  export_path,
  github_location,
  cu_pre_var = "Qout",
  cu_post_var = "Qout",
  doc_title = "auto",
  cu_threshold = c(-10, -20, -30),
  include_appendices = c()
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{runid}}{runid to summarize}

\item{\code{export_path}}{where to store the file}

\item{\code{github_location}}{library to load rendering code}

\item{\code{cu_pre_var}}{variable for "BEFORE" flow values}

\item{\code{cu_post_var}}{variable for "AFTER"}

\item{\code{doc_title}}{document title page text, default generated from feature}

\item{\code{cu_threshold}}{thresholds of change to code yellow, orange, red}

\item{\code{include_appendices}}{options 'hydropower', ...}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
file path of rendered output
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WatershedModelNode-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WatershedModelNode-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WatershedModelNode$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
